Module Bash_lib.Ast

AST

type var = string * string

Variable reference in the form of name[subscript]

val pp_var : Ppx_deriving_runtime.Format.formatter -> var -> Ppx_deriving_runtime.unit

val show_var : var -> Ppx_deriving_runtime.string

type arithm =

| Num of int

| Var of var

| Plus of arithm * arithm

| Minus of arithm * arithm

| Mul of arithm * arithm

| Div of arithm * arithm

| Less of arithm * arithm

| Greater of arithm * arithm

| LessEq of arithm * arithm

| GreaterEq of arithm * arithm

| Equal of arithm * arithm

| NEqual of arithm * arithm

Arithmetical expression

val pp_arithm : Ppx_deriving_runtime.Format.formatter -> arithm -> Ppx_deriving_runtime.unit

val show_arithm : arithm -> Ppx_deriving_runtime.string

type param_exp =

| Param of var	(*
$name, $\{name\}
*)

| PosParam of int	(*
Positional parameters denoted with a number
*)

| Length of var	(*
$\{#name\}
*)

| Substring of var * arithm * arithm option	(*
$\{name:offset[:length]\}
*)

| CutMinBeg of var * string	(*
$\{name#pattern\}
*)

| CutMaxBeg of var * string	(*
$\{name##pattern\}
*)

| CutMinEnd of var * string	(*
$\{name%pattern\}
*)

| CutMaxEnd of var * string	(*
$\{name%%pattern\}
*)

| SubstOne of var * string * string	(*
$\{name/pattern[/string]\}
*)

| SubstAll of var * string * string	(*
$\{name//pattern[/string]\}
*)

| SubstBeg of var * string * string	(*
$\{name/#pattern[/string]\}
*)

| SubstEnd of var * string * string	(*
$\{name/%pattern[/string]\}
*)

Parameter expansion

val pp_param_exp : Ppx_deriving_runtime.Format.formatter -> param_exp -> Ppx_deriving_runtime.unit

val show_param_exp : param_exp -> Ppx_deriving_runtime.string

type word =

| BraceExp of word list	(*
Brace expansion
*)

| ParamExp of param_exp	(*
Parameter expansion
*)

| CmdSubst of cmd	(*
$(command)
*)

| ArithmExp of arithm	(*
$((...))
*)

| FilenameExp of string	(*
A string that includes *, ? or [
*)

| Word of string	(*
If none of the expansions above are applicable
*)

Token that may be subject to expansions

and cmd =

| Assignt of assignt list	(*
name1=val1 name1=val2 ...
*)

| Command of assignt list * word list	(*
[ assignments ] command [ parameters ]
*)

Simple command

and assignt =

| SimpleAssignt of var * word	(*
variable=[ value ]
*)

| IndArrAssignt of string * word list	(*
name=(word1 word2 ...), if no words are provided, the indexed array is not set
*)

| AssocArrAssignt of string * (string * word) list	(*
name=(key1=value1 key2=value2 ...), if no pairs are provided, the array is not set
*)

Assignment (differs from the original Bash not to depend on the declare built-in)

val pp_word : Ppx_deriving_runtime.Format.formatter -> word -> Ppx_deriving_runtime.unit

val show_word : word -> Ppx_deriving_runtime.string

val pp_cmd : Ppx_deriving_runtime.Format.formatter -> cmd -> Ppx_deriving_runtime.unit

val show_cmd : cmd -> Ppx_deriving_runtime.string

val pp_assignt : Ppx_deriving_runtime.Format.formatter -> assignt -> Ppx_deriving_runtime.unit

val show_assignt : assignt -> Ppx_deriving_runtime.string

type redir =

| RedirInp of int * word	(*
[n]<word
*)

| RedirOtp of int * word	(*
[n]>word
*)

| AppendOtp of int * word	(*
[n]>>word
*)

| DuplInp of int * word	(*
[n]<&word
*)

| DuplOtp of int * word	(*
[n]>&word
*)

Redirection

val pp_redir : Ppx_deriving_runtime.Format.formatter -> redir -> Ppx_deriving_runtime.unit

val show_redir : redir -> Ppx_deriving_runtime.string

type pipeline_list =

| Pipeline of pipeline	(*
list containing a single pipeline
*)

| PipelineAndList of pipeline * pipeline_list	(*
pipeline && other-pipelines
*)

| PipelineOrList of pipeline * pipeline_list	(*
pipeline || other-pipelines
*)

List of pipelines

and pipeline = bool * compound * compound list

Pipeline in the form of [ ! ] command1 [ | command2 [ | ... ] ]

and compound =

| While of while_loop * redir list

| ForList of for_list_loop * redir list

| ForExpr of for_expr_loop * redir list

| If of if_stmt * redir list

| Case of case_stmt * redir list

| ArithmExpr of arithm * redir list	(*
(( ... ))
*)

| SimpleCommand of cmd * redir list

Compound command

and while_loop = pipeline_list * pipeline_list

While loop in the form of while list; do list; done

and for_list_loop = string * word list * pipeline_list

For loop in the form of or name in [ word ... ]; do list ; done

and for_expr_loop = arithm * arithm * arithm * pipeline_list

For loop in the form of for (( expr1 ; expr2 ; expr3 )) ; do list ; done

and if_stmt = pipeline_list * pipeline_list * pipeline_list option

If statement in the form of if list; then list; [ else list; ] fi

and case_stmt = word * case_item list

Case statement in the form of case word in [ case_item ] ... esac

and case_item = word * word list * pipeline_list

An element of a case statement in the form of [(] pattern [ | pattern ] ... ) list ;;

val pp_pipeline_list : Ppx_deriving_runtime.Format.formatter -> pipeline_list -> Ppx_deriving_runtime.unit

val show_pipeline_list : pipeline_list -> Ppx_deriving_runtime.string

val pp_pipeline : Ppx_deriving_runtime.Format.formatter -> pipeline -> Ppx_deriving_runtime.unit

val show_pipeline : pipeline -> Ppx_deriving_runtime.string

val pp_compound : Ppx_deriving_runtime.Format.formatter -> compound -> Ppx_deriving_runtime.unit

val show_compound : compound -> Ppx_deriving_runtime.string

val pp_while_loop : Ppx_deriving_runtime.Format.formatter -> while_loop -> Ppx_deriving_runtime.unit

val show_while_loop : while_loop -> Ppx_deriving_runtime.string

val pp_for_list_loop : Ppx_deriving_runtime.Format.formatter -> for_list_loop -> Ppx_deriving_runtime.unit

val show_for_list_loop : for_list_loop -> Ppx_deriving_runtime.string

val pp_for_expr_loop : Ppx_deriving_runtime.Format.formatter -> for_expr_loop -> Ppx_deriving_runtime.unit

val show_for_expr_loop : for_expr_loop -> Ppx_deriving_runtime.string

val pp_if_stmt : Ppx_deriving_runtime.Format.formatter -> if_stmt -> Ppx_deriving_runtime.unit

val show_if_stmt : if_stmt -> Ppx_deriving_runtime.string

val pp_case_stmt : Ppx_deriving_runtime.Format.formatter -> case_stmt -> Ppx_deriving_runtime.unit

val show_case_stmt : case_stmt -> Ppx_deriving_runtime.string

val pp_case_item : Ppx_deriving_runtime.Format.formatter -> case_item -> Ppx_deriving_runtime.unit

val show_case_item : case_item -> Ppx_deriving_runtime.string

type func = string * compound

Function in the form of function name [()] compound or name () compound

val pp_func : Ppx_deriving_runtime.Format.formatter -> func -> Ppx_deriving_runtime.unit

val show_func : func -> Ppx_deriving_runtime.string

type script =

| Empty

| Script of script_elem * script

AST root

and script_elem =

| Func of func

| Pipelines of pipeline_list

A function declaration or a pipeline list

val pp_script : Ppx_deriving_runtime.Format.formatter -> script -> Ppx_deriving_runtime.unit

val show_script : script -> Ppx_deriving_runtime.string

val pp_script_elem : Ppx_deriving_runtime.Format.formatter -> script_elem -> Ppx_deriving_runtime.unit

val show_script_elem : script_elem -> Ppx_deriving_runtime.string