Module Interpreter.Eval

Parameters

module M : MONAD_FAIL

Signature

module Map : sig ... end

type var_t =

| Val of string	(*
Simple variable
*)

| IndArray of string list	(*
Indexed array
*)

| AssocArray of string Map.t	(*
Associative array
*)

Container for values of variables

type fun_t = string list -> string * string * int

Container for functions (takes arguments and produces contents of stdout, stderr, and return code)

type variables = var_t Map.t

Environment containing variables available in the current scope

type functions = fun_t Map.t

Environment containing functions defined in the current scope

type environment = {

vars : variables;

funs : functions;

retcode : int;

}

Complete environment

type environments = environment list

All available environments with the first element as the innermost scope

val find_var : string -> environments -> var_t option

Searches for a variable with the given name in the provided environments

val find_fun : string -> environments -> fun_t option

Searches for a function with the given name in the provided environments

val ev_var : environments -> Ast.var -> string M.t

Evaluate variable

val ev_arithm : environments -> Ast.arithm -> int M.t

Evaluate arithmetic

val ev_script : 'a -> 'b M.t

Evaluate Bash script