Module Bash_lib.Parser

val chainl1 : 'a Angstrom.t -> ('a -> 'a -> 'a) Angstrom.t -> 'a Angstrom.t

chainl1 e op parses one or more occurrences of e, separated by op. Returns a value obtained by a left associative application of op to the values returned by e.

val str_exists : (char -> bool) -> string -> bool

Check if string s contains a char that satisfies predicate p

val int_p : int Angstrom.t

Integer parser

val reserved : string list

val is_blank : char -> bool

val blank : string Angstrom.t

val blank1 : string Angstrom.t

val trim : 'a Angstrom.t -> 'a Angstrom.t

val parens : 'a Angstrom.t -> 'a Angstrom.t

val is_delim : char -> bool

val delim1 : string Angstrom.t

val is_meta : char -> bool

val meta : string Angstrom.t

val non_meta : string Angstrom.t

val name_p : Ast.name Angstrom.t

Name parser

val var_p : Ast.var Angstrom.t

Variable parser

val plus : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val minus : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val mul : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val div : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val less : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val greater : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val lesseq : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val greatereq : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val equal : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val nequal : (Ast.arithm -> Ast.arithm -> Ast.arithm) Angstrom.t

val arithm_p : Ast.arithm Angstrom.t

Arithmetic parser

val word_p : ?brc:bool -> ?prm:bool -> ?cmd:bool -> ?ari:bool -> ?fln:bool -> unit -> Ast.word Angstrom.t

Word parser. Parameters determine which expansions may be performed.

val brace_exp : Ast.word Angstrom.t

val param_exp_p : Ast.param_exp Angstrom.t

Parameter expansion parser

val inn_cmd_subst : unit -> Ast.word Angstrom.t

val arithm_exp : Ast.word Angstrom.t

val filename_exp : Ast.word Angstrom.t

val inn_assignt_p : unit -> Ast.assignt Angstrom.t

val inn_cmd_p : unit -> Ast.cmd Angstrom.t

val assignt_p : Ast.assignt Angstrom.t

Assignment parser

val cmd_p : Ast.cmd Angstrom.t

Simple command parser

val redir_p : Ast.redir Angstrom.t

Redirection parser

val ctrl_m : string -> string Angstrom.t

val ctrl_e : string -> string Angstrom.t

val inn_pipeline_list_p : unit -> Ast.pipeline_list Angstrom.t

val inn_pipeline_p : unit -> Ast.pipeline Angstrom.t

val inn_compound_p : unit -> Ast.compound Angstrom.t

val inn_while_loop_p : unit -> Ast.while_loop Angstrom.t

val inn_for_loop_p : unit -> Ast.for_loop Angstrom.t

val inn_if_stmt_p : unit -> Ast.if_stmt Angstrom.t

val inn_case_stmt_p : unit -> Ast.case_stmt Angstrom.t

val inn_case_item_p : unit -> Ast.case_item Angstrom.t

val pipeline_list_p : Ast.pipeline_list Angstrom.t

Pipeline list parser

val pipeline_p : Ast.pipeline Angstrom.t

Pipeline parser

val compound_p : Ast.compound Angstrom.t

Compound parser

val while_loop_p : Ast.while_loop Angstrom.t

While loop parser

val for_loop_p : Ast.for_loop Angstrom.t

For loop parser

val if_stmt_p : Ast.if_stmt Angstrom.t

If statement parser

val case_stmt_p : Ast.case_stmt Angstrom.t

Case statement parser

val case_item_p : Ast.case_item Angstrom.t

Case item parser

val func_p : Ast.func Angstrom.t

Function parser

val script_elem_p : Ast.script_elem Angstrom.t

Script element parser

val script_p : Ast.script Angstrom.t

Bash script parser

val parse : string -> (Ast.script, string) Stdlib.result

Parses the given string as a Bash script