Module Bindings__GParam

type flags =

| Readable
the parameter is readable

| Writable
the parameter is writable

| Readwrite
alias for G_PARAM_READABLE | G_PARAM_WRITABLE

| Construct
the parameter will be set upon object construction

| Construct_only
the parameter can only be set upon object construction

| Lax_validation
upon parameter conversion (see g_param_value_convert()) strict validation is not required

| Static_name
the string used as name when constructing the parameter is guaranteed to remain valid and unmodified for the lifetime of the parameter. Since 2.8

| Static_nick
the string used as nick when constructing the parameter is guaranteed to remain valid and unmmodified for the lifetime of the parameter. Since 2.8

| Static_blurb
the string used as blurb when constructing the parameter is guaranteed to remain valid and unmodified for the lifetime of the parameter. Since 2.8

| Explicit_notify
calls to g_object_set_property() for this property will not automatically result in a "notify" signal being emitted: the implementation must call g_object_notify() themselves in case the property actually changes. Since: 2.42.

| Deprecated
the parameter is deprecated and will be removed in a future version. A warning will be generated if it is used while running with G_ENABLE_DIAGNOSTIC=1. Since 2.26

val string_of_flag : flags -> string

module Flags : functor (T : Cstubs.Types.TYPE) -> sig ... end