Module Csharp_Exc_Lib.Base_lib

Copyright 2021-2023, Georgy Sichkar

SPDX-License-Identifier: LGPL-3.0-or-later

This file describes the library that is included in each program and is available in the global scope

module type UNIT_METHOD = sig ... end

module File_Info : sig ... end

A module that allows you to interact with files (open)

module Exception : sig ... end

Module describing the base exception class

val classes_with_system_classes : Common_types.code_ident list

val get_system_method_opt : Common_types.code_ident -> Ast.ident -> (unit, 'a) Monads.Eval.t option

val get_system_constr_opt : Common_types.code_ident -> (Common_types.address -> (unit, 'a) Monads.Eval.t) option

val run_sys_constructor_or_normal : (Common_types.address, 'a) Monads.Eval.t -> Common_types.code_ident -> (Common_types.address, 'a) Monads.Eval.t

val main_method : Ast.ident

val base_lib_decl : Ast.tast option

Get Ast of base lib