Module Csharp_Exc_Lib.Common_types

Copyright 2021-2023, Georgy Sichkar

SPDX-License-Identifier: LGPL-3.0-or-later

type code_ident =

| Code_ident of Ast.ident

val pp_code_ident : Stdlib.Format.formatter -> code_ident -> unit

val show_code_ident : code_ident -> string

val equal_code_ident : code_ident -> code_ident -> bool

module Code_id : sig ... end

module Code_Map : sig ... end

module Ident : sig ... end

module Ident_Map : sig ... end

type address =

| Link of int

Represents the address of an object instance

val pp_address : Stdlib.Format.formatter -> address -> unit

val show_address : address -> string

val equal_address : address -> address -> bool

val incr_ : address -> address

val ln : int -> address

module Mem_Address : sig ... end

module Mem_Map : sig ... end

type internal_address =

| ILink of int

Represents a reference to a description of the internal fields of a system class instance from Base_lib This is necessary because some field types are not representable in the language, such as out_channel

val pp_internal_address : Stdlib.Format.formatter -> internal_address -> unit

val show_internal_address : internal_address -> string

val equal_internal_address : internal_address -> internal_address -> bool

val i_incr_ : internal_address -> internal_address

val i_ln : int -> internal_address

module Intern_Mem_Address : sig ... end

module Intern_Mem : sig ... end

type base_lib_id =

| Fl_descriptors

Type used as a key for Intern_Mems

module Sys_id : sig ... end

module Sys_Map : sig ... end

type file_st =

| W_File of Stdlib.out_channel

type sys_mem_el =

| Files of internal_address * file_st Intern_Mem.t
(*
Next available internal_address * Map
*)

type sys_memory = sys_mem_el Sys_Map.t

type code_ctx =

| Exception_ctx of Ast.class_decl

| Class_ctx of Ast.class_decl

val get_class_decl : code_ctx -> Ast.class_decl

type text = code_ctx Code_Map.t