Module ClassLoader.ClassLoader

Parameters

module M : SupportTypes.MONADERROR

Signature

val is_public : Ast.modifiers list -> bool

val is_private : Ast.modifiers list -> bool

val is_protected : Ast.modifiers list -> bool

val is_abstract : Ast.modifiers list -> bool

val is_static : Ast.modifiers list -> bool

val is_const : Ast.modifiers list -> bool

val is_virtual : Ast.modifiers list -> bool

val is_override : Ast.modifiers list -> bool

val is_new : Ast.modifiers list -> bool

val check_mutually_exclusive_modifiers : Ast.modifiers list -> bool

Return false if modifiers have public&private, public&protected, private&protected

val get_element_option : Ast.KeyMap.key -> 'a Ast.KeyMap.t -> 'a option

val update_element : 'a Ast.KeyMap.t -> Ast.KeyMap.key -> 'b -> 'a Ast.KeyMap.t

val list_map : 'a list -> ( 'b -> 'c M.t ) -> 'd -> 'e M.t

val seq_iter : 'a Stdlib.Seq.t -> ( 'b -> 'c M.t ) -> 'd -> 'e M.t

val merge_map_function : 'a -> 'b option -> 'b option -> 'b option

val prepare_object_class : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t M.t

Add hardcoded class Object and its methods Equals() and ToString() to hashtable

val get_modifiers_list : unit

val check_modifiers_class_element : (Ast.modifiers list * Ast.fields) -> unit M.t

Checking modifiers for class methods --- list of modifiers (e.g. const, static, public, etc.), parsed object element (field/method/constructor)

val check_modifiers_interface_element : (Ast.modifiers list * Ast.fields) -> unit M.t

Checking modifiers for interface methods --- list of modifiers (e.g. const, static, public, etc.), parsed object element (field/method/constructor)

val check_modifiers_object : Ast.objects -> unit M.t

val many_add_objects : Ast.objects list -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t M.t

val get_default_constructor : SupportTypes.table_class -> SupportTypes.table_constructor Ast.KeyMap.t

Make default (empty) constructor for class

val build_new_table_class : ('a * SupportTypes.table_class) Stdlib.Seq.t -> ('b * SupportTypes.table_class) Stdlib.Seq.t -> ('b * SupportTypes.table_class) Stdlib.Seq.t

Add default (empty) constructor to class for all non-static classes in sequence

val many_add_default_constructor : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t M.t

Add an empty constructor to all classes that don't have a constructor

val check_children : SupportTypes.table_class Ast.KeyMap.t -> Ast.KeyMap.key list -> bool -> unit M.t

Checks the list of parents for the existence of parents -- table of all classes, list of parent keys and is already inheritance from class flag

val iterate_for_checking_childs : SupportTypes.table_class Ast.KeyMap.t -> ('a * SupportTypes.table_class) Stdlib.Seq.t -> unit M.t

Iterate by all class sequence and check parent list for errors

val build_parents_table : 'a -> Ast.KeyMap.key list -> string -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t

Create class table with updated children key field. Without any check

val iterate_for_add_children : 'a -> (Ast.KeyMap.key * SupportTypes.table_class) Stdlib.Seq.t -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t

Iterate by all classes in sequence and update their parents 'children_keys' field

val many_add_children : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t M.t

Update children keys for all parents

val build_new_children_fields : (string * SupportTypes.table_field) Stdlib.Seq.t -> SupportTypes.table_class -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t * SupportTypes.table_class

val many_field_inheritance : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class -> SupportTypes.table_class -> (SupportTypes.table_class Ast.KeyMap.t * SupportTypes.table_class) M.t

val is_base_method : Ast.expressions -> bool

val is_this_method : Ast.expressions -> bool

val many_check_parent_constructor : 'a -> SupportTypes.table_class -> 'b M.t

val many_check_children_constructor : 'a -> SupportTypes.table_class -> 'b M.t

val iterate_for_checking_methods : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class -> (string * SupportTypes.table_method) Stdlib.Seq.t -> unit M.t

Iterate by all methods in parent sequence and check children methods for errors

val build_new_children_methods : (string * SupportTypes.table_method) Stdlib.Seq.t -> SupportTypes.table_class -> SupportTypes.table_class -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t * SupportTypes.table_class

val many_method_inheritance : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class -> SupportTypes.table_class -> (SupportTypes.table_class Ast.KeyMap.t * SupportTypes.table_class) M.t

Adds explicitly unimplemented methods of parent classes to children

val inheritance_object_class_methods : (string * SupportTypes.table_method) Stdlib.Seq.t -> SupportTypes.table_class -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t * SupportTypes.table_class

val many_object_class_inheritance : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class -> (SupportTypes.table_class Ast.KeyMap.t * SupportTypes.table_class) M.t

Adds object class methods (e.g. Equals() or ToString()) to every class

val many_check_method_override : 'a -> SupportTypes.table_class -> SupportTypes.table_class -> 'b M.t

val many_check_interface_inheritance : 'a -> SupportTypes.table_class -> SupportTypes.table_class -> 'b M.t

Checks that classes implement all methods of parent interfaces

val iter_by_childs : 'a -> ( 'b -> 'c -> 'd M.t ) -> 'e list -> 'f M.t

val transfert : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class -> SupportTypes.table_class -> SupportTypes.table_class Ast.KeyMap.t M.t

val transfert_on_children : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class -> SupportTypes.table_class Ast.KeyMap.t M.t

val inheritance : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t M.t

val load : Ast.objects list -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class Ast.KeyMap.t M.t