Module Interpreter.Interpretation

Parameters

module M : SupportTypes.MONADERROR

Signature

type variable = {

v_type : Ast.types;

v_key : string;

is_const : bool;

assignment_count : int;

v_value : Ast.values;

visibility_level : int;

}

val pp_variable : Ppx_show_runtime.Format.formatter -> variable -> unit

val show_variable : variable -> string

type signal =

| WasBreak

| WasContinue

| WasReturn

| NoSignal

val pp_signal : Ppx_show_runtime.Format.formatter -> signal -> unit

val show_signal : signal -> string

type context = {

cur_object : Ast.object_references;

var_table : variable Ast.KeyMap.t;

last_expr_result : Ast.values;

runtime_signal : signal;

curr_method_type : Ast.types;

is_main_scope : bool;

nested_loops_cnt : int;

visibility_level : int;

cur_constr_key : string option;

prev_context : context option;

obj_created_cnt : int;

is_creation : bool;

constr_affilation : string option;

class_for_cast : SupportTypes.table_class option;

}

val pp_context : Ppx_show_runtime.Format.formatter -> context -> unit

val show_context : context -> string

val make_context : Ast.object_references -> variable Ast.KeyMap.t -> context M.t

val get_type_default_value : Ast.types -> Ast.values

val seq_hd : ( unit -> 'a Stdlib.Seq.node ) -> 'b

val mfold_left2 : ( 'a -> 'b -> 'c -> 'd M.t ) -> 'e -> 'f list -> 'g list -> 'h M.t

Applies the F function to the elements of each list and the already accumulated value and returns the accumulated result

val mfold_right : ( 'a -> 'b -> 'c M.t ) -> 'd list -> 'e -> 'f M.t

val mfold_left : ( 'a -> 'b -> 'c M.t ) -> 'd -> 'e list -> 'f M.t

val get_field_list : Ast.objects -> Ast.fields list

val convert_element_pair_list : ('a * ('b * 'c) list) -> ('d * 'e * 'f) list

val get_variable_field_pairs_list_typed : Ast.objects -> (Ast.types * Ast.names * Ast.expressions option) list

val get_object_fields : Ast.object_references -> Ast.field_references Ast.KeyMap.t

val get_object_info : Ast.object_references -> string * string option * Ast.field_references Ast.KeyMap.t * int

val get_ref_identifier : Ast.types -> string M.t

val get_elem_if_present_m : Ast.KeyMap.key -> 'a Ast.KeyMap.t -> 'b M.t

val find_main_class : SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class M.t

val find_parent_class_key : Ast.KeyMap.key list -> SupportTypes.table_class Ast.KeyMap.t -> Ast.KeyMap.key option -> Ast.KeyMap.key option

val startswith : string -> Stdlib.String.t -> bool

val check_parent_tree : SupportTypes.table_class Ast.KeyMap.t -> string -> Ast.KeyMap.key -> bool

Сheck that the left class is a parent of the right

val find_last_overriden : Ast.KeyMap.key -> SupportTypes.table_class -> Ast.KeyMap.key -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class M.t

val find_nearest_child : Ast.KeyMap.key -> Ast.KeyMap.key list -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_class M.t

Finding the nearest direct descendant of the class cast to

val expr_type_check : Ast.expressions -> context -> SupportTypes.table_class Ast.KeyMap.t -> Ast.types M.t

val check_classname_assign : string -> Ast.KeyMap.key -> SupportTypes.table_class Ast.KeyMap.t -> Ast.types M.t

Сheck that the right class can be assigned to left and return right type

val check_classname_bool : string -> Ast.KeyMap.key -> SupportTypes.table_class Ast.KeyMap.t -> bool

Сheck that the left class is a parent of the right

val check_method : SupportTypes.table_class -> Stdlib.String.t -> Ast.expressions list -> context -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_method M.t

Verifies that the called method with the passed number of arguments exists in the singular in the class, and that the argument types are the same as expected. Get current class, name of method for check, its arguments, context, table of all classes

val check_constructor : SupportTypes.table_class -> Ast.expressions list -> context -> SupportTypes.table_class Ast.KeyMap.t -> SupportTypes.table_constructor M.t

Verifies that the called constructor with the passed number of arguments exists in the singular in the class, and that the argument types are the same as expected. Get current class, constructor arguments, context, table of all classes

val get_obj_value : Ast.values -> Ast.object_references

val add_var_to_context : context -> variable -> context

val inc_visibility_level : context -> context

val dec_visibility_level : context -> context

val check_assign_cnt_f : Ast.field_references -> unit M.t

val check_assign_cnt_v : variable -> unit M.t

val can_access_modifier : Ast.modifiers list -> Ast.KeyMap.key -> Ast.object_references -> SupportTypes.table_class Ast.KeyMap.t -> bool

val delete_scope_var : context -> context M.t

Delete nested variables (e.g. in loops)

val is_good_for_stmt : Ast.expressions -> bool

Expressions that can be used as statements (on single line)

val eval_stmt : Ast.statements -> context -> SupportTypes.table_class Ast.KeyMap.t -> context M.t

Interpretate statements

val eval_expr : Ast.expressions -> context -> SupportTypes.table_class Ast.KeyMap.t -> context M.t

val update_identifier_v : Ast.KeyMap.key -> Ast.values -> context -> context M.t

val update_field_v : Ast.expressions -> Ast.KeyMap.key -> context -> SupportTypes.table_class Ast.KeyMap.t -> context M.t

val update_object_state_exn : Ast.object_references -> Ast.KeyMap.key -> Ast.values -> context -> context

val prepare_table_with_args_exn : variable Ast.KeyMap.t -> Ast.expressions list -> (Ast.types * Ast.names) list -> context -> SupportTypes.table_class Ast.KeyMap.t -> (variable Ast.KeyMap.t * context) M.t

Make table of variables by arguments. Concatenate arguments for calling method and base/this method/constructor and return variable table of args. Context doesn't update

val prepare_constructor_block : Ast.statements -> SupportTypes.table_class -> Ast.expressions option -> SupportTypes.table_class Ast.KeyMap.t -> Ast.statements M.t

Add call base/this constructor expression to statement body of constructor

val execute : SupportTypes.table_class Ast.KeyMap.t -> context M.t