Module Math.ComplexField

Complexification of a field

For instance, ComplexNumbers is equivalent to ComplexField (RealNumbers).

Parameters

F : Field

Signature

include Field

include Ring

include Rng

include AbelianGroup

type t

val zero : t

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val mul : t → t → t

val (*) : t → t → t

val one : t

val mone : t

minus one

val of_int : int → t

type names

If the type names remains private, it means it has no use in the current module implementation.

val to_tex : ?⁠names:names → t → string

val of_tex : string → t

val inv : t → t

Will raise Division_by_zero on the zero element.

val div : t → t → t

val (/) : t → t → t

Can raise Division_by_zero.

val of_float : float → t

Try to create an element of the Field from the given float.

val i : t

val mi : t

val mone : t