Module Cypher_lib.Interpret

Copyright 2023-2024, Arseniy Baytenov

SPDX-License-Identifier: MIT

module NameMap : sig ... end

module NameSet : sig ... end

type entity = int * NameSet.t * Ast.constant NameMap.t

module Node : sig ... end

module Edge : sig ... end

type graph = Graph.Persistent.Digraph.ConcreteLabeled(Node)(Edge).t

graph type

type outvalue =

| OutConstant of Ast.constant

| OutNode of int * string list * (string * Ast.constant) list

| OutRel of int * int * string list * (string * Ast.constant) list * int

type of value that can be in output

val pp_outvalue : Stdlib.Format.formatter -> outvalue -> unit

val show_outvalue : outvalue -> string

type output = (int * int) * (int * int) * (string * outvalue) list list

type of output: ((num_creates_nodes, num_created_rels), (num_deleted_nodes, num_deleted_rels), RETURN_output)

val pp_output : Stdlib.Format.formatter -> output -> unit

val show_output : output -> string

type expr_error =

| Type_mismatch

| Division_by_zero

| Undefined_id of string

| Deleted_entity of int

type of error for evaluating expressions

val pp_expr_error : Stdlib.Format.formatter -> expr_error -> unit

val show_expr_error : expr_error -> string

type interpreter_error =

| Type_mismatch

| Multiple_def of string

| Expr_err of expr_error

| Unable_node_deletion of int

| ID_overflow

| Undirected_rel

type of error for interpreting request

val pp_interpreter_error : Stdlib.Format.formatter -> interpreter_error -> unit

val show_interpreter_error : interpreter_error -> string

type req_error =

| Parser_err of string

| Interpreter_err of interpreter_error

type of error for parsing and interpreting request

val pp_req_error : Stdlib.Format.formatter -> req_error -> unit

val show_req_error : req_error -> string

val create_empty_graph : graph

creates empty graph

val interpret_request : graph -> Ast.clause -> (graph * output, req_error) Stdlib.result

takes a graph and a request, interprets it and returns a new graph and an output or an error

val parse_and_interpret_request : graph -> string -> (graph * output, req_error) Stdlib.result

takes a graph and a request, parses and interprets it and returns a new graph and an output or an error