Module Cypher_lib.Interpreter

type value =

| VString of string

| VInt of int

| VBool of bool

| VNull of string

val pp_value : Ppx_deriving_runtime.Format.formatter -> value -> Ppx_deriving_runtime.unit

val show_value : value -> Ppx_deriving_runtime.string

val pp_value_ret : Stdlib.Format.formatter -> value -> unit

val pp_vproperties_ret : Stdlib.Format.formatter -> (string * value) list -> unit

val pp_type_ret : Stdlib.Format.formatter -> string list -> unit

val pp_labels_ret : Stdlib.Format.formatter -> string list -> unit

type vproperty = string * value

val pp_vproperty : Ppx_deriving_runtime.Format.formatter -> vproperty -> Ppx_deriving_runtime.unit

val show_vproperty : vproperty -> Ppx_deriving_runtime.string

type vproperties = vproperty list

val pp_vproperties : Ppx_deriving_runtime.Format.formatter -> vproperties -> Ppx_deriving_runtime.unit

val show_vproperties : vproperties -> Ppx_deriving_runtime.string

type elm_data_src_dst = int * string list * vproperties

val pp_elm_data_src_dst : Ppx_deriving_runtime.Format.formatter -> elm_data_src_dst -> Ppx_deriving_runtime.unit

val show_elm_data_src_dst : elm_data_src_dst -> Ppx_deriving_runtime.string

type elm_data = (int * string list * vproperties) * (elm_data_src_dst option * elm_data_src_dst option)

Uses the same type for vertices and edges to make it easier to store data in a single environment.

Node: (id, labels, vproperties), (None, None) Edge: (id. labels, vproperties), (src, dst)

val pp_elm_data : Ppx_deriving_runtime.Format.formatter -> elm_data -> Ppx_deriving_runtime.unit

val show_elm_data : elm_data -> Ppx_deriving_runtime.string

type abstvalue =

| AValue of value

| AElm of elm_data

val pp_abstvalue : Ppx_deriving_runtime.Format.formatter -> abstvalue -> Ppx_deriving_runtime.unit

val show_abstvalue : abstvalue -> Ppx_deriving_runtime.string

val gen_sym : unit -> int

module ElmData : sig ... end

module Graph : sig ... end

type graph = Graph.t

type env = graph * (string * elm_data) list

val pp_graph : Stdlib.Format.formatter -> Graph.t -> unit

val let* : ( 'a, 'b ) Stdlib.result -> ( 'a -> ( 'c, 'b ) Stdlib.result ) -> ( 'c, 'b ) Stdlib.result

type err =

| IncorrectType

| DivByZero

| NotBound of string

| NoLabelEdge of string

| IncorrectProps

| UnstblElm

| ElmNotValid

| IncorrectWhere of string

| TypeNotValid of string

| UnderectNotValidInCrt

val pp_err : Ppx_deriving_runtime.Format.formatter -> err -> Ppx_deriving_runtime.unit

val show_err : err -> Ppx_deriving_runtime.string

type elms =

| ENode

| EEdge

val find_typeofrelat_env : 'a -> ('a * (('b * string list * 'c) * ('d option * 'e option))) list -> ( abstvalue list, err ) Stdlib.result

val find_valueofelm_env : 'a -> 'b -> ('a * (('c * 'd * ('b * value) list) * ('e * 'f))) list -> ( abstvalue list, err ) Stdlib.result

val find_idofelm_env : 'a -> ('a * ((int * 'b * 'c) * ('d * 'e))) list -> ( abstvalue list, err ) Stdlib.result

val find_elm_env : 'a -> ('a * elm_data) list -> ( abstvalue list, err ) Stdlib.result

val check_var_props : 'a -> ('b * value) list -> 'a -> 'b -> ( value, err ) Stdlib.result

val check_var_id : int -> 'a -> 'a -> ( value, err ) Stdlib.result

val interpret_rec_expr_where : int -> Ast.expr -> string -> (string * value) list -> ( value, err ) Stdlib.result

val interpret_expr_where : int -> Ast.expr -> string -> (string * value) list -> ( bool, err ) Stdlib.result

Initially, the interpreter checks for WHERE that WHERE will return boolean.

val interpret_expr : (string * elm_data) list -> Ast.expr -> ( abstvalue list, err ) Stdlib.result

val interpret_expr_for_ret : Ast.expr -> ( string, 'a ) Stdlib.result

val get_props : (string * elm_data) list -> ('a * Ast.expr) list -> ( ('a * value) list, err ) Stdlib.result

val save_node : 'a option -> string list -> vproperties -> Graph.t -> ('a * ((int * string list * vproperties) * ('b option * 'c option))) list -> ( Graph.t * ('a * ((int * string list * vproperties) * ('b option * 'c option))) list * ((int * string list * vproperties) * ('d option * 'e option)), 'f ) Stdlib.result

val add_node : string option -> string list option -> Graph.t -> (string * elm_data) list -> (string * Ast.expr) list option -> ( Graph.t * (string * elm_data) list * ((int * string list * vproperties) * ('a option * 'b option)), err ) Stdlib.result

val peek_or_add_node : (string * elm_data) list -> Graph.t -> int -> Ast.nodedata -> ( Graph.t * (string * elm_data) list * elm_data * int, err ) Stdlib.result

val save_edge : 'a option -> string -> vproperties -> Graph.t -> ('a * ((int * string list * vproperties) * (elm_data_src_dst option * elm_data_src_dst option))) list -> ElmData.t -> ElmData.t -> ( Graph.t * ('a * ((int * string list * vproperties) * (elm_data_src_dst option * elm_data_src_dst option))) list * ((int * string list * vproperties) * (elm_data_src_dst option * elm_data_src_dst option)), 'b ) Stdlib.result

val add_edge : string option -> Graph.t -> (string * elm_data) list -> ElmData.t -> ElmData.t -> string -> (string * Ast.expr) list option -> ( Graph.t * (string * elm_data) list * ((int * string list * vproperties) * (elm_data_src_dst option * elm_data_src_dst option)), err ) Stdlib.result

val check_data : 'a list -> 'a list -> 'b -> 'c -> ('b * 'c) list -> ('b * 'c) list -> ( ('b * 'c) list * ('b * 'c) list, 'd ) Stdlib.result

val check_data_with_where : int -> (string * value) list -> (string * value) list -> Ast.expr -> string -> ( bool, err ) Stdlib.result

val send_check_datas_without_where : ('a * Ast.expr) list option -> (string * elm_data) list -> 'b list -> ('a * value) list -> string -> elm_data -> (string * elm_data) list -> 'b list option -> ( (string * elm_data) list * (string * elm_data) list, err ) Stdlib.result

val send_check_datas_with_where : int -> (string * Ast.expr) list option -> (string * elm_data) list -> 'a list -> (string * value) list -> string -> Ast.expr -> 'a list option -> ( bool, err ) Stdlib.result

val send_check_datas : int -> (string * Ast.expr) list option -> (string * value) list -> 'a list -> string -> elm_data -> (string * elm_data) list -> (string * elm_data) list -> 'a list option -> elms -> Ast.cmdwithmatch option -> ( (string * elm_data) list * (string * elm_data) list, err ) Stdlib.result

val iter_edges : ('a * elm_data * 'b) list -> string option -> string option -> (string * Ast.expr) list option -> Ast.cmdwithmatch option -> (string * elm_data) list -> (string * elm_data) -> (string * elm_data) -> (string * elm_data) list -> ( (string * elm_data) list * (string * elm_data) list, err ) Result.result

val check_type_edges : Ast.cmdwithmatch option -> Graph.t -> (string * elm_data) list -> (string * elm_data) -> (string * elm_data) -> (string * elm_data) list -> Graph.vertex -> Graph.vertex -> Ast.edgedata -> ( (string * elm_data) list * (string * elm_data) list, err ) Result.result

The function finds all edges between two nodes, iterates one by one and sends it for comparison with user-specified data.

val find_nodes : Ast.cmdwithmatch option -> string -> string list option -> (string * Ast.expr) list option -> Graph.t -> (string * elm_data) list -> elms -> ( (string * elm_data) list * (string * elm_data) list, err ) Stdlib.result

val find_edges : Ast.cmdwithmatch option -> Ast.nodedata -> Ast.edgedata -> Ast.nodedata -> Graph.t -> (string * elm_data) list -> ( (string * elm_data) list * (string * elm_data) list, err ) Stdlib.result

val interp_crt : Ast.elm list -> Graph.t -> (string * elm_data) list -> ( Graph.t * (string * elm_data) list, err ) Stdlib.result

val exe_cmd_vars : Graph.vertex -> (int * string list * vproperties) option -> (int * string list * vproperties) option -> Graph.t -> 'a -> ( Graph.t * 'a, 'b ) Stdlib.result

val interp_cmd_vars : 'a list -> Graph.t -> ('a * Graph.vertex) list -> ( Graph.t * ('a * Graph.vertex) list, err ) Stdlib.result

The function can be used for all commands followed by a list of elements.

val sep_env : 'a -> ('a * 'b) list -> ( ('a * 'b) list * ('a * 'b) list, err ) Stdlib.result

val sort_env : Ast.expr -> (string * (('a * 'b * (string * 'c) list) * ('d * 'e))) list -> int -> ( (string * (('a * 'b * (string * 'c) list) * ('d * 'e))) list, err ) Stdlib.result

"Failwith" in this function is written to get rid of the warning. The code won't get here. In the case of "IncorrectType", the program will stop using "Result.error". But this will not happen either, since only 3 types described below are supported at the parsing stage.

val interp_ret : Ast.expr list -> Ast.orderby option -> 'a -> (string * elm_data) list -> ( 'a * (string * elm_data) list, err ) Stdlib.result

val interp_match : Ast.elm list -> (Graph.t * (string * Graph.vertex) list) -> Ast.cmdmatch list -> Ast.cmdwithmatch option -> ( Graph.t * (string * Graph.vertex) list, err ) Stdlib.result

val interpret_command : (Graph.t * (string * elm_data) list) -> Ast.command -> ( Graph.t * (string * elm_data) list, err ) Stdlib.result

val interpret_program : Ast.command list -> ( Graph.t * 'a list, err ) Stdlib.result