Module Ast

type id = string

val pp_id : Ppx_deriving_runtime.Format.formatter -> id -> Ppx_deriving_runtime.unit

val show_id : id -> Ppx_deriving_runtime.string

type const =

| Int of int
(*
Integer constants, e.g. -1, 2, 0.
*)

| Str of string
(*
String constants, e.g. {|meow|}, {|miniML|}.
*)

| Bool of bool
(*
Boolean constants, e.g. true, false.
*)

| Unit
(*
()
*)

Represents constant values.

val pp_const : Ppx_deriving_runtime.Format.formatter -> const -> Ppx_deriving_runtime.unit

val show_const : const -> Ppx_deriving_runtime.string

type typ =

| TypInt
(*
Integer type, e.g. int.
*)

| TypStr
(*
String type, e.g. string.
*)

| TypBool
(*
Boolean type, e.g. bool.
*)

| TypUnit
(*
Unit type, e.g. unit.
*)

| TypList of typ
(*
List type, e.g. int list, string list.
*)

Represents basic types in the language.

val pp_typ : Ppx_deriving_runtime.Format.formatter -> typ -> Ppx_deriving_runtime.unit

val show_typ : typ -> Ppx_deriving_runtime.string

type binop =

| Mul
(*
Multiplication, e.g. a * b.
*)

| Div
(*
Division, e.g. a / b.
*)

| Add
(*
Addition, e.g. a + b.
*)

| Sub
(*
Subtraction, e.g. a - b.
*)

| Eq
(*
Equality comparison, e.g. a = b.
*)

| Ne
(*
Inequality comparison, e.g. a <> b.
*)

| Lt
(*
Less-than comparison, e.g. a < b.
*)

| Le
(*
Less-than-or-equal comparison, e.g. a <= b.
*)

| Gt
(*
Greater-than comparison, e.g. a > b.
*)

| Ge
(*
Greater-than-or-equal comparison, e.g. a >= b.
*)

| And
(*
Logical AND, e.g. a && b.
*)

| Or
(*
Logical OR, e.g. a || b.
*)

Represents binary operators, such as arithmetic or logical operations.

val pp_binop : Ppx_deriving_runtime.Format.formatter -> binop -> Ppx_deriving_runtime.unit

val show_binop : binop -> Ppx_deriving_runtime.string

type pat =

| PatConst of const
(*
Matches a constant value, e.g. 42, true.
*)

| PatVar of id
(*
Matches any value and binds it to a variable, e.g. x.
*)

| PatAny
(*
Matches any value without binding it, e.g. _.
*)

| PatTup of pat * pat * pat list
(*
Matches tuples, e.g. (x, y), (a, b, c).
*)

Represents patterns for matching values in expressions.

val pp_pat : Ppx_deriving_runtime.Format.formatter -> pat -> Ppx_deriving_runtime.unit

val show_pat : pat -> Ppx_deriving_runtime.string

type rec_state =

| Rec
(*
Recursive binding, e.g. let rec fact = ....
*)

| NonRec
(*
Non-recursive binding, e.g. let x = ....
*)

Indicates whether a let binding is recursive or non-recursive.

val pp_rec_state : Ppx_deriving_runtime.Format.formatter -> rec_state -> Ppx_deriving_runtime.unit

val show_rec_state : rec_state -> Ppx_deriving_runtime.string

type expr =

| Var of id
(*
Variable reference, e.g. x, my_var.
*)

| Const of const
(*
Constant value, e.g. 42, "hello", true.
*)

| BinOp of binop * expr * expr
(*
Binary operation, e.g. x + y, a >= b.
*)

| Let of let_binding * expr
(*
let expression, e.g. let x = 5 in e.
*)

| App of expr * expr
(*
Function application, e.g. e1 e2, (fun x -> x) 42.
*)

| Fun of pat * expr
(*
Function definition, e.g. fun p -> e.
*)

| Branch of expr * expr * expr
(*
Conditional expression, e.g. if e1 then e2 else e3.
*)

| Tup of expr * expr * expr list
(*
Tuple expression, e.g. (e1, e2), (x, y, z).
*)

| List of expr list
(*
List expression, e.g. [], [e1; e2; e3].
*)

| Match of expr * match_case * match_case list
(*
Pattern matching, e.g. match x with | 0 -> "zero" | _ -> "nonzero".
*)

Represents expressions in the language.

and let_binding = {

is_rec : rec_state;
(*
Whether the binding is recursive or non-recursive.
*)

pat : pat;
(*
The pattern being bound, e.g. x, (a, b).
*)

expr : expr;
(*
The expression being assigned, e.g. 42, fun x -> x + 1.
*)

}

Represents a binding in a let expression.

and match_case = {

match_pat : pat;
(*
The pattern to match, e.g. 0, _, (x, y).
*)

match_expr : expr;
(*
The expression to evaluate if the pattern matches.
*)

}

Represents a single case in a match expression.

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

val pp_let_binding : Ppx_deriving_runtime.Format.formatter -> let_binding -> Ppx_deriving_runtime.unit

val show_let_binding : let_binding -> Ppx_deriving_runtime.string

val pp_match_case : Ppx_deriving_runtime.Format.formatter -> match_case -> Ppx_deriving_runtime.unit

val show_match_case : match_case -> Ppx_deriving_runtime.string

type top_level_item =

| EvalExpr of expr
(*
An expression to be evaluated but not bound, e.g. 1 + 2, print_endline "Hi".
*)

| Binding of let_binding
(*
A value or function binding, e.g. let x = 5, let rec fact n = ....
*)

Represents a top-level item in a program.

val pp_top_level_item : Ppx_deriving_runtime.Format.formatter -> top_level_item -> Ppx_deriving_runtime.unit

val show_top_level_item : top_level_item -> Ppx_deriving_runtime.string

type program = top_level_item list

Represents an entire program as a list of top-level items.

val pp_program : Ppx_deriving_runtime.Format.formatter -> program -> Ppx_deriving_runtime.unit

val show_program : program -> Ppx_deriving_runtime.string