Module Forest.Ast

Copyright 2024, Victoria Lutsyuk

SPDX-License-Identifier: MIT

type id = string

Expression identifier

val pp_id : Ppx_deriving_runtime.Format.formatter -> id -> Ppx_deriving_runtime.unit

val show_id : id -> Ppx_deriving_runtime.string

type const =

| Int of int
(*
Integer constants, e.g. -1, 2, 0.
*)

| Str of string
(*
String constants, e.g. {|meow|}, "miniML".
*)

| Bool of bool
(*
Boolean constants, e.g. true, false.
*)

| Unit
(*
()
*)

Represents constant values.

val pp_const : Ppx_deriving_runtime.Format.formatter -> const -> Ppx_deriving_runtime.unit

val show_const : const -> Ppx_deriving_runtime.string

type binop =

| Mul
(*
Multiplication, e.g. a * b.
*)

| Div
(*
Division, e.g. a / b.
*)

| Add
(*
Addition, e.g. a + b.
*)

| Sub
(*
Subtraction, e.g. a - b.
*)

| Eq
(*
Equality comparison, e.g. a = b.
*)

| Ne
(*
Inequality comparison, e.g. a <> b.
*)

| Lt
(*
Less-than comparison, e.g. a < b.
*)

| Le
(*
Less-than-or-equal comparison, e.g. a <= b.
*)

| Gt
(*
Greater-than comparison, e.g. a > b.
*)

| Ge
(*
Greater-than-or-equal comparison, e.g. a >= b.
*)

| And
(*
Logical AND, e.g. a && b.
*)

| Or
(*
Logical OR, e.g. a || b.
*)

Represents binary operators, such as arithmetic or logical operations.

val pp_binop : Ppx_deriving_runtime.Format.formatter -> binop -> Ppx_deriving_runtime.unit

val show_binop : binop -> Ppx_deriving_runtime.string

type pat =

| PatAny
(*
Matches any value without binding it, e.g. _.
*)

| PatConst of const
(*
Matches a constant value, e.g. 42, true.
*)

| PatListCons of pat * pat
(*
Matches a cons pattern for lists, e.g. x::y.
*)

| PatVar of id
(*
Matches any value and binds it to a variable, e.g. x.
*)

| PatTup of pat * pat * pat list
(*
Matches tuples, e.g. x, y, a, b, c.
*)

| PatList of pat list

| PatType of pat * TypesTree.typ

Represents patterns for matching values in expressions.

val pp_pat : Ppx_deriving_runtime.Format.formatter -> pat -> Ppx_deriving_runtime.unit

val show_pat : pat -> Ppx_deriving_runtime.string

type unary =

| Minus

| Plus

| Not

val pp_unary : Ppx_deriving_runtime.Format.formatter -> unary -> Ppx_deriving_runtime.unit

val show_unary : unary -> Ppx_deriving_runtime.string

type rec_state =

| Rec
(*
Recursive binding, e.g. let rec fact = ....
*)

| NonRec
(*
Non-recursive binding, e.g. let x = ....
*)

Indicates whether a let binding is recursive or non-recursive.

val pp_rec_state : Ppx_deriving_runtime.Format.formatter -> rec_state -> Ppx_deriving_runtime.unit

val show_rec_state : rec_state -> Ppx_deriving_runtime.string

type expr =

| Var of id
(*
Variable reference, e.g. x, my_var.
*)

| Const of const
(*
Constant value, e.g. 42, "hello", true.
*)

| Unary of unary * expr

| BinOp of binop * expr * expr
(*
Binary operation, e.g. x + y, a >= b.
*)

| Option of expr option
(*
Options, e.g. Some("meow"), None.
*)

| Let of rec_state * let_binding * let_binding list * expr
(*
let expression, e.g. let x = 5 in e
*)

| App of expr * expr
(*
Function application, e.g. e1 e2, (fun x -> x) 42.
*)

| Fun of pat * pat list * expr
(*
Function definition, e.g. fun p -> e.
*)

| Branch of expr * expr * expr
(*
Conditional expression, e.g. if e1 then e2 else e3.
*)

| Tup of expr * expr * expr list
(*
Tuple expression, e.g. (e1, e2), (x, y, z).
*)

| List of expr list
(*
List expression, e.g. [], [e1; e2; e3].
*)

| Type of expr * TypesTree.typ

Represents expressions in the language.

and let_binding =

| Binding of pat * expr
(*
The pattern being bound, e.g. x, (a, b). The expression being assigned, e.g. 42, fun x -> x + 1.
*)

Represents a binding in a let expression.

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

val pp_let_binding : Ppx_deriving_runtime.Format.formatter -> let_binding -> Ppx_deriving_runtime.unit

val show_let_binding : let_binding -> Ppx_deriving_runtime.string

type structure_item =

| Eval of expr
(*
An expression to be evaluated but not bound, e.g. 1 + 2, print_endline "Hi".
*)

| Value of rec_state * let_binding * let_binding list
(*
A value or function binding, e.g. let x = 5, let rec fact n = ....
*)

Represents a top-level item in a program.

val pp_structure_item : Ppx_deriving_runtime.Format.formatter -> structure_item -> Ppx_deriving_runtime.unit

val show_structure_item : structure_item -> Ppx_deriving_runtime.string

type program = structure_item list

Represents an entire program as a list of top-level items.

val pp_program : Ppx_deriving_runtime.Format.formatter -> program -> Ppx_deriving_runtime.unit

val show_program : program -> Ppx_deriving_runtime.string