Module EMigunova_lib.Inference

Copyright 2025, Migunova Anastasia

SPDX-License-Identifier: LGPL-3.0-or-later

type error = [

| `No_variable_rec
(*
Represents an error where a recursive variable is not allowed because that would lead to infinite recursion. E.g. let rec x = x + 1
*)

| `No_arg_rec
(*
Represents an error where the left-hand side of the recursive binding is not a var. E.g. let rec [ a; b ] = ..
*)

| `Bound_several_times of string
(*
Represents an error where a pattern bound the variable multiple times. E.g. let x, x = ..
*)

| `Occurs_check of string * Ast.ttype
(*
Represents an occurs check failure. This occurs when attempting to unify types, and one type is found to occur within another in a way that violates the rules of type systems. E.g. let rec f x = f
*)

| `No_variable of string
(*
Represents an error indicating that a variable could not be found in the current scope.
*)

| `Unification_failed of string * Ast.ttype * Ast.ttype
(*
Represents that type unification has failed. This occurs when two types cannot made equivalent during type inference.
*)

]

module VarSet : sig ... end

type scheme =

| Scheme of VarSet.t * Ast.ttype

module TypeEnv : sig ... end

val env_with_print_funs : TypeEnv.t

val print_error : error -> unit

val print_type : Ast.ttype -> unit

val run_inferencer : Ast.let_binding list -> ((Ast.ident * Ast.ttype) list, error) Stdlib.result