Module Parse.Common

Copyright 2024, Vlasenco Daniel and Kudrya Alexandr

SPDX-License-Identifier: MIT

This file contains parsers for part of F# 4.1 grammar, taken from https://fsharp.org/specs/language-spec/4.1/FSharpSpec-4.1-latest.pdf, page 292

val is_whitespace : char -> bool

' ', '\n' or '\r'.

val is_builtin_op : string -> bool

returns true if operator is builtin, false otherwise

val skip_ws : unit Angstrom.t

skip_ws skips 0 or more whitespaces, newlines included

val skip_ws_nl : unit Angstrom.t

skip_ws_nl skips 0 or more whitespaces, fails if '\n' wasn't met

val skip_ws_no_nl : unit Angstrom.t

skip_ws_no_nl skips 0 or more whitespaces, stops when '\n' is met

val skip_ws1 : unit Angstrom.t

skip_ws1 skips 1 or more whitespaces.

val skip_token : string -> unit Angstrom.t

skip_token str skips str exactly, separated with 0 or more whitespaces before and after it.

val is_id_char : char -> bool

is_id_char returns true if char can be used in an identificator.

val is_id_stchar : char -> bool

is_id_stchar returns true if char can be used as a first char of an identificator.

val pid : string Angstrom.t

pid parses identificator and returns it.

val pid_or_op : string Angstrom.t

pid_or_op parses identificator or an operator inside parentheses and returns it.

val pchar : char Angstrom.t

pchar parses char with single quotes and returns it. Cannot parse surrouning whitespaces.

val pstring : string Angstrom.t

pstring parses string with double quotes and returns it. Cannot parse surrouning whitespaces.

val pbool : bool Angstrom.t

pbool parses bool and returns it. Cannot parse surrouning whitespaces.

val pint : int Angstrom.t

pint parses unsigned int and returns it. Should be used for expression parsing. Cannot parse surrouning whitespaces.

val psint : int Angstrom.t

psint parses signed int and returns it. Should be used for pattern parsing. Cannot parse surrouning whitespaces.

val pfloat : float Angstrom.t

pfloat parses unsigned float and returns it. Should be used for expression parsing. Cannot parse surrouning whitespaces.

val psfloat : float Angstrom.t

psfloat parses signed float and returns it. Should be used for pattern parsing. Cannot parse surrouning whitespaces.

val chainl : 'a Angstrom.t -> ('a -> 'a -> 'a) Angstrom.t -> 'a Angstrom.t

chainl pa psep runs pa's with psep's in betweens, applying psep's returned functions left-associatively.

val chainr : 'a Angstrom.t -> ('a -> 'a -> 'a) Angstrom.t -> 'a Angstrom.t

chainr pa psep runs pa's with psep's in betweens, applying psep's returned functions right-associatively.