Module Parse.Units_of_measure

Copyright 2024, Vlasenco Daniel and Kudrya Alexandr

SPDX-License-Identifier: MIT

This file contains parsers for part of F# 4.1 grammar, taken from https://fsharp.org/specs/language-spec/4.1/FSharpSpec-4.1-latest.pdf, page 309

val pm_num : Ast.measure_num Angstrom.t

pm_num parses number before unit of measure and returns it.

val pm : Ast.measure Angstrom.t

pm parses identifiers, products, quotients and powers as measures.

val puom : Ast.unit_of_measure Angstrom.t

puom parses int or float number, then parses a measure in angular brackets and returns them as a unit of measure.