Module Filichkin_lib.Ast

type binop =

| Plus
(*
Addition: `+`
*)

| Minus
(*
Subtraction: `-`
*)

| Mult
(*
Multiplication: `*`
*)

| Div
(*
Division: `/`
*)

| Equal
(*
Equality: `=`
*)

| More
(*
Greater than: `>`
*)

| Less
(*
Less than: `<`
*)

| EMore
(*
Greater than or equal: `>=`
*)

| ELess
(*
Less than or equal: `<=`
*)

type rec_flag =

| NonRec
(*
Non-recursive binding
*)

| Rec
(*
Recursive binding
*)

type expr =

| Int of int
(*
Integer literal
*)

| Var of string
(*
Variable name
*)

| BinOp of binop * expr * expr
(*
Binary operation: operator, left and right operands
*)

| If of expr * expr * expr option
(*
Conditional expression: condition, then branch, optional else branch
*)

| Let of rec_flag * string * expr * expr option
(*
Let expression: recursiveness flag, name, bound expression, body (optional) The last parameter is the optional expression body (for parsing convenience)
*)

| Abs of expr * expr
(*
Abstraction (lambda function): parameter list, function body Syntactic sugar for functions with multiple arguments
*)

| App of expr * expr
(*
Application (function call): function, argument
*)