Module GObject_introspection.Conversions

type carray_of_strings = char Ctypes.ptr Ctypes.ptr

C pointer of null terminated array of C strings

val carray_of_strings : carray_of_strings Ctypes.typ

val carray_of_strings_to_list : char Ctypes.ptr Ctypes.ptr -> string list

Converts C array of strings to OCaml list of strings

val carray_of_strings_to_array : carray_of_strings -> string array

Converts C array of strings to OCaml array of strings

type glist

GList struct

val glist : glist Ctypes.structure Ctypes.typ

val glist_data : (unit Ctypes_static.ptr, glist Ctypes.structure) Ctypes.field

val glist_next : (glist Ctypes.structure Ctypes_static.ptr option, glist Ctypes.structure) Ctypes.field

val glist_prev : (glist Ctypes.structure Ctypes_static.ptr option, glist Ctypes.structure) Ctypes.field

val g_free : unit Ctypes_static.ptr -> unit

val g_free_t : (unit Ctypes_static.ptr -> unit) Ctypes.fn

val glist_free_full : glist Ctypes.structure Ctypes_static.ptr -> (unit Ctypes_static.ptr -> unit) -> unit

val g_list_next : (glist, [ `Struct ]) Ctypes.structured Ctypes.ptr -> glist Ctypes.structure Ctypes_static.ptr option

Get the next element of a glist

val g_list_data : (glist, [ `Struct ]) Ctypes.structured Ctypes.ptr -> unit Ctypes_static.ptr

Get the void ptr data of the current element

val glist_of_strings_to_list : (glist, [ `Struct ]) Ctypes.structured Ctypes.ptr -> string list

Transform a GList of strings to an OCaml list of strings

type gslist

GSList struct

val gslist : gslist Ctypes.structure Ctypes.typ

val gslist_data : (unit Ctypes_static.ptr, gslist Ctypes.structure) Ctypes.field

val gslist_next : (gslist Ctypes.structure Ctypes_static.ptr option, gslist Ctypes.structure) Ctypes.field

val g_slist_next : (gslist, [ `Struct ]) Ctypes.structured Ctypes.ptr -> gslist Ctypes.structure Ctypes_static.ptr option

Get the next element of a gslist

val g_slist_data : (gslist, [ `Struct ]) Ctypes.structured Ctypes.ptr -> unit Ctypes_static.ptr

Get the void ptr data of the current element

val gslist_of_strings_to_list : (gslist, [ `Struct ]) Ctypes.structured Ctypes.ptr -> string list

Transform a GSList of strings to an OCaml list of strings