Module GObject_introspection.Enum_info

type t

A Enum_info represents an enumeration and a Value_info struct represents a value of an enumeration. The Enum_info contains a set of values and a type The Value_info is fetched by calling Enum_info.get_value on a Enum_info.

val enuminfo : t Ctypes.structure Ctypes.typ

val get_n_values : t Ctypes.structure Ctypes.ptr -> int

Obtain the number of values this enumeration contains.

val get_n_methods : t Ctypes.structure Ctypes.ptr -> int

Obtain the number of methods that this enum type has.

val get_method : t Ctypes.structure Ctypes.ptr -> int -> Function_info.t Ctypes.structure Ctypes.ptr

Obtain an enum type method at index n .

val get_value : t Ctypes.structure Ctypes.ptr -> int -> Value_info.t Ctypes.structure Ctypes.ptr option

Obtain a value for this enumeration.

val get_error_domain : t Ctypes.structure Ctypes.ptr -> string option

Obtain the string form of the quark for the error domain associated with this enum, if any.

val get_storage_type : t Ctypes.structure Ctypes.ptr -> Bindings.Types.tag

Obtain the tag of the type used for the enum in the C ABI. This will be a signed or unsigned integral type. Note that in the current implementation the width of the type is computed correctly, but the signed or unsigned nature of the type may not match the sign of the type used by the C compiler.

val cast_from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes base info to enum info.

val cast_to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes enum info to base info

val from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Enum_info.t from a Base_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Return a Base_info.t form a Enum_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val cast_from_registeredtypeinfo : Registered_type_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes registeredtype info to enum info.

val cast_to_registeredtypeinfo : t Ctypes.structure Ctypes.ptr -> Registered_type_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes enum info to registeredtype info

val from_registeredtypeinfo : Registered_type_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Enum_info.t from a Registered_type_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Registered_type_info.registeredtypeinfo_unref.

val to_registeredtypeinfo : t Ctypes.structure Ctypes.ptr -> Registered_type_info.t Ctypes.structure Ctypes.ptr

Return a Registered_type_info.t form a Enum_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Registered_type_info.registeredtypeinfo_unref.