Module GObject_introspection.Field_info

type t

A Field_info struct represents a field of a struct (see Struct_info), union (see Union_info) or an object (see Object_info). The Field_info is fetched by calling Struct_info.get_field, Union_info.get_field or Object_info.get_field. A field has a size, type and a struct offset asssociated and a set of flags, which is currently GI_FIELD_IS_READABLE or GI_FIELD_IS_WRITABLE.

val fieldinfo : t Ctypes.structure Ctypes.typ

val get_flags : t Ctypes.structure Ctypes.ptr -> Bindings.Field_info.flags list

Obtain the flags for this Field_info. See Field_info.flags for possible flag values.

val get_offset : t Ctypes.structure Ctypes.ptr -> int

Obtain the offset in bits of the field member, this is relative to the beginning of the struct or union.

val get_size : t Ctypes.structure Ctypes.ptr -> int

Obtain the size in bits of the field member, this is how much space you need to allocate to store the field.

val get_type : t Ctypes.structure Ctypes.ptr -> Type_info.t Ctypes.structure Ctypes.ptr

Obtain the type of a field as a Type_info.

val cast_from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes base info to field info.

val cast_to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes field info to base info

val add_unref_finaliser : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Add unref of the C underlying structure whith Gc.finalise.

val from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Field_info.t from a Base_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Return a Base_info.t from a Field_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.