Module GObject_introspection.Property_info

type t

Property_info represents a property. A property belongs to either a Object_info or a Interface_info.

val propertyinfo : t Ctypes.structure Ctypes.typ

val get_ownership_transfer : t Ctypes.structure Ctypes.ptr -> Bindings.Arg_info.transfer

Obtain the ownership transfer for this property. See GITransfer for more information about transfer values.

val get_type : t Ctypes.structure Ctypes.ptr -> Type_info.t Ctypes.structure Ctypes.ptr

Obtain the type information for the property info .

val get_flags : t Ctypes.structure Ctypes.ptr -> Bindings.GParam.flags list

Obtain the flags for this property info. See GParamFlags for more information about possible flag values.

val cast_from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes base info to property info.

val cast_to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes property info to base info

val add_unref_finaliser : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Add unref of the C underlying structure whith Gc.finalise.

val from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Property_info.t from a Base_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Return a Base_info.t from a Property_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.