Module GObject_introspection.Registered_type_info

type t

val registeredtypeinfo : t Ctypes.structure Ctypes.typ

val get_type_name : t Ctypes.structure Ctypes.ptr -> string option

Obtain the type name of the struct within the GObject type system. This type can be passed to g_type_name() to get a #GType.

val get_g_type : t Ctypes.structure Ctypes.ptr -> Repository.gtype option

Obtain the gtype for this registered type or None which a special meaning. It means that either there is no type information associated with this info or that the shared library which provides the type_init function for this info cannot be called.

val get_type_init : t Ctypes.structure Ctypes.ptr -> string option

Obtain the type init function for info . The type init function is the function which will register the GType within the GObject type system. Usually this is not called by langauge bindings or applications, use Registered_type_info.get_g_type directly instead.

val cast_from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes base info to registeredtype info.

val cast_to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes registeredtype info to base info

val add_unref_finaliser : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Add unref of the C underlying structure whith Gc.finalise.

val from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Registered_type_info.t from a Base_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Return a Base_info.t from a Registered_type_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.