Module GObject_introspection.Signal_info

type t

Signal_info represents a signal. It's a sub-struct of Callable_info and contains a set of flags and a class closure. See Callable_info for information on how to retreive arguments and other metadata from the signal.

val signalinfo : t Ctypes.structure Ctypes.typ

val true_stops_emit : t Ctypes.structure Ctypes.ptr -> bool

Obtain if the returning true in the signal handler will stop the emission of the signal.

val get_flags : t Ctypes.structure Ctypes.ptr -> Bindings.GSignal.flags list

Obtain the flags for this signal info. See GSignalFlags for more information about possible flag values.

val get_class_closure : t Ctypes.structure Ctypes.ptr -> Callable_info.t Ctypes.structure Ctypes.ptr option

Obtain the class closure for this signal if one is set. The class closure is a virtual function on the type that the signal belongs to. If the signal lacks a closure None will be returned. In order to avoid circular call graph between Signal_info and VFunc_info, this function will return a Callable_info. It is upto the user to use GIVSignalInfo.from_callableinfo in order to have the VFunc_info.

val cast_from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes base info to signal info.

val cast_to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes signal info to base info

val add_unref_finaliser : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Add unref of the C underlying structure whith Gc.finalise.

val from_baseinfo : Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Signal_info.t from a Base_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_baseinfo : t Ctypes.structure Ctypes.ptr -> Base_info.t Ctypes.structure Ctypes.ptr

Return a Base_info.t from a Signal_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val cast_from_callableinfo : Callable_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes callable info to signal info.

val cast_to_callableinfo : t Ctypes.structure Ctypes.ptr -> Callable_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes signal info to callable info

val from_callableinfo : Callable_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Signal_info.t from a Callable_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_callableinfo : t Ctypes.structure Ctypes.ptr -> Callable_info.t Ctypes.structure Ctypes.ptr

Return a Callable_info.t from a Signal_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.