Module GObject_introspection__Base_info

type t

Base_info is the common base struct of all other *Info structs accessible through the Repository API. All other structs can be casted to a Base_info

val baseinfo : t Ctypes.structure Ctypes.typ

val base_info_ref : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Increases the reference count of underlying Base_info *info.

val base_info_unref : t Ctypes.structure Ctypes.ptr -> unit

Decreases the reference count of underlying Base_info *info . When its reference count drops to 0, the info is freed.

val get_name : t Ctypes.structure Ctypes.ptr -> string option

Obtain the name of the info . What the name represents depends on the GIInfoType of the info . For instance for Function_info it is the name of the function.

val equal : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr -> bool

Compare two Base_info. Using pointer comparison is not practical since many functions return different instances of Base_info that refers to the same part of the TypeLib; use this function instead to do Base_info comparisons.

val get_namespace : t Ctypes.structure Ctypes.ptr -> string

Obtain the namespace of info

val is_deprecated : t Ctypes.structure Ctypes.ptr -> bool

Obtain whether the info is represents a metadata which is deprecated or not.

val get_container : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr option

Obtain the container of the info . The container is the parent Base_info. For instance, the parent of a Function_info is an Object_info or Interface_info.

val get_type : t Ctypes.structure Ctypes.ptr -> Bindings.Base_info.info_type

Obtain the info type of the Base_info.

val add_unref_finaliser : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr