Module GObject_introspection__Constant_info

Constant_info — Struct representing a constant

type t

Constant_info represents a constant. A constant has a type associated which can be obtained by calling Constant_info.get_type and a value, which can be obtained by calling GIConstant.get_value.

val constantinfo : t Ctypes.structure Ctypes.typ

val get_type : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Type_info.t Ctypes.structure Ctypes.ptr

Obtain the type of the constant as a Type_info.

val cast_from_baseinfo : GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes base info to constant info.

val cast_to_baseinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes constant info to base info

val add_unref_finaliser : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Add unref of the C underlying structure whith Gc.finalise.

val from_baseinfo : GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Constant_info.t from a Base_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_baseinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr

Return a Base_info.t form a Constant_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val get_value : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Types.argument_t Ctypes.union Ctypes.ptr

Obtain the value associated with the Constant_info and store it in the value parameter. argument needs to be allocated before passing it in. The size of the constant value stored in argument will be returned. Free the value with Constant_info.free_value.