Module GObject_introspection__Interface_info

Interface_info — Struct representing a GInterface

type t

Interface_info represents a GInterface type. A GInterface has methods, fields, properties, signals, interfaces, constants, virtual functions and prerequisites.

val interfaceinfo : t Ctypes.structure Ctypes.typ

val get_n_prerequisites : t Ctypes.structure Ctypes.ptr -> int

Obtain the number of prerequisites for this interface type. A prerequisites is another interface that needs to be implemented for interface, similar to an base class for GObjects.

val get_prerequisite : t Ctypes.structure Ctypes.ptr -> int -> GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr

Obtain an interface type prerequisites index n .

val get_n_properties : t Ctypes.structure Ctypes.ptr -> int

Obtain the number of properties that this interface type has.

val get_property : t Ctypes.structure Ctypes.ptr -> int -> GObject_introspection.Property_info.t Ctypes.structure Ctypes.ptr

Obtain an interface type property at index n .

val get_n_methods : t Ctypes.structure Ctypes.ptr -> int

Obtain the number of methods that this interface type has.

val get_method : t Ctypes.structure Ctypes.ptr -> int -> GObject_introspection.Function_info.t Ctypes.structure Ctypes.ptr

Obtain an interface type method at index n .

val find_method : t Ctypes.structure Ctypes.ptr -> string -> GObject_introspection.Function_info.t Ctypes.structure Ctypes.ptr option

Obtain a method of the interface type given a name . NULL will be returned if there's no method available with that name.

val get_n_signals : t Ctypes.structure Ctypes.ptr -> int

Obtain the number of signals that this interface type has.

val get_signal : t Ctypes.structure Ctypes.ptr -> int -> GObject_introspection.Signal_info.t Ctypes.structure Ctypes.ptr

Obtain an interface type signal at index n .

val find_signal : t Ctypes.structure Ctypes.ptr -> string -> GObject_introspection.Signal_info.t Ctypes.structure Ctypes.ptr option

Find a signal of the interface

val get_n_constants : t Ctypes.structure Ctypes.ptr -> int

Obtain the number of constants that this interface type has.

val get_constant : t Ctypes.structure Ctypes.ptr -> int -> GObject_introspection.Constant_info.t Ctypes.structure Ctypes.ptr

Obtain an interface type constant at index n .

val get_iface_struct : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Struct_info.t Ctypes.structure Ctypes.ptr option

Returns the layout C structure associated with this GInterface.

val get_n_vfuncs : t Ctypes.structure Ctypes.ptr -> int

Obtain the number of virtual functions that this interface type has.

val get_vfunc : t Ctypes.structure Ctypes.ptr -> int -> GObject_introspection.VFunc_info.t Ctypes.structure Ctypes.ptr

Obtain an interface type virtual function at index n .

val find_vfunc : t Ctypes.structure Ctypes.ptr -> string -> GObject_introspection.VFunc_info.t Ctypes.structure Ctypes.ptr option

Locate a virtual function slot with name name . See the documentation for g_object_info_find_vfunc() for more information on virtuals.

val cast_from_baseinfo : GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes base info to interface info.

val cast_to_baseinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes interface info to base info

val add_unref_finaliser : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Add unref of the C underlying structure whith Gc.finalise.

val from_baseinfo : GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Interface_info.t from a Base_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_baseinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr

Return a Base_info.t from a Interface_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val cast_from_registeredtypeinfo : GObject_introspection.Registered_type_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes registeredtype info to enum info.

val cast_to_registeredtypeinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Registered_type_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes enum info to registeredtype info

val from_registeredtypeinfo : GObject_introspection.Registered_type_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a Interface_info.t from a Registered_type_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Registered_type_info.registeredtypeinfo_unref.

val to_registeredtypeinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Registered_type_info.t Ctypes.structure Ctypes.ptr

Return a Registered_type_info.t form a Interface_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Registered_type_info.registeredtypeinfo_unref.