Module GObject_introspection__VFunc_info

VFunc_info — Struct representing a virtual function.

type t

GIVfuncInfo represents a virtual function. A property belongs to either a Object_info or a Interface_info.

val vfuncinfo : t Ctypes.structure Ctypes.typ

val get_offset : t Ctypes.structure Ctypes.ptr -> int

Obtain the offset of the function pointer in the class struct. The value 0xFFFF indicates that the struct offset is unknown.

val get_signal : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Callable_info.t Ctypes.structure Ctypes.ptr option

Obtain the signal for the virtual function if one is set. The signal comes from the object or interface to which this virtual function belongs. In order to avoid circular call graph between VFunc_info and Signal_info, this function will return a Callable_info. It is upto the user to use Signal_info.from_callableinfo in order to have the Signal_info.

val get_flags : t Ctypes.structure Ctypes.ptr -> Bindings.VFunc_info.flags list

Obtain the flags for this virtual function info. See VFunc_infoFlags for more information about possible flag values.

val cast_from_baseinfo : GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes base info to vfunc info.

val cast_to_baseinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes vfunc info to base info

val add_unref_finaliser : t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Add unref of the C underlying structure whith Gc.finalise.

val from_baseinfo : GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a VFunc_info.t from a Base_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_baseinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Base_info.t Ctypes.structure Ctypes.ptr

Return a Base_info.t from a VFunc_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val cast_from_callableinfo : GObject_introspection.Callable_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes callable info to vfunc info.

val cast_to_callableinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Callable_info.t Ctypes.structure Ctypes.ptr

Just cast OCaml Ctypes vfunc info to callable info

val from_callableinfo : GObject_introspection.Callable_info.t Ctypes.structure Ctypes.ptr -> t Ctypes.structure Ctypes.ptr

Return a VFunc_info.t from a Callable_info.t, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.

val to_callableinfo : t Ctypes.structure Ctypes.ptr -> GObject_introspection.Callable_info.t Ctypes.structure Ctypes.ptr

Return a Callable_info.t from a VFunc_info, the underlying C structure ref count is increased and the value is Gc.finalis"ed" with Base_info.baseinfo_unref.