Module Build.Players

module Players is ...

module Player : sig ... end

...

module Map : sig ... end

...

Set of Players

include sig ... end

type elt = Player.t

type t = Stdlib__Set.Make(Player).t

val empty : t

val add : elt -> t -> t

val singleton : elt -> t

val remove : elt -> t -> t

val union : t -> t -> t

val inter : t -> t -> t

val disjoint : t -> t -> bool

val diff : t -> t -> t

val cardinal : t -> int

val elements : t -> elt list

val min_elt : t -> elt

val min_elt_opt : t -> elt option

val max_elt : t -> elt

val max_elt_opt : t -> elt option

val choose : t -> elt

val choose_opt : t -> elt option

val find : elt -> t -> elt

val find_opt : elt -> t -> elt option

val find_first : (elt -> bool) -> t -> elt

val find_first_opt : (elt -> bool) -> t -> elt option

val find_last : (elt -> bool) -> t -> elt

val find_last_opt : (elt -> bool) -> t -> elt option

val iter : (elt -> unit) -> t -> unit

val fold : (elt -> 'acc -> 'acc) -> t -> 'acc -> 'acc

val map : (elt -> elt) -> t -> t

val filter : (elt -> bool) -> t -> t

val filter_map : (elt -> elt option) -> t -> t

val partition : (elt -> bool) -> t -> t * t

val split : elt -> t -> t * bool * t

val is_empty : t -> bool

val mem : elt -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val subset : t -> t -> bool

val for_all : (elt -> bool) -> t -> bool

val exists : (elt -> bool) -> t -> bool

val to_list : t -> elt list

val of_list : elt list -> t

val to_seq_from : elt -> t -> elt Stdlib.Seq.t

val to_seq : t -> elt Stdlib.Seq.t

val to_rev_seq : t -> elt Stdlib.Seq.t

val add_seq : elt Stdlib.Seq.t -> t -> t

val of_seq : elt Stdlib.Seq.t -> t

val show : t -> string

val add_role : Roles.Role.t -> t -> t

add_role x ys takes a Roles.Role.t x and makes a fresh Player.t with index equal to cardinal.

val create : Player.t list -> t

Filter Functions

exception NoPlayersWithRole

val kinds : Roles.Role.Kind.t -> t -> t

val aligned : Roles.Role.Alignment.t -> t -> t

val random : ?f:(t -> t) -> t -> Player.t

Player Fields

Status

val alive : t -> t

val dead : t -> t

val poisoned : t -> t

val status : Player.Status.t -> t -> t

Groups

val allied : Player.t -> t -> t

val opposed : Player.t -> t -> t

exception NoPlayerWithRole of Roles.Role.t

val group : Roles.Group.t -> t -> t

val incl_self : (Player.t -> t -> t) -> Player.t -> t -> t

helper function for applying filters that require target to be in resulting set.

Player Neighbours

module Neighbours : sig ... end

val neighbours : ?f:(t -> t) -> Player.t -> t -> Neighbours.t

neighbours x ys returns the neighbours of x in ys, i.e., those indexed either side of x. Note: requires that x be in ys. Note: if there is only one valid neighbour, then both Neighbours.t.left and Neighbours.t.right will refer to the same Player.t. Note: raises NoNeighbour in the event that x would be it's own neighbours.

val allied_neighbours : Player.t -> t -> Neighbours.t

val opposed_neighbours : Player.t -> t -> Neighbours.t

val alive_neighbours : Player.t -> t -> Neighbours.t

val dead_neighbours : Player.t -> t -> Neighbours.t

val exists_role_kind : Roles.Role.Kind.t -> bool Roles.Map.t -> bool

Setup

exception NoNewRole

exception NoOldRole

val assert_can_change : Roles.Role.Kind.t -> bool Roles.Map.t -> exn -> unit

val log_replace_kind : Player.t -> Roles.Role.t -> unit

val replace_player_role_kind : Roles.Role.Kind.t -> Roles.Role.Kind.t -> bool Roles.Map.t -> t -> t -> Player.t

replace_player_role_kind old new map players exclude replaces the role of one player in players that is not in exclude with kind old is replaced with a new role with kind new. We use map to ensure that we don't introduce duplicate roles into the game, and to that we can't re-add a role that has been removed.

val replace_n_player_role_kinds : ?acc:t -> int -> Roles.Role.Kind.t -> Roles.Role.Kind.t -> bool Roles.Map.t -> t -> unit