Module Role.Kind

type t = Kind.t =

| Townsfolk
(*
a Good character with a beneficial ability.
*)

| Outsider
(*
a Good character with a hindering ability.
*)

| Minion
(*
An Evil character with an ability that hinders the Good team.
*)

| Demon
(*
An Evil character with the ability to kill at Night.
*)

| Traveller
(*
May join the game at any time, and may leave the game at any time.
*)

| Fabled
(*
Characters for the Storyteller.
*)

| Loric
(*
Characters for the Storyteller.
*)

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val equal : t -> t -> Ppx_deriving_runtime.bool