Module Jaml_lib.LL_ast

Copyright 2023-2024, Ilya Pankratov, Maxim Drumov

SPDX-License-Identifier: LGPL-2.1-or-later

Almost the same ast as Typedtree. Changes: Arguments in let and let rec are represented by a list of arguments. There are no TFuns, so anonymous functions are not expressible.

type typed_id = string * Ty.ty

Typed function name

type typed_arg =

| Used of string * Ty.ty
(*
Represent usual argument
*)

| Unused of Ty.ty
(*
Represent unused arguments (wildcard for example)
*)

Typed argument to separate the arguments used and no

type typed_binop = Ast.bin_op * Ty.ty

Typed sign of a binary operation

type llexpr =

| LConst of Ast.const * Ty.ty
(*
Typed expression for the constant
*)

| LVar of string * Ty.ty
(*
Typed expression for the variables
*)

| LTuple of llexpr list * Ty.ty
(*
Typed expression for the tuples
*)

| LBinop of typed_binop * llexpr * llexpr
(*
Typed expression for the binary operations
*)

| LApp of llexpr * llexpr * Ty.ty
(*
Typed expression for the function application to the arguments
*)

| LIfThenElse of llexpr * llexpr * llexpr * Ty.ty
(*
Typed expression for condition statement
*)

| LLetIn of typed_id * llexpr * llexpr
(*
Typed expression for let in declaration
*)

| LTake of llexpr * int
(*
A constructor for getting rid of tuples as a matchmaking pattern. By index in the list.
*)

Typed expression type

type llbinding =

| LLet of typed_id * typed_arg list * llexpr
(*
Typed expression for let declaration
*)

| LLetRec of typed_id * typed_arg list * llexpr
(*
Typed expression for let rec declaration
*)

Typed binding type

type llstatements = llbinding list

Typed statements type