Module Monads.CompilerMonad

The monad is actually a large result-state monad, which encapsulates some of complexities of generating an assembler for an x86-64 machine. The monad works accoring to SYSTEM V ABI calling conventions.

See https://refspecs.linuxbase.org/elf/x86_64-abi-0.99.pdf 3.2 Function Calling Sequence for reference.

What parts of calling convenction are handeled:

First six arguments are passed via registers and others arguments are pushed on stack.

Monad do care about stack alignment by 16 before making function call.

Monad do save caller save registers except RAX. Monad user has to save it excplicetly himself.

What parts of calling convenciton are not handled:

Callee save registers do not save.

What benefits are offered by the monad:

Monad encapsulate working with stack. User does not have access to RBS, RSP registers. Also he does not have access to push and pop commands. All stack work is done by monad.

Monad introduce user abstraction -- variable. It can be saved using the command: let* cmd = save_return_value my_var

Monad takes care of the saved variables and prevents the user from accidentally overwriting them.

Monad prohibits the user from exiting a function twice and creating a function twice in a row without using doing function exit.

type register

The type represents registers: rax, rbx, rcx, ...

type label

The type represents assembly labels

type asm_value

The type represent all values from assembly: labels, memory, registers

type command

Represent assembly mnemonic with operands

type error

Returned from monad error

type ('a, 'error) t

val pp_error : Stdlib.Format.formatter -> error -> unit

Print error

val return : 'a -> ('a, 'error) t

val fail : 'error -> ('a, 'error) t

val bind : ('a, 'error) t -> ('a -> ('b, 'error) t) -> ('b, 'error) t

val let* : ('a, 'error) t -> ('a -> ('b, 'error) t) -> ('b, 'error) t

val fresh : string -> (string, 'error) t

val monad_map : 'a list -> f:('a -> ('b, 'c) t) -> ('b list, 'c) t

val run : ('a, 'error) t -> ('a, 'error) Stdlib.result

module R : sig ... end

Registers available for monad user

val empty_cmd : command

Just empty command which does nothing.

val mov_cmd : ?comment:string option -> register -> asm_value -> (command list, 'error) t

Create mov mnemonic with two operands

val add_cmd : ?comment:string option -> register -> asm_value -> (command list, error) t

Create add mnemonic with two operands

val sub_cmd : ?comment:string option -> register -> asm_value -> (command list, error) t

Create sum mnemonic with two operands

val idiv_cmd : ?comment:string option -> asm_value -> (command list, error) t

Create idiv mnemonic with two operands

val imul_cmd : ?comment:string option -> register -> asm_value -> (command list, error) t

Create imul mnemonic with two operands

val xor_cmd : ?comment:string option -> register -> asm_value -> (command list, error) t

Create xor mnemonic with two operands

val and_cmd : ?comment:string option -> register -> asm_value -> (command list, error) t

Create and mnemonic with two operands

val or_cmd : ?comment:string option -> register -> asm_value -> (command list, error) t

Create or mnemonic with two operands

val cmp_cmd : ?comment:string option -> register -> asm_value -> (command list, error) t

Create cmp mnemonic with two operands

val sete_cmd : ?comment:string option -> register -> (command list, error) t

Create setemnemonic with one operand

val setne_cmd : ?comment:string option -> register -> (command list, error) t

Create setne mnemonic with one operand

val setg_cmd : ?comment:string option -> register -> (command list, error) t

Create setg mnemonic with one operand

val setl_cmd : ?comment:string option -> register -> (command list, error) t

Create setl mnemonic with one operand

val setge_cmd : ?comment:string option -> register -> (command list, error) t

Create setge mnemonic with one operand

val setle_cmd : ?comment:string option -> register -> (command list, error) t

Create setle mnemonic with one operand

val jmp_cmd : ?comment:string option -> label -> (command, error) t

Create jmp mnemonic with one operand

val jne_cmd : ?comment:string option -> label -> (command, error) t

Create jne mnemonic with one operand

val je_cmd : ?comment:string option -> label -> (command, error) t

Create je mnemonic with one operand

val cqo_cmd : ?comment:string option -> (command, error) t

val make_comment : string -> (command, 'error) t

Create assembler comment

val add_global : label -> (command, error) t

Create global declaration

val add_extern : label -> (command, error) t

Create extern declaration. All extern functions must be declared using THIS command.

val make_label : string -> (label, 'error) t

Convert string to label type. And add the label to the environment

val make_bool : bool -> (asm_value, error) t

Convert bool value to asm_value

val make_integer : int -> (asm_value, error) t

Convert integer value to asm_value

val label_to_asmvalue : label -> asm_value

Convert label to asm_value

val reg_to_asmvalue : register -> asm_value

Convert register to asm_value

val get_data_from_environment : string -> (command * asm_value, error) t

Used to get variable or saved label

val make_label_cmd : label -> (command, error) t

Create command form label. For example main label can be declared using the command.

val make_func_call : asm_value -> asm_value list -> (command list, error) t

Takes function and her argument and make call. Before call:

All arguments are passed according to SYSTEM V ABI calling convenction.

val make_func_preambula : string -> string list -> (command list, error) t

Make function preambula: save old rsp register, bind all arguments to the stack/registers.

val make_func_ret : (command list, error) t

Make exit from the function

val save_return_value : string -> (command, error) t

The function can be used to bind any value stored in rax to the variable. After binding you don't have to be scared of lossing variable by writing something at rax. Monad will protected binding variable from owerriting and if necessary will place it on stack from register. Binded vatiable can be obtained using get_data_from_environment

val commands_to_str : command list -> string

Converts all commands to string

val pp_commands : Stdlib.Format.formatter -> command list -> unit

Print commands

val commands_to_file : string -> command list -> unit

Save commands to specified file

val start_branch : (unit, 'a) t

Should be called before 'then' or 'else' branch

val end_branch : (unit, 'a) t

Should be called after 'then' or 'else' branch