Module KREML_lib.Ll_ast

Copyright 2023-2024, Anton Kraev and Polina Badreeva

SPDX-License-Identifier: LGPL-3.0-or-later

Simplified ast representation which is the result of lambda lifting stage. The difference with cc_expr is that there are no nested and anonymous functions

type ll_expr =

| LLiteral of Ast.literal

| LIdentifier of Ast.identifier

| LUnaryOp of Ast.unary_op * ll_expr

| LBinaryOp of Ast.binary_op * ll_expr * ll_expr

| LApp of ll_expr * ll_expr

| LIfThenElse of ll_expr * ll_expr * ll_expr

| LLetIn of Ast.identifier * ll_expr * ll_expr

val pp_ll_expr : Ppx_deriving_runtime.Format.formatter -> ll_expr -> Ppx_deriving_runtime.unit

val show_ll_expr : ll_expr -> Ppx_deriving_runtime.string

type ll_binding =

| LVal of Ast.identifier * ll_expr

| LFun of Ast.identifier * Ast.identifier list * ll_expr

val pp_ll_binding : Ppx_deriving_runtime.Format.formatter -> ll_binding -> Ppx_deriving_runtime.unit

val show_ll_binding : ll_binding -> Ppx_deriving_runtime.string