Module Kappa_data_structures

module Base64 : sig ... end

Base64 RFC4648 implementation.

module Bigbuffer : sig ... end

Extensible buffers.

module Buffers : sig ... end

module Cache : sig ... end

module Circular_buffers : sig ... end

module Color : sig ... end

module Crc32 : sig ... end

module DynamicArray : sig ... end

module ExceptionDefn : sig ... end

module Fifo : sig ... end

FIFO

module Fractions : sig ... end

module GenArray : sig ... end

Signature of array only limited by max_int

module Hashed_list : sig ... end

module Infinite_buffers : sig ... end

module IntCollection : sig ... end

Imperative int set

module JsonUtil : sig ... end

Parsing utils

module LargeArray : sig ... end

Implementation of array only limited by max_int

module List_util : sig ... end

module Loc : sig ... end

Module for type Loc.t annotating structured data with the line range * in a file which was used to define it

module Mods : sig ... end

Datastructures' functors instantiation

module NamedDecls : sig ... end

Stores a bunch of stuff the user gave a name to

module Nbr : sig ... end

Kappa numbers (either float, int or int64) and operations on them

module Operator : sig ... end

Compiled algebraic expression

module Option_util : sig ... end

Utilities on option

module Pp : sig ... end

module Pp_html : sig ... end

module Random_tree : sig ... end

module Renaming : sig ... end

Functions from a subset of nat to a subset of nat

module Result_util : sig ... end

module SetMap : sig ... end

Our own implementattion of Set and Map

module Stop : sig ... end

module Tools : sig ... end

Misc utilities

module ValMap : sig ... end