Module Kappa_data_structures.Base64

Base64 RFC4648 implementation.

Base64 is a group of similar binary-to-text encoding schemes that represent binary data in an ASCII string format by translating it into a radix-64 representation. It is specified in RFC 4648.

Release %%VERSION%% - %%PKG_HOMEPAGE%%

val default_alphabet : string

A 64-character string specifying the regular Base64 alphabet.

val uri_safe_alphabet : string

A 64-character string specifying the URI- and filename-safe Base64 alphabet.

val decode : ?alphabet:string -> string -> Bigbuffer.bigstring

decode s decodes the string s that is encoded in Base64 format. Will leave trailing NULLs on the string, padding it out to a multiple of 3 characters. alphabet defaults to default_alphabet.

raises Not_found
if s is not a valid Base64 string.

val encode : ?pad:bool -> ?alphabet:string -> Bigbuffer.bigstring -> string

encode s encodes the string s into base64. If pad is false, no trailing padding is added. pad defaults to true, and alphabet to default_alphabet.