Module Kappa_runtime.Data

type snapshot = {

snapshot_event : int;

snapshot_time : float;

snapshot_agents : (int * Kappa_mixtures.User_graph.connected_component) list;

snapshot_tokens : (string * Kappa_generic_toolset.Nbr.t) array;

}

type din_data = {

din_kind : Kappa_terms.Primitives.din_kind;

din_start : float;

din_hits : int array;

din_fluxs : float array array;

}

type din = {

din_rules : string array;

din_data : din_data;

din_end : float;

}

type file_line = {

file_line_name : string option;

file_line_text : string;

}

type t =

| DIN of string * din

| DeltaActivities of int * (int * (float * float)) list

| Plot of Kappa_generic_toolset.Nbr.t array	(*
Must have length >= 1 (at least T or E)
*)

| Print of file_line

| TraceStep of Trace.step

| Snapshot of string * snapshot

| Log of string

| Species of string * float * Kappa_mixtures.User_graph.connected_component

| Warning of Kappa_generic_toolset.Locality.t option * Stdlib.Format.formatter -> unit

val print_snapshot : ?uuid:int -> Stdlib.Format.formatter -> snapshot -> unit

val print_dot_snapshot : ?uuid:int -> Stdlib.Format.formatter -> snapshot -> unit

val write_snapshot : Bi_outbuf.t -> snapshot -> unit

Output a JSON value of type snapshot.

val string_of_snapshot : ?len:int -> snapshot -> string

Serialize a value of type snapshot into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_snapshot : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> snapshot

Input JSON data of type snapshot.

val snapshot_of_string : string -> snapshot

Deserialize JSON data of type snapshot.

val print_dot_din : ?uuid:int -> Stdlib.Format.formatter -> din -> unit

val print_html_din : Stdlib.Format.formatter -> din -> unit

val write_din : Bi_outbuf.t -> din -> unit

Output a JSON value of type din.

val string_of_din : ?len:int -> din -> string

Serialize a value of type din into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_din : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> din

Input JSON data of type din.

val din_of_string : string -> din

Deserialize JSON data of type din.

type plot = {

plot_legend : string array;

plot_series : float option array list;

}

val add_plot_line : Kappa_generic_toolset.Nbr.t array -> plot -> plot

val init_plot : Kappa_terms.Model.t -> plot

val write_plot : Bi_outbuf.t -> plot -> unit

Output a JSON value of type plot.

val string_of_plot : ?len:int -> plot -> string

Serialize a value of type plot into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_plot : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> plot

Input JSON data of type plot.

val plot_of_string : string -> plot

Deserialize JSON data of type plot.

val print_plot_legend : is_tsv:bool -> Stdlib.Format.formatter -> string array -> unit

val print_plot_line : is_tsv:bool -> ( Stdlib.Format.formatter -> 'a -> unit ) -> Stdlib.Format.formatter -> 'a array -> unit

val export_plot : is_tsv:bool -> plot -> string

val print_initial_inputs : ?uuid:int -> Kappa_terms.Configuration.t -> Kappa_terms.Model.t -> Stdlib.Format.formatter -> (Kappa_terms.Primitives.alg_expr * Kappa_terms.Primitives.elementary_rule) list -> unit

val print_warning : ?pos:Kappa_generic_toolset.Locality.t -> Stdlib.Format.formatter -> ( Stdlib.Format.formatter -> unit ) -> unit