Module Kappa_runtime.Trace

Trace of simulation

module Simulation_info : sig ... end

type event_kind =

| RULE of int

| INIT of int list
(*
the agents
*)

| PERT of string
(*
the rule
*)

val print_event_kind : ?env:Kappa_terms.Model.t -> Stdlib.Format.formatter -> event_kind -> unit

val print_event_kind_dot_annot : Kappa_terms.Model.t -> Stdlib.Format.formatter -> event_kind -> unit

type step =

| Subs of int * int

| Rule of int * Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.event * unit Simulation_info.t

| Pert of string * Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.event * unit Simulation_info.t

| Init of Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.action list

| Obs of string * Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.test list list * unit Simulation_info.t

| Dummy of string

type t = step list

val dummy_step : string -> step

val subs_step : int -> int -> step

val step_is_obs : step -> bool

val step_is_init : step -> bool

val step_is_subs : step -> bool

val step_is_rule : step -> bool

val step_is_pert : step -> bool

val has_creation_of_step : step -> bool

val tests_of_step : step -> Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.test list

val actions_of_step : step -> Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.action list * (Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.site * Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.binding_state) list

returns
(actions, side_effects)

val side_effects_of_step : step -> Kappa_terms.Instantiation.concrete Kappa_terms.Instantiation.site list

val simulation_info_of_step : step -> unit Simulation_info.t option

val creation_of_actions : ('a -> 'b) -> 'a Kappa_terms.Instantiation.action list -> 'b list

val creation_of_step : step -> int list

val print_step : ?compact:bool -> ?env:Kappa_terms.Model.t -> Stdlib.Format.formatter -> step -> unit

val print_label_of_step : ?env:Kappa_terms.Model.t -> Stdlib.Format.formatter -> step -> unit

val step_to_yojson : step -> Yojson.Basic.t

val json_dictionnary : string

val write_step : Stdlib.Buffer.t -> step -> unit

Output a JSON value of type step.

val string_of_step : ?len:int -> step -> string

Serialize a value of type step into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_step : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> step

Input JSON data of type step.

val step_of_string : string -> step

Deserialize JSON data of type step.

val write_json : Stdlib.Buffer.t -> t -> unit

val read_json : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> t

val init_trace_file : uuid:int -> Kappa_terms.Model.t -> Stdlib.out_channel -> unit

val fold_trace : (Kappa_terms.Model.t -> 'a -> step -> 'a) -> (Kappa_terms.Model.t -> 'a) -> Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> Kappa_terms.Model.t * 'a

val fold_trace_file : (Kappa_terms.Model.t -> 'a -> step -> 'a) -> (Kappa_terms.Model.t -> 'a) -> string -> Kappa_terms.Model.t * 'a

val get_headers_from_file : string -> int option * Kappa_terms.Model.t