Module Ast.Ast

type ident = string

val pp_ident : Ppx_deriving_runtime.Format.formatter -> ident -> Ppx_deriving_runtime.unit

val show_ident : ident -> Ppx_deriving_runtime.string

type unop =

| Not	(*
not
*)

| USub	(*

*)

val pp_unop : Ppx_deriving_runtime.Format.formatter -> unop -> Ppx_deriving_runtime.unit

val show_unop : unop -> Ppx_deriving_runtime.string

type logic_binop =

| And	(*
and
*)

| Or	(*
or
*)

val pp_logic_binop : Ppx_deriving_runtime.Format.formatter -> logic_binop -> Ppx_deriving_runtime.unit

val show_logic_binop : logic_binop -> Ppx_deriving_runtime.string

type arithm_binop =

| Add	(*

*)

| Mul	(*
*
*)

| Sub	(*

*)

| Div	(*
/
*)

| Pow	(*
^
*)

val pp_arithm_binop : Ppx_deriving_runtime.Format.formatter -> arithm_binop -> Ppx_deriving_runtime.unit

val show_arithm_binop : arithm_binop -> Ppx_deriving_runtime.string

type compare_binop =

| Le	(*
<=
*)

| Ge	(*
>=
*)

| Lt	(*
>
*)

| Gt	(*
<
*)

| Eq	(*
==
*)

| Ne	(*
~=
*)

val pp_compare_binop : Ppx_deriving_runtime.Format.formatter -> compare_binop -> Ppx_deriving_runtime.unit

val show_compare_binop : compare_binop -> Ppx_deriving_runtime.string

type string_binop =

| Concat	(*
..
*)

val pp_string_binop : Ppx_deriving_runtime.Format.formatter -> string_binop -> Ppx_deriving_runtime.unit

val show_string_binop : string_binop -> Ppx_deriving_runtime.string

type binop =

| LOp of logic_binop	(*
Logic operators: and, or
*)

| AOp of arithm_binop	(*
Arithmetic operators: +. -. /, *, ^
*)

| COp of compare_binop	(*
Comparative operators: <=, >=, <, >, ==, ~=
*)

| SOp of string_binop	(*
String operators: ..
*)

val pp_binop : Ppx_deriving_runtime.Format.formatter -> binop -> Ppx_deriving_runtime.unit

val show_binop : binop -> Ppx_deriving_runtime.string

type block = statement list

and expr_table_entry =

| JustExpr of expr	(*
Just an expressinon, its index will be determined by position
*)

| PairExpr of expr * expr	(*
Key/Value Pair (key=value)
*)

and l_function = ident list * block

and const =

| Bool of bool	(*
true | false
*)

| Number of float	(*
1, 2.3, 3, ...
*)

| String of string	(*
"abc"
*)

| Function of l_function	(*
function(args) block end
*)

| Nil	(*
nil
*)

and expr =

| Const of const	(*
number or string or ..., written in const
*)

| Variable of ident	(*
obv. variable
*)

| TableGet of expr * expr	(*
exprexpr
*)

| TableInit of expr_table_entry list	(*
xpr, expr, expr, ...
*)

| BinOp of binop * expr * expr	(*
expr binop expr
*)

| UnOp of unop * expr	(*
unop expr
*)

| ExprApply of apply	(*
expr(expr, ..., expr)
*)

and apply =

| Call of expr * expr list	(*
expr(expr, ..., expr)
*)

and statement =

| Do of block	(*
do sttmn sttmnt end
*)

| Set of lvalue list * expr list	(*
a, b, c = d, e, f
*)

| While of expr * block	(*
while expr do block
*)

| Repeat of block * expr	(*
repeat block until expr
*)

| If of expr * block * elseif_block list * block option	(*
if expr then block elseif block else block end
*)

| Fornum of ident * expr * expr * expr option * block	(*
for var = expr, expr, ?expr, block
*)

| Forin of ident list * expr list * block	(*
for expr, expr, ... in expr, expr, ... do block
*)

| Local of ident list * expr list	(*
local a = ?expr
*)

| Return of expr option	(*
return expr
*)

| Break	(*
break
*)

| StatementApply of apply	(*
just a fun call, needed because of side effects
*)

| FunctionDeclare of lvalue * ident list * block	(*
function name(args) body end
*)

| Expr of expr	(*
just an expression, will be printed, for REPL
*)

and lvalue =

| Index of lvalue * expr	(*
a123..nnn
*)

| Ident of ident	(*
var
*)

and elseif_block = expr * block

elseif expr then block

val pp_block : Ppx_deriving_runtime.Format.formatter -> block -> Ppx_deriving_runtime.unit

val show_block : block -> Ppx_deriving_runtime.string

val pp_expr_table_entry : Ppx_deriving_runtime.Format.formatter -> expr_table_entry -> Ppx_deriving_runtime.unit

val show_expr_table_entry : expr_table_entry -> Ppx_deriving_runtime.string

val pp_l_function : Ppx_deriving_runtime.Format.formatter -> l_function -> Ppx_deriving_runtime.unit

val show_l_function : l_function -> Ppx_deriving_runtime.string

val pp_const : Ppx_deriving_runtime.Format.formatter -> const -> Ppx_deriving_runtime.unit

val show_const : const -> Ppx_deriving_runtime.string

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

val pp_apply : Ppx_deriving_runtime.Format.formatter -> apply -> Ppx_deriving_runtime.unit

val show_apply : apply -> Ppx_deriving_runtime.string

val pp_statement : Ppx_deriving_runtime.Format.formatter -> statement -> Ppx_deriving_runtime.unit

val show_statement : statement -> Ppx_deriving_runtime.string

val pp_lvalue : Ppx_deriving_runtime.Format.formatter -> lvalue -> Ppx_deriving_runtime.unit

val show_lvalue : lvalue -> Ppx_deriving_runtime.string

val pp_elseif_block : Ppx_deriving_runtime.Format.formatter -> elseif_block -> Ppx_deriving_runtime.unit

val show_elseif_block : elseif_block -> Ppx_deriving_runtime.string

type ast = block

val pp_ast : Ppx_deriving_runtime.Format.formatter -> ast -> Ppx_deriving_runtime.unit

val show_ast : ast -> Ppx_deriving_runtime.string

type t = ast