Module Make_lib.Ast

Copyright 2021-2022, ol-imorozko and contributors

SPDX-License-Identifier: LGPL-3.0-or-later

type substitute =

| Regular of substitute Base.list

| Pattern

| None of Base.string

| Asterisk

| At

| Lesser

val pp_substitute : Ppx_deriving_runtime.Format.formatter -> substitute -> Ppx_deriving_runtime.unit

val show_substitute : substitute -> Ppx_deriving_runtime.string

val regular : substitute Base.list -> substitute

val pattern : substitute

val none : Base.string -> substitute

val asterisk : substitute

val at : substitute

val lesser : substitute

module Variants_of_substitute : sig ... end

val substitute_of_sexp : Base__.Ppx_sexp_conv_lib.Sexp.t -> substitute

val sexp_of_substitute : substitute -> Base__.Ppx_sexp_conv_lib.Sexp.t

val compare_substitute : substitute -> substitute -> Base.int

type word = substitute Base.list

val pp_word : Ppx_deriving_runtime.Format.formatter -> word -> Ppx_deriving_runtime.unit

val show_word : word -> Ppx_deriving_runtime.string

val word_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> word

val sexp_of_word : word -> Ppx_sexp_conv_lib.Sexp.t

val compare_word : word -> word -> Base.int

type var =

| Recursive of word * word Base.list

| Simply of word * word Base.list

| Conditional of word * word Base.list

val pp_var : Ppx_deriving_runtime.Format.formatter -> var -> Ppx_deriving_runtime.unit

val show_var : var -> Ppx_deriving_runtime.string

val recursive : word -> word Base.list -> var

val simply : word -> word Base.list -> var

val conditional : word -> word Base.list -> var

module Variants_of_var : sig ... end

type recipe =

| Echo of word

| Silent of word

val pp_recipe : Ppx_deriving_runtime.Format.formatter -> recipe -> Ppx_deriving_runtime.unit

val show_recipe : recipe -> Ppx_deriving_runtime.string

val echo : word -> recipe

val silent : word -> recipe

module Variants_of_recipe : sig ... end

val recipe_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> recipe

val sexp_of_recipe : recipe -> Ppx_sexp_conv_lib.Sexp.t

type rule = {

targets : word * word Base.list;

prerequisites : word Base.list;

recipes : recipe Base.list;

}

val pp_rule : Ppx_deriving_runtime.Format.formatter -> rule -> Ppx_deriving_runtime.unit

val show_rule : rule -> Ppx_deriving_runtime.string

type expr =

| Rule of rule

| Var of var

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

type ast = (rule * Base.int) * expr Base.list

Ast for the Make syntax

Makefile, essentially, is an expr list. However, is should contain at least one rule. We're ensure it by extracting that rule from the list. But we need to remember it's position in the list as this is important when this rule contains a substitute variable.

val pp_ast : Ppx_deriving_runtime.Format.formatter -> ast -> Ppx_deriving_runtime.unit

val show_ast : ast -> Ppx_deriving_runtime.string