Module Mpd

module Client : sig ... end

Provides functions and type in order to communicate to the mpd server with commands and requests.

module Client_lwt : sig ... end

Provides functions and type in order to communicate to the mpd server with commands and requests in Lwt threads.

module Connection : sig ... end

Offer functions and type in order to handle connections to the mpd server at the socket level.

module Connection_lwt : sig ... end

Offer functions and type in order to handle connections to the mpd server at the socket level in Lwt thread.

module Loggin : sig ... end

module Music_database : sig ... end

Music_database module: regroups data base related commands.

module Music_database_lwt : sig ... end

Music_database module with Lwt: regroups data base related commands.

module Playback : sig ... end

Controlling playback functions. https://www.musicpd.org/doc/protocol/playback_commands.html

module Playback_lwt : sig ... end

Controlling playback functions in Lwt thread. https://www.musicpd.org/doc/protocol/playback_commands.html

module Playback_options : sig ... end

functions that configure all the playbackoptions

module Playback_options_lwt : sig ... end

functions that configure all the playbackoptions in a Lwt thread.

module Protocol : sig ... end

Define the Mpd response and error types

module Queue : sig ... end

Module for Mpd current playlist manipulation.

module Queue_lwt : sig ... end

Module for Mpd current playlist manipulation in Lwt threads.

module Song : sig ... end

Song module used to store and retrieve information of a song based on * mpd tags.

module Status : sig ... end

Status : get informations on the current status of the Mpd server.

module Stored_playlists : sig ... end

Stored_playlists : Playlists are stored inside the configured playlist directory. They are addressed with their file name (without the directory and without the .m3u suffix).

module Stored_playlists_lwt : sig ... end

LwtStoredPlaylists : Playlists are stored inside the configured playlist directory. They are addressed with their file name (without the directory and without the .m3u suffix). This is module is based on Lwt.

module Tags : sig ... end

module Utils : sig ... end

Set of helpers for the mpdlibclient library.