Module Mpd.Stored_playlists

Stored_playlists : Playlists are stored inside the configured playlist directory. They are addressed with their file name (without the directory and without the .m3u suffix).

val listplaylists : Client.t -> (string list, string) Stdlib.result

Print a list of the playlist names.

val load : Client.t -> string -> ?range:(int * int) -> unit -> Protocol.response

Load the playlist into the current queue. Playlist plugins are supported. A range may be specified to load only a part of the playlist.

val playlistadd : Client.t -> string -> string -> Protocol.response

Add URI to the playlist NAME.m3u. NAME.m3u will be created if it does not exist.

val playlistclear : Client.t -> string -> Protocol.response

Clear the playlist NAME.m3u.

val playlistdelete : Client.t -> string -> int -> Protocol.response

Delete SONGPOS from the playlist NAME.m3u.

val playlistmove : Client.t -> string -> int -> int -> Protocol.response

Move the song at position FROM in the playlist NAME.m3u to the position TO.

val rename : Client.t -> string -> string -> Protocol.response

Rename the playlist NAME.m3u to NEW_NAME.m3u.

val rm : Client.t -> string -> Protocol.response

Remove the playlist NAME.m3u from the playlist directory.

val save : Client.t -> string -> Protocol.response

Save the current playlist to NAME.m3u in the playlist directory.