Module OCamlLabeledArgs_lib.Parsetree

Copyright 2022-2023, Denis Porsev and contributors

SPDX-License-Identifier: LGPL-3.0-or-later

The parsed AST representation of our mini language.

Parse tree

type id = string

Built-in string type for variable names

type const =

| Bool of bool	(*
true or false
*)

| Int of int	(*
integer literal
*)

| Unit	(*
()
*)

Type for constants

type bin_op =

| Plus	(*

operator

*)

| Minus	(*

operator

*)

| Mult	(*
* operator
*)

| Divide	(*
/ operator
*)

| Mod	(*
mod (remainder) operator
*)

| Eq	(*
= operator
*)

| Neq	(*
<> operator
*)

| Lt	(*
< operator
*)

| Ltq	(*
<= operator
*)

| Gt	(*
> operator
*)

| Gtq	(*
>= operator
*)

Enumeration of all available binary operators

type arg_label =

| ArgNoLabel	(*
T -> ...
*)

| ArgLabeled of id	(*
label:T -> ...
*)

| ArgOptional of id	(*
?label:T -> ...
*)

Representation of function arguments

type expr =

| Const of const	(*
Constant expressions
*)

| Var of id	(*
Variable names expressions
*)

| Binop of bin_op * expr * expr	(*
Expressions involving binary operations
*)

| Fun of arg_label * expr option * id * expr	(*
Anonymous functions
*)

| App of expr * arg_label * expr	(*
Function application
*)

| IfThenElse of expr * expr * expr	(*
Conditional operator if-then-else
*)

| Let of id * expr * expr	(*
Expression for "let x = e1 in e2"
*)

| LetRec of id * expr * expr	(*
Fixpointed Let
*)

Type for expressions

type definition = id * expr

Type for definitions

val pp_definition : Ppx_deriving_runtime.Format.formatter -> definition -> Ppx_deriving_runtime.unit

val show_definition : definition -> Ppx_deriving_runtime.string

Top-level tree

type command =

| Help	(*
#help command prints a list of all available commands
*)

| Quit	(*
#quit command exits the toplevel loop and terminates this program
*)

| Use of string	(*
#use <file> command reads and evaluates source phrases from the given file
*)

Top-level input

type toplevel =

| Definition of definition	(*
Definition is for representing let expressions in top-level input
*)

| Expression of expr	(*
Expression is for representing expr type in top-level input
*)

| Command of command	(*
Command is for representing available REPL commands in top-level input
*)