Module Infer.TyError

Copyright 2023-2024, Danil S, Andrei

SPDX-License-Identifier: MIT

type t =

| UnificationMismatch
(*
Lists of types to unify have different lengths
*)

| UnificationFailTy of Types.Ty.t * Types.Ty.t
(*
Failed to unify two types
*)

| UnificationFailEff of Types.Eff.t * Types.Eff.t
(*
Failed to unify two effects
*)

| UnboundVariable of Misc.Ident.t
(*
Failed to find a variable in the environment
*)

| OccursInTy of Types.Var.t * Types.Ty.t
(*
Type variable occurs in a type
*)

| OccursInEff of Types.Var.t * Types.Eff.t
(*
Effect variable occurs in an effect
*)

| RecursiveEffRows
(*
Recursive effect rows
*)

| PatVarBoundSeveralTimes of Misc.Ident.t
(*
Pattern(s) bound the same variable several times. E.g. `let x, x = ..`
*)

| VarsMismatchOrPattern of Misc.Ident.t
(*
Sides of | pattern bound different variables. E.g. `Some x | Some y`
*)

| ConstructorArityMismatch of Misc.Ident.t
(*
Constructor is not applied (`Some`), or applied when not needed (`None 1`)
*)

| NotVarLHSRec
(*
The left hand side of the recursive binding is not a var. E.g. `let rec _ = ..`
*)

| NotAllowedRHSRec of Ast.expression
(*
The expression is not allowed as right-hand side of `let rec'. E.g. `let rec x = x + 1`
*)

| UnboundTypeVariable of Base.string
(*
Type variable is unbound in type declaration. E.g. `type bar = Bar of 'a` ('a is unbound)
*)

| UnboundType of Misc.Ident.t
(*
Type is unbound in type declaration. E.g. `type bar = Bar of foo` (foo is unbound)
*)

| TypeArityMismatch of Misc.Ident.t
(*
Type expects more/less arguments than applied. E.g. `type foo = Foo` (foo expects 0 arguments) `type bar = Bar of int foo`
*)

| NotImplemented of Base.string
(*
Feature is not yet implemented in the type checker. The argument contains the feature description
*)

val pp : Base.Format.formatter -> t -> Base.unit