Module Ocamladt_lib.Parser

Copyright 2024-2025, Rodion Suvorov, Mikhail Gavrilenko

SPDX-License-Identifier: LGPL-3.0-or-later

Parses a given input string into a list of structure items.

parameter str
The input string to parse.

returns
An Ok containing the parsed structure or an Error message.

val pident_cap : string Angstrom.t

val pident_lc : string Angstrom.t

val ppatvar : Ast.Pattern.t Angstrom.t

val pletbinding : Ast.Expression.t Angstrom.t -> Ast.Expression.t Ast.Expression.value_binding Angstrom.t

val pifexpr : Ast.Expression.t Angstrom.t -> Ast.Expression.t Angstrom.t

val papplyexpr : (Ast.Expression.t -> Ast.Expression.t -> Ast.Expression.t) Angstrom.t

val pfunexpr : Ast.Expression.t Angstrom.t -> Ast.Expression.t Angstrom.t

val parsebinop : string -> (Ast.Expression.t -> Ast.Expression.t -> Ast.Expression.t) Angstrom.t

val padd : (Ast.Expression.t -> Ast.Expression.t -> Ast.Expression.t) Angstrom.t

val psub : (Ast.Expression.t -> Ast.Expression.t -> Ast.Expression.t) Angstrom.t

val pdiv : (Ast.Expression.t -> Ast.Expression.t -> Ast.Expression.t) Angstrom.t

val pmul : (Ast.Expression.t -> Ast.Expression.t -> Ast.Expression.t) Angstrom.t

val pcompops : (Ast.Expression.t -> Ast.Expression.t -> Ast.Expression.t) Angstrom.t

val plogops : (Ast.Expression.t -> Ast.Expression.t -> Ast.Expression.t) Angstrom.t

val pexpr : Ast.Expression.t Angstrom.t

val pseval : Ast.Structure.structure_item Angstrom.t

val pstr_item : Ast.Structure.structure_item Angstrom.t

val pstructure : Ast.Structure.structure_item list Angstrom.t

val parse : string -> (Ast.Structure.structure_item list, string) Stdlib.result

val parse_str : string -> Ast.Structure.structure_item list