Module OcamlOOP_lib.Typedtree

Copyright 2023, Artem-Rzhankoff

SPDX-License-Identifier: LGPL-3.0-or-later

module VarSet : sig ... end

type binder = int

val equal_binder : binder -> binder -> Ppx_deriving_runtime.bool

val pp_binder : Ppx_deriving_runtime.Format.formatter -> binder -> Ppx_deriving_runtime.unit

val show_binder : binder -> Ppx_deriving_runtime.string

type binder_set = VarSet.t

val pp_binder_set : Ppx_deriving_runtime.Format.formatter -> binder_set -> Ppx_deriving_runtime.unit

val show_binder_set : binder_set -> Ppx_deriving_runtime.string

type constr =

| C of binder * bool

Represents type sharing for objects, which displayed using the as constract. A shared sub-term is represented by a type variable, when object's type is recursive (flag = true) This makes sense since this language doesn't support subtyping

val equal_constr : constr -> constr -> Ppx_deriving_runtime.bool

val pp_constr : Ppx_deriving_runtime.Format.formatter -> constr -> Ppx_deriving_runtime.unit

val show_constr : constr -> Ppx_deriving_runtime.string

type ty =

| TPrim of string
(*
Ground types: int, bool
*)

| TVar of binder
(*
Variables: 'a, 'b
*)

| TTuple of ty list
(*
TTuple [t1;..;tn] ==> (t1 * ... * tn)
*)

| TArrow of ty * ty
(*
TArrow (t1, t2) ==> e1 -> e2
*)

| TList of ty
(*
['a; 'b; ...]
*)

| TObject of ty * (string * ty) list * constr
(*
Represents:

< .. > as 'a when constr is true

< .. > when constr is false

TObject(`f1:t1;..;fn:tn',_,_) ==> < f1:t1; ...; fn:tn> f1, fn are represented a linked list of types using TField and TNil constructors.

The second parameter represents the instance variables.
*)

| TField of string * ty * ty
(*
TField ("foo", t, ts) ==> <..; foo : t; ts>
*)

| TNil
(*
TNil ==> <...;>
*)

| TPoly of binder
(*
Represents anonymous row variable <;..>
*)

val equal_ty : ty -> ty -> Ppx_deriving_runtime.bool

val pp_ty : Ppx_deriving_runtime.Format.formatter -> ty -> Ppx_deriving_runtime.unit

val show_ty : ty -> Ppx_deriving_runtime.string

type scheme =

| S of binder_set * ty

val pp_scheme : Ppx_deriving_runtime.Format.formatter -> scheme -> Ppx_deriving_runtime.unit

val show_scheme : scheme -> Ppx_deriving_runtime.string