Module Engine

Oratio's proof construction engine

This module exposes a langage as well as a generic evaluator for proofs construction programs.

Multiple backends could be used together with this engine to produce indistinclty proof-trees, explanation texts, typed lambda-terms or any other proof objects. Evaluators with custom backends are produced using the Make functor.

The proof construction language

The proof construction language is composed of 12 elementary instructions (see instructions). Instructions are evaluated by a stack machine as follows :

ElimBot p

Pop a thm from the stack, apply elim_bot with p as second argument on it and push the result

ElimImpl

Pop 2 thm from the stack, apply elim_impl on them and push the result

ElimAndL or ElimAndR

Pop a thm from the stack, apply elim_and_l (respectively elim_and_r) on it and push the result

ElimOr

Pop 3 thm from the stack, apply elim_or on them and push the result

IntroBot

Pop 2 thm from the stack, apply intro_bot on them and push the result

IntroImpl p

Pop a thm from the stack, apply intro_impl with p as second argument on it and push the result

IntroAnd

Pop 2 thm from the stack, apply intro_and on them and push the result

IntroOrL p or IntroOrL p

Pop a thm from the stack, apply intro_or_l (respectively intro_or_r) with p as second argument on it and push the result

Axiom l p

Push axiom l p on the stack

module type EVAL_MODEL = sig ... end

Signature of backend modules for the engine

type 'prop instructions =

| IntroBot

| IntroImpl of 'prop

| IntroAnd

| IntroOrL of 'prop

| IntroOrR of 'prop

| ElimBot of 'prop

| ElimImpl

| ElimAndL

| ElimAndR

| ElimOr

| Axiom of 'prop list * 'prop

Proof construction language

module Make : functor (X : EVAL_MODEL) -> sig ... end