Module Tactic

Tactic

Tactics for interative theorem proving.

Each tactic is a function maping an interactive proof environement to another proof environement and generating instructions for the proof construction engine.

The program generated by succesive applications of tactics can then be evaluated using the engine module together with the backend of your choice.

The function qed evaluate the construction program contained in a proof environement with the rules backend for Oratio's engine.

type env = {

ctx : Kernel.Logic.prop list list;

goals : Kernel.Logic.prop list;

proof : Kernel.Logic.prop Engine.instructions list;

}

val add_frame : 'a list list -> 'a list list

val drop_frame : 'a list -> 'a list

val top_frame : 'a list list -> 'a list

val add_hyp : 'a -> 'a list list -> 'a list list

val find : 'a -> 'a list list -> bool

val check : Kernel.Rules.prop -> Kernel.Rules.prop Engine.instructions list -> bool

val init : Kernel.Logic.prop -> env

val apply : Kernel.Logic.prop -> env -> env

val elim : Kernel.Logic.prop -> env -> env

val elimn : int -> env -> env

val left : env -> env

val right : env -> env

val contradiction : Kernel.Logic.prop -> env -> env

val exfalso : env -> env

val intro : env -> env

val assumption : env -> env

val debug : env -> env

val help : 'a -> 'a

val qed : Kernel.Rules.prop -> env -> unit