Module Prolog_lib.Types

Copyright 2021-2022, Ilya Shchuckin

SPDX-License-Identifier: LGPL-3.0-or-later

Defines main types that are used in the interpreter.

type failure =

| Critical of string

| EvalError of string

Critical error forces program to stop it's calculations, while EvalError is the typical failure of goals evaluation.

val failure_to_string : failure -> string

type 'a result =

| Ok of 'a

| Error of failure

module type MonadFail = sig ... end

module Result : MonadFail

type clause = {

head : Ast.term;

goal : Ast.term;

}

val pp_clause : Ppx_deriving_runtime.Format.formatter -> clause -> Ppx_deriving_runtime.unit

val show_clause : clause -> Ppx_deriving_runtime.string

type db = clause list

val pp_db : Ppx_deriving_runtime.Format.formatter -> db -> Ppx_deriving_runtime.unit

val show_db : db -> Ppx_deriving_runtime.string

module type Config = sig ... end

Contains database and execution mode (debug) for the current run.

type unifier = (Ast.term * Ast.term) list

val pp_unifier : Ppx_deriving_runtime.Format.formatter -> unifier -> Ppx_deriving_runtime.unit

val show_unifier : unifier -> Ppx_deriving_runtime.string

type choicepoint =

| Choicepoint of {

goal : Ast.term;

substitution : unifier;

candidates : clause list;

}

| Conjuction of {

goal : Ast.term;

choicepoints : choicepoints;

substitution : unifier;

}

| Clause of {

goal : Ast.term;

substitution : unifier;

candidates : clause list;

head : Ast.term;

body : Ast.term;

}

Defines three possible types of a choicepoint. Type of a choicepoints depends on the predicate while executing which it was created.

and choicepoints = choicepoint list

val pp_choicepoint : Ppx_deriving_runtime.Format.formatter -> choicepoint -> Ppx_deriving_runtime.unit

val show_choicepoint : choicepoint -> Ppx_deriving_runtime.string

val pp_choicepoints : Ppx_deriving_runtime.Format.formatter -> choicepoints -> Ppx_deriving_runtime.unit

val show_choicepoints : choicepoints -> Ppx_deriving_runtime.string

type interp_res =

| InterpretationResult of unifier * ( unit -> interp_res result ) option

Consists of substitution and optionally a function that starts backtracking.