Module Prolog_lib.Utils

Copyright 2021-2022, Ilya Shchuckin

SPDX-License-Identifier: LGPL-3.0-or-later

val get_pi : Ast.term -> string * int

get_pi term returns predicate indicator (7.1.6.6 in ISO) of term.

val str_of_pi : (string * int) -> string

str_of_pi pi converts predicate indicator pi to string.

val get_vars_from_term : Ast.term -> Ast.term list

get_vars_from_term term returns a list of variables that occur in term.

val apply_substitution : Ast.term -> (Ast.term * Ast.term) list -> Ast.term

apply_substitution term substitution replaces variables in term with matching substitions from substitution.

val is_empty : 'a list -> bool

Check whether a list is empty.