Module Qcheck_test.Run

val run : int -> (Stdlib.Format.formatter -> Ast.program -> unit) -> bool -> unit

Runs a QuickCheck tests.

parameter n
Number of generated Ast.

parameter printer
Printer that formats generated Ast on a failure case.

parameter dparse
Don't check anything, only generate Ast and print it.