Parameter MinMaxAB.1-X

include Spec.SPEC

include Spec.UNIT_SPEC

type state

Type of the states

type action

Type of the actions

val init : state

Initial states

val goal : state -> bool

Test if a state is a goal state

val actions : state -> action list

Compute the actions available in a state

val apply : state -> action -> state

Apply an action to a state

val cost : action -> int

Cost of an action

val which : state -> Spec.player

Determine which player is supposed to play in a given state

val utility : state -> Spec.player -> int

Compute the utility of a terminal state. If the given state is not terminal, utility is allowed to have any arbitrary behavior. utility is never called on non-terminal states.