Module Puzzle.Search

module type SOLVER = functor (X : Spec.SPEC) -> sig ... end

Type of solvers

module Solver : SOLVER

A Simple solver based on Dijsktra algorithm. If the problem specified by X has a solution, Solver(X).solve () returns an optimal sequence of actions leading to a goal state.

module BfsSolver : SOLVER

A solver Breadth First Search resolution algorithm. If the problem specified by X has a solution, BfsSolver(X).solve () returns the shortest sequence of actions leading to a goal state.