Module Search.BfsSolver

A solver Breadth First Search resolution algorithm. If the problem specified by X has a solution, BfsSolver(X).solve () returns the shortest sequence of actions leading to a goal state.

Note that this solver is usually much faster than Solver on problems where the cost of actions is constant.

Parameters

X : Spec.SPEC

Signature

val solve : unit -> X.action list option