Module Search.HSolver

A solver based on A* algorithm. If the problem specified by X has a solution, HSolver(X).solve () returns an optimal sequence of actions leading to a goal state. The search takes advantage of the function heuristic provided in the problem specification.

If no heuristic can be provided, Solver can be used instead.

Parameters

X : Spec.HSPEC

Signature

val solve : unit -> X.action list option