Module Search.Solver

A Simple solver based on Dijkstra algorithm. If the problem specified by X has a solution, Solver(X).solve () returns an optimal sequence of actions leading to a goal state.

If the cost of all actions is the same, it is recommended to use BfsSolver instead.

Parameters

X : Spec.SPEC

Signature

val solve : unit -> X.action list option