Module Puzzle.Spec

DSL(s) for generic problem specification

module type UNIT_SPEC = sig ... end

module type SPEC = sig ... end

SPEC is similar to UNIT_SPEC but allows to assign a cost to the actions. Cost informations can be used by solvers to search for an optimal solution (a solution with the lowest possible cost).