Module type Spec.UNIT_SPEC

Specification of unit costs search problems

UNIT_SPEC is a language to describe search problems where no notion of "cost" is involved.

Such problems are described by a state-space graph with unlabelled edges. The graphs are not explicitely represented because they can have an infinite number of nodes. Graphs are instead represented by the following elements :

a value init representing the initial state

a function actions listing the possible actions in a given state

a function apply to compute a new state from a previous state and an action

a function goal describing the goal states

type state

Type of the states

type action

Type of the actions

val init : state

Initial states

val goal : state -> bool

Test if a state is a goal state

val actions : state -> action list

Compute the actions available in a state

val apply : state -> action -> state

Apply an action to a state