Module Puzzle__.Puzzle_intf

Puzzle - an OCaml library and DSL for problem solving

Puzzle is beginner-friendly library to model puzzles in a declarative way using OCaml as a description language.

Models can then be solved using one of the provided solvers.

How to use Puzzle

To use Puzzle, the first step is to open the Puzzle library:

open Puzzle

module type SPEC = sig ... end

DSL for problem specification