Module Puzzle__Search

module Paths : sig ... end

A module to ease the incremental construction of paths in a state space graph. The state space is supposed to be characterised by a type state of states and a type action of actions to move from a state to another.

module Costs : sig ... end

A module to keep track of the costs associated to nodes in a search algorithm

module HSolver : functor (X : Puzzle.Spec.HSPEC) -> sig ... end

A solver based on A* algorithm. If the problem specified by X has a solution, HSolver(X).solve () returns an optimal sequence of actions leading to a goal state. The search takes advantage of the function heuristic provided in the problem specification.

module Solver : functor (X : Puzzle.Spec.SPEC) -> sig ... end

A Simple solver based on Dijkstra algorithm. If the problem specified by X has a solution, Solver(X).solve () returns an optimal sequence of actions leading to a goal state.

module HillClimbing : functor (X : Puzzle.Spec.HSPEC) -> sig ... end

Hill climbing solver

module BfsSolver : functor (X : Puzzle.Spec.UNIT_SPEC) -> sig ... end

A Breadth First Search resolution algorithm. If the problem specified by X has a solution, BfsSolver(X).solve () returns a shortest sequence of actions leading to a goal state.