Module Puzzle__Search.Costs

A module to keep track of the costs associated to nodes in a search algorithm

type 'a t

Type of costs tables

val init : 'a -> 'a t

Create a costs table with an initial node. The cost associated to the node is always 0

val improve : 'a t -> 'a -> int -> bool

improve tbl x cx checks if the cost cx is better than the current cost of x in tbl. If so, the cost is updated in place and true is returned. Otherwise, tbl is left unchanged and false is returned.