Module Puzzle__Search.Paths

A module to ease the incremental construction of paths in a state space graph. The state space is supposed to be characterised by a type state of states and a type action of actions to move from a state to another.

type ('state, 'action) t

Type of paths tables

val init : 'state -> ('state, 'action) t

Create a new table with a given initial state

val set : ('state, 'action) t -> 'state -> 'action -> 'state -> unit

Update the table. set tbl x act y updates the table tbl such that x is declared as the new predecessor of y and act is the action to perform to go from x to y

val build : ('state, 'action) t -> 'state -> 'action list

Try to extract a path (a list of actions) from the initial state to a given state. This function might not terminate if there's a cycle in the table. This function might throw an exception if the table is ill-formed (if there is a state in the table whose predecessor is not in the table)