Module Puzzle__Spec

DSL(s) for generic problem specification

module type UNIT_SPEC = sig ... end

UNIT_SPEC is a language to describe search problems where no notion of "cost" is involved.

module type SPEC = sig ... end

SPEC is similar to UNIT_SPEC but allows to assign a cost to the actions. Cost informations can be used by solvers to search for an optimal solution (a solution with the lowest possible cost).

module type HSPEC = sig ... end

HSPEC is similar to SPEC but allows to additionally specify a heuristic function. The solver Search.HSolver can take advantage of this heuristic to speed up the search.

type player =

| P1

| P2

module type GAME2 = sig ... end

2-players games

module type GAME = sig ... end

N-players games