Module type Puzzle__Spec.SPEC

SPEC is similar to UNIT_SPEC but allows to assign a cost to the actions. Cost informations can be used by solvers to search for an optimal solution (a solution with the lowest possible cost).

include UNIT_SPEC

type state

Type of the states

type action

Type of the actions

val init : state

Initial states

val goal : state -> bool

Test if a state is a goal state

val actions : state -> action list

Compute the actions available in a state

val apply : state -> action -> state

Apply an action to a state

val cost : action -> int

Cost of an action