Module Math.RatPoly

Polynomials with rational coefficients with arbitrary number of variables

include Algebra

type scalar

include Ring

include Rng

include AbelianGroup

type t

val zero : t

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val mul : t → t → t

val (*) : t → t → t

val one : t

val mone : t

minus one

val of_int : int → t

type names

If the type names remains private, it means it has no use in the current module implementation.

val to_tex : ?⁠names:names → t → string

val of_tex : string → t

val of_scalar : scalar → t

of_scalar s is scal_mul s one

val scal_mul : scalar → t → t

type monomial = Monomial_generic.t

val const : scalar → t

Constant polynomials

val xi : int → t

xi i is the polynomial 'x_i'

val of_monomial : monomial → t

val add_monomial : monomial → scalar → t → t

add_monomial c m p adds the monomial m with coefficient c to the polynomial p

val of_list : (scalar * monomial) list → t

Create a polynomial by adding all monomials with their given coefficients.