Module Math.RatPoly1

Polynomials in one variable with rational coefficients.

For instance, the following code

let open RatPoly1 in
    let p1 = of_array [| Q.one; Q.(of_int 1 / of_int 3); Q.one |] in
    let p2 = of_array [| Q.zero; Q.zero; Q.one; Q.of_int 2 |] in
    let p = p1 * p2 in
    Printf.printf "(%s) (%s) = %s\n" (to_tex p1) (to_tex p2) (to_tex p)

should print:

(1 + \frac{1}{3}x + x^2) (x^2 + 2x^3) = x^2 + \frac{7}{3}x^3 + \frac{5}{3}x^4 + 2x^5

include Polynomial.S with type names = string

include Algebra

type scalar

Type of the scalar Ring.

include Ring

include Rng

include AbelianGroup

type t

val zero : t

Neutral element for addition.

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val mul : t → t → t

val (*) : t → t → t

Ring multiplication.

val one : t

Ring unit element.

val mone : t

minus one.

val of_int : int → t

type names

If the type names remains private, it means it has no use in the current module implementation.

val to_tex : ?⁠names:names → t → string

val of_tex : string → t

val of_scalar : scalar → t

of_scalar s is scal_mul s one

val scal_mul : scalar → t → t

type monomial

val const : scalar → t

Constant polynomials

val xi : int → t

xi i is the polynomial 'x_i'

val of_monomial : monomial → t

val add_monomial : monomial → scalar → t → t

add_monomial c m p adds the monomial m with coefficient c to the polynomial p

val of_list : (scalar * monomial) list → t

Create a polynomial by adding all monomials with their given coefficients.

type generic = RatPoly.t

val x : t

The "x" polynomial

val of_array : scalar array → t

For instance of_array [| 3; 4; 5|] is the polynomial 3 + 4x + 5x^2.

val to_generic : int → t → generic

to_generic i p transforms the 1D polyomial p into a generic polyomial where the original variable is replaced by "x_i".