Module Roles.Role

type t =

| Washerwoman
(*
You start knowing that 1 of 2 players is a particular Townsfolk.
*)

| Librarian
(*
You start knowing that 1 of 2 players is a particular Outsider (or that zero are in play).
*)

| Investigator
(*
You start knowing that 1 of 2 players is a particular Minion.
*)

| Chef
(*
You start knowing how many pairs of Evil players there are.
*)

| Empath
(*
Each Night, learn how many of your two Neighbours are Evil.
*)

| FortuneTeller
(*
Each Night, choose 2 players: you learn if either is a Demon. There is a Good player that registers as a Demon to you (i.e., red herring).
*)

| Undertaker
(*
Each Night, you learn which Role died by execution the previous Day.
*)

| Monk
(*
Each Night, choose a player (not yourself): they are safe from the Demon tonight.
*)

| Ravenkeeper
(*
If you die at Night, you are woken to choose a player: you learn their Role.
*)

| Virgin
(*
The first time you are nominated, if the nominator is a Townsfolk, they are executed immediately.
*)

| Slayer
(*
Once per game, during the Day, publically choose a player: if they are the Demon, they die.
*)

| Soldier
(*
You are safe from the Demon.
*)

| Mayor
(*
If only 3 players live & no execution occurs, your team wins. If you die at Night, another player might die instead.
*)

| Butler
(*
Each Night, choose a player (not yourself): tomorrow, you may only vote if they are voting too.
*)

| Drunk
(*
You do not know you are the drunk, i.e., you think you are a Townsfolk character, but you are not.
*)

| Recluse
(*
You might register as Evil and as a Minion or Demon, even if dead.
*)

| Saint
(*
If you die by execution, your team loses.
*)

| Poisoner
(*
Each Night, choose a player: they are poisoned tonight and tomorrow-day. (This secretly removes/negates their ability.)
*)

| Spy
(*
Each Night, you see the grimoire. You might register as Good and as a Townsfolk or Outsider, even if dead. (The grimoire shows the state-of-the-game.)
*)

| ScarletWoman
(*
If there are 5 or more players alive and the Demon dies, you become the Demon. (Travellers don't count.)
*)

| Baron
(*
There are extra Outsiders in play (+2).
*)

| Imp
(*
Each Night, choose a player: they die. If you kill yourself this way, a Minion becomes the Imp.
*)

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val min : int

val max : int

val to_enum : t -> int

val of_enum : int -> t option

val equal : t -> t -> Ppx_deriving_runtime.bool

exception EnumOutOfBounds

val hash : t -> int

val compare : t -> t -> int

Kind

module Kind = Kind

val kind : t -> Kind.t

val is_townfolk : t -> bool

val is_outsider : t -> bool

val is_minion : t -> bool

val is_demon : t -> bool

Alignment

module Alignment = Alignment

val is_good : t -> bool

val is_evil : t -> bool

exception CannotDetermineAlignment of t

val alignment : t -> Alignment.t

val allied : t -> t -> bool

val opposed : t -> t -> bool