Module Rounds

module Phase = Game.Build.Rounds.Phase

module Round = Game.Build.Rounds.Round

Rounds

type t =

| Initial of initial
(*
Initial game-state configuration.
*)

| Round of round
(*
Some round.
*)

linked-list of rounds

and round = {

this : Round.t;

prev : t;

}

and initial = {

phase : Phase.t;
(*
Starting phase.
*)

players : Game.Build.Players.t;
(*
Initial players.
*)

rolemap : bool Roles.Map.t;
(*
Role selection map. (Required for implementing the Barons setup ability.)
*)

}

val make : Phase.t -> bool Roles.Map.t -> Game.Build.Players.t -> t

val initialize : initial -> t

val players : t -> Game.Build.Players.t

Traversal

val initial : t -> initial

val initial_phase : t -> Phase.t

val initial_players : t -> Game.Build.Players.t

val rolemap : t -> bool Roles.Map.t

Navigation

exception SetupRound

val this : t -> Round.t

exception FirstRound

val prev : t -> t

prev x returns prev of x.

raises SetupRound
if x is Initial _.

raises FirstRound
if x.prev is Initial _.

val step : t -> t

step x is the progression to the next phase of x, which ultimately depends on the initial.phase.

Query

val is_initial : t -> bool

is_initial x is true if x is Initial, else false.

val count : t -> int

count x is the number of rounds there have been in x. (Where Initial is 0.)

val num : t -> int

num x is the number of times the number of rounds x modulo the enum phases.

Round Phase

val phase_of_int : t -> int -> Phase.t

val phase : t -> Phase.t

Phase Query

exception SetupPhase

SetupPhase is raised when an Initial round is unexpectedly found.

val is_first_phase : t -> bool

is_first_phase x is true if x.prev is Initial, else false.

raises SetupPhase
if x is Initial _.

val is_phase_start : t -> bool

...

raises SetupPhase
if x is Initial _ (since this would be trivially true).

val is_phase_end : t -> bool

...

raises SetupPhase
if x is Initial _ (since something with no start cannot end).