Module Scaml_lib.Llast

Copyright 2023-2024, Grigory Aseev and Matvey Kalashnikov

SPDX-License-Identifier: LGPL-2.1-or-later

Ast without anonymous functions, arguments are listed, the recursiveness of nested binding has been removed, since it makes no sense to store it for a variable

type llexpr =

| LLConst of Ast.const
(*
constants such as integers, boolean, unit
*)

| LLBinOp of Ast.bin_op * llexpr * llexpr
(*
binary operations such as add, multiply, equality checking etc.
*)

| LLVar of Ast.id
(*
identifier for variables
*)

| LLIf of llexpr * llexpr * llexpr
(*
conditionals
*)

| LLApp of llexpr * llexpr
(*
applying a function to an argument
*)

| LLLetIn of Ast.id * llexpr * llexpr
(*
binding to a variable
*)

val pp_llexpr : Ppx_deriving_runtime.Format.formatter -> llexpr -> Ppx_deriving_runtime.unit

val show_llexpr : llexpr -> Ppx_deriving_runtime.string

type llbinding =

| LLLet of bool * Ast.id * Ast.pattern list * llexpr

binding to a function with arguments

val pp_llbinding : Ppx_deriving_runtime.Format.formatter -> llbinding -> Ppx_deriving_runtime.unit

val show_llbinding : llbinding -> Ppx_deriving_runtime.string

type llprogram = llbinding list

type containing functions at the top level

val pp_llprogram : Ppx_deriving_runtime.Format.formatter -> llprogram -> Ppx_deriving_runtime.unit

val show_llprogram : llprogram -> Ppx_deriving_runtime.string