Module Sql_lib.Meta

Copyright 2021-2022, Michael Polyntsov and contributors

SPDX-License-Identifier: LGPL-3.0-or-later

Functions and types to work with DBMS entities such as catalog, databases, tables and columns

type column_type =

| IntCol

| StringCol

Type of values that the column stores

val yojson_of_column_type : column_type -> Ppx_yojson_conv_lib.Yojson.Safe.t

val column_type_of_yojson : Ppx_yojson_conv_lib.Yojson.Safe.t -> column_type

val column_type_to_string : column_type -> string

val column_type_of_string : string -> column_type

type column

Table attribute, constitutes column's name and type

val column_type : column -> column_type

type header

Header of the table, i.e. list of attributes

type table

Table in the database represented by a name and a header

type database

Database represented as a named list of tables

type catalog

Catalog represented as list of databases, there is only one catalog

Catalog is stored with this directory structure: _catalog/┌──►Catalog.meta.json │ ├──►Database1_name/┌───►Table1 │ │ │ └───►Table2 │ └──►Database2_name/┌───►Table1 │ ├───►Table2 │ └───►Table3 Catalog.meta.json contains all information about this catalog, i.e. all databases (identified by their names), their tables, headers of the tables

exception AmbiguousEntity

module Table : sig ... end

module Database : sig ... end

module Catalog : sig ... end