Module Meta.Catalog

val create : string -> catalog

create path creates new catalog in path directory, throws an exception if one already exists

val load : string -> catalog

load path loads catalog from the path directory and returns it

val init : string -> catalog

init path as load but if catalog does not exist yet creates a new empty one

val recreate : string -> catalog

recreate path creates new catalog in path directory as create, but if it already exists drops it and creates a new emtpy one

val dump : catalog -> unit

dump c writes catalog c (its meta and creates internal structure for data) to the disk to catalog's path

val dump_meta : catalog -> unit

dump_meta c writes meta of catalog c to the disk to catalog's path

val drop : catalog -> unit

drop c deletes catalog c from the disk

val create_db : string -> catalog -> database * catalog

create_db name c creates a new empty database named name in corresponding catalog c

val create_table : string -> database -> catalog -> table * database * catalog

create_table name db creates a new empty table named name in corresponding database db

val create_cols : (string * column_type) list -> table -> catalog -> table * database * catalog

create_cols cols_list table creates columns from cols_list list of <name, type> pairs and inserts them into table

val to_string : catalog -> string

to_string c converts c to json string

val get_db : string -> catalog -> database option

get_db name c returns Some database with name name or None if none exists

val get_table_path : table -> catalog -> string

get_table_path table returns path to the table

val get_dbs : catalog -> database list

get_dbs c returns all databases in c

val get_table_types : table -> column_type list

get_table_types t returns a list of table t column types