Module Meta.Database

val get_name : database -> string

val get_tables : database -> table list

get_tables db c returns all tables in database db

val get_table_ci : string -> database -> table

get_table name db returns table with name equal to name ignoring case

raises Not_found
if there is no table named name in db

raises AmbigousEntity
if there is more than one table named name in db

val get_table : string -> database -> table option

get_table name db returns table named name as Some table or None if there is no such table in database db

val get_col_table : column -> database -> table

val get_col_ci : string -> table list -> table * column

get_col_ci name tables searches for column named name among all tables in tables ignoring case and returns it

raises Not_found
if there is not column named name among all tables in tables

raises AmbigousEntity
if there is more than one column named name

val get_col_by_fullname_ci : string -> database -> table * column