Module Meta.Table

val name : table -> string

val col_fullname : column -> table -> string

val cols_as_fullnames : table -> string list

val get_col_ci : string -> table -> column

get_table name t returns column in table t with name equal to name ignoring case

raises Not_found
if there is no column named name in t

raises AmbigousEntity
if there is more than one column named name in t

val get_table_by_name : string -> table list -> table