Module Algo.AlgoLog

All functions in this module have no effect on the flow of the algorithm. Their only purpose is to log or print information about the current state of the algorithm.

val show_stat : float -> int -> int -> unit

val show_steps : int -> int -> Base.unit

val show_counterexamples : Common.ProblemDefs.refinement_loop_state -> (Lang__Term.term, Base__Comparator.Make(Lang__Term.Terms.E).comparator_witness) Base.Set.t -> Base.unit

val show_summary : (string * string * string) -> Lang.PMRS.t -> Base.unit

val show_pmrs : Lang.PMRS.t -> Base.unit

val show_new_rskel : Base.Int.t -> Common.ProblemDefs.psi_def -> Base.unit

val msg_too_many_opts : unit -> Base.unit

val pp_unrealizability_witness : Base.Formatter.t -> Common.ProblemDefs.unrealizability_witness -> Base.unit

val show_unrealizability_witnesses : (Lang.Term.VarSet.elt, 'a) Base.Set.t -> Common.ProblemDefs.equation Base.List.t -> Common.ProblemDefs.unrealizability_witness list -> unit

val no_spurious_witness : unit -> Base.unit

val spurious_violates_requires : int -> Base.unit

val print_infeasible_message : (Lang.Term.term, 'a) Base.Set.t -> Base.unit

val announce_new_term_state : Common.ProblemDefs.witness -> Base.unit

val announce_new_lemma_synthesis : Base.int -> Common.ProblemDefs.term_info -> Base.unit

val lemma_not_proved_correct : Utils.Stats.verif_method -> unit

val lemma_proved_correct : Utils.Stats.verif_method -> Common.ProblemDefs.term_info -> Lang.Term.term -> Base.unit

val violates_ensures : Common.ProblemDefs.psi_def -> Common.ProblemDefs.witness Base.List.t -> unit

val positives_ensures : Common.ProblemDefs.psi_def -> Lang.Term.term list -> unit

val show_new_ensures_predicate : Lang.Term.variable -> Lang.Term.term -> Base.unit

val image_witness_class : Common.ProblemDefs.psi_def -> Common.ProblemDefs.witness -> Smtlib.SmtLib.solver_response -> Utils.Stats.verif_method -> unit

val requires_witness_class : Lang.PMRS.t -> Common.ProblemDefs.witness -> Smtlib.SmtLib.solver_response -> Utils.Stats.verif_method -> unit