Module Algo.Equations

val check_equation : p:AState.psi_def -> AState.equation -> Base.bool

val compute_lhs : AState.psi_def -> Lang.Term.term -> Lang.Term.term

Compute the left hand side of an equation of p from term t. The result is a maximally reduced term with some applicative terms of the form (p.psi_reference x) where x is a variable.

val remap_rec_calls : AState.psi_def -> Lang.Term.term -> Lang.Term.term

val compute_rhs_with_replacing : AState.psi_def -> Lang.Term.term -> Lang.Term.term

val compute_rhs : ?⁠force_replace_off:Base.bool -> AState.psi_def -> Lang.Term.term -> Lang.Term.term

val compute_preconds : p:AState.psi_def -> term_state:AState.term_state -> (Lang.Term.term -> Lang.Term.term) -> Lang.Term.term -> Lang.Term.term Base.Option.t

val filter_elims : (Lang.Term.term * 'a) Base.List.t -> Lang.Term.term -> (Lang.Term.term * 'a) Base.List.t

val make : ?⁠force_replace_off:Base.bool -> p:AState.psi_def -> term_state:AState.term_state -> lifting:AState.lifting -> Lang.Term.TermSet.t -> AState.equation Base.list * AState.lifting

val revert_projs : Lang.Term.VarSet.t -> (Base.int, Lang.Term.variable Base.list, Base.Int.comparator_witness) Base.Map.t -> (Base.string * Lang.Term.variable Base.list * Lang.Term.term) Base.list -> (Base.string * Lang.Term.variable Base.list * Lang.Term.term) Base.list

val free_vars_of_equations : AState.equation Base.list -> Lang.Term.VarSet.t

type partial_soln = (Base.string * Lang.Term.variable Base.list * Lang.Term.term) Base.list

val pp_partial_soln : Base.Formatter.t -> (string * Lang.Term.Variable.t list * Lang.Term.term) list -> unit

val combine : ?⁠verb:bool -> ((string * Lang.Term.Variable.t list * Lang.Term.term) Base__List.t, 'a Base__List.t) Base.Either.t -> ('b * ((string * Lang.Term.Variable.t list * Lang.Term.term) Base__List.t, 'a Base__List.t) Base.Either.t) -> 'b * ((string * Lang.Term.Variable.t list * Lang.Term.term) Base__List.t, 'a Base__List.t) Base.Either.t

val solve_syntactic_definitions : Lang.Term.VarSet.t -> AState.equation Base.list -> (Base.string * Lang__Term.variable Base.List.t * Lang.Term.term) Base.List.t * (Lang__Term.Variable.t, Lang__Term.Variable.comparator_witness) Base.Set.t * AState.equation Base.List.t

val synthfuns_of_unknowns : ?⁠bools:Base.bool -> ?⁠eqns:AState.equation Base.List.t -> ?⁠ops:(Lang__Term.Operator.t, Lang__Term.Operator.comparator_witness) Base.Set.t -> Lang.Term.VarSet.t -> Syguslib.Sygus.command Base.List.t

val constraints_of_eqns : AState.equation Base.list -> Syguslib.Sygus.command Base.list

val aux_solve : Base.bool -> Lang.Term.VarSet.t -> AState.equation Base.list -> (Syguslib.Sygus.solver_response * ((Syguslib.Sygus.symbol * Lang__Term.variable Base.List.t * Lang.Term.term) Base.List.t, 'a list) Base.Either.t) Lwt.t * Base.int Lwt.u

val solve_eqns : Lang.Term.VarSet.t -> AState.equation Base.list -> Syguslib.Sygus.solver_response * (partial_soln, Counterexamples.unrealizability_ctex Base.list) Base.Either.t

val solve_eqns_proxy : Lang.Term.VarSet.t -> AState.equation Base.list -> Syguslib.Sygus.solver_response * ((Base.string * Lang.Term.Variable.t Base.List.t * Lang.Term.term) Base__List.t, Counterexamples.unrealizability_ctex Base__List.t) Base.Either.t

val solve_constant_eqns : Lang.Term.VarSet.t -> AState.equation Base.list -> (Base.string * 'a list * Lang.Term.term) Base.List.t * (Lang__Term.Variable.t, Lang__Term.Variable.comparator_witness) Base.Set.t * AState.equation Base.List.t

val split_solve : (Base.string * Lang.Term.Variable.t Base.List.t * Lang.Term.term) Base__List.t -> Lang.Term.VarSet.t -> AState.equation Base.list -> Syguslib.Sygus.solver_response * ((Base.string * Lang.Term.Variable.t Base.List.t * Lang.Term.term) Base__List.t, Counterexamples.unrealizability_ctex Base__List.t) Base.Either.t

val solve_stratified : Lang.Term.VarSet.t -> AState.equation Base.list -> Syguslib.Sygus.solver_response * ((Base.string * Lang.Term.Variable.t Base.List.t * Lang.Term.term) Base__List.t, Counterexamples.unrealizability_ctex Base__List.t) Base.Either.t

type preprocessing_action_result = {

pre_unknowns : Lang.Term.VarSet.t;

pre_equations : AState.equation Base.list;

pre_postprocessing : (Syguslib.Sygus.solver_response * (partial_soln, Counterexamples.unrealizability_ctex Base.list) Base.Either.t) -> Syguslib.Sygus.solver_response * (partial_soln, Counterexamples.unrealizability_ctex Base.list) Base.Either.t;

}

A preprocessing action should return a new system of equations, and optionally a new set of unknowns together with a postprocessing function.

val preprocess_none : Lang.Term.VarSet.t -> AState.equation Base.list -> preprocessing_action_result

An empty preprocessing action.

val preprocess_detuple : Lang.Term.VarSet.t -> AState.equation Base.list -> preprocessing_action_result

Preprocessing action that projects each unknown returning a tuple into a tuple of unknowns and change the equations accordingly. Postprocessing consist of rebuilding the tuples.

val preprocess_deconstruct_if : Lang.Term.VarSet.t -> AState.equation Base.list -> preprocessing_action_result

Preprocessing action that transforms constraints with conditionals into sets of constraints where conditions have been moved in the precondition of the constraint.

val preprocess_factor_subexpressions : Lang.Term.VarSet.t -> AState.equation Base.list -> preprocessing_action_result

For each equation, search within the system of equations whether there is another equation that constrains a subexpression.

val solve : p:AState.psi_def -> AState.equation Base.list -> Syguslib.Sygus.solver_response * (partial_soln, Counterexamples.unrealizability_ctex Base.list) Base.Either.t

Main entry point: solve a system of equations by synthesizing the unknowns. Returns either a solution as a list of implementations for the unknowns (a triple of unknown name, arguments of a function and body of a function) or a list of unrealizability counterexamples.