Module Algo.Lifting

val msg_lifting : unit -> Base.unit

module LiftingMap : sig ... end

val empty_lifting : AState.lifting

val is_empty_lifting : AState.lifting -> bool

val alpha_component_count : unit -> int

val decompose_t : AState.psi_def -> Lang.Term.term -> (Base.int * Lang.Term.term) Base.option

val recompose_t : AState.psi_def -> Lang.Term.term -> Base.int -> Lang.Term.term

val get_mapped_value : p:AState.psi_def -> AState.lifting -> Lang.Term.term -> Lang.Term.term Base.Option.t

Find the expression of the lifting for an input term.

val interactive_add_lifting_expression : p:AState.psi_def -> AState.lifting -> Lang.Term.term -> Base.int -> AState.lifting * Lang.Term.term Base.option

Interactively add an expression for the lifting.

val replace_boxed_expressions : p:AState.psi_def -> AState.lifting -> Lang.Term.term -> Lang.Term.term

val is_lifted : AState.psi_def -> Base.bool

A problem definition is lifted if the ouput type of the target recursion skeleton is different from the output type of the reference function (which is unchanged and stored in AState._alpha)

val lift_count : AState.psi_def -> Base.int

val proj_to_non_lifting : AState.psi_def -> Lang.Term.term Base.option

Return a projection function that can be symbolically evaluated.

val is_proj_function : AState.psi_def -> Lang.Term.term -> Base.bool

val proj_to_lifting : AState.psi_def -> (Lang.Term.term -> Lang.Term.term) Base.option

Return function f such that f(a) = a.i, .. a.i + n if i, .. i+n where i, ..., i+n are the components of the lifting.

val compose_parts : AState.psi_def -> Lang.Term.term Base.option

Returns a symbolic function f such that f original_part lifting_part computes the tuple that corresponds to the lifted function from the original components of the function in oringal_part and the components of the lifting only in lifting_part

val ith_type : AState.psi_def -> Base.int -> Lang.RType.t Base.option

val apply_lifting : p:AState.psi_def -> Lang.RType.t Base.list -> AState.psi_def

apply_lifting ~p l

val deduce_lifting_expressions : p:AState.psi_def -> AState.lifting -> Lang.Term.term Base.option -> Lang.Term.term -> Lang.Term.term -> AState.lifting

val scalar : p:AState.psi_def -> AState.refinement_loop_state -> 'a -> (AState.psi_def * AState.refinement_loop_state, Syguslib.Sygus.solver_response) Base.Result.t

Perform a scalar lifting.