Module Algo.Projection

val simple_flattening : Lang.Term.term Base.list -> Lang.Term.term Base.List.t

Flatten a list of terms that may be tuples into a flat list with no tuples.

val mk_projs : Lang.RType.t Base.list -> Lang.RType.t Base.list -> Lang.Term.Variable.t -> Lang.Term.Variable.t Base.List.t

Construct the projection functions for the given variable. The character '$' is used to build projections and should not be used anywhere else. (It is compatible with usage in smtlib but not in the ocaml program.)

val apply_projections : (Base.int, Lang.Term.variable Base.list, Base.Int.comparator_witness) Base.Map.t -> Lang.Term.term -> Lang.Term.term

apply_projections projection_map t replaces in t all the variables whose id appears in the projection map by the associated tuple of variables. Does not reduce the term after applying the projections.

val projection_eqns : Lang.Term.term -> Lang.Term.term -> (Lang.Term.term * Lang.Term.term) Base.List.t

projection_eqns lhs rhs returns a list of equations that corresponds to equating every components of lhs and rhs. If rhs and lhs are tuples of the same length n, then a list of equations of length n is returned. Otherwise [(lhs,rhs)] is returned.

val invar : (Lang.Term.term, 'a) Base.Set.t -> Lang.Term.term -> Lang.Term.term -> Lang.Term.term Base.option

invar invariants lhs_e rhs_e filters the terms in `invariants` that have no free variable in common with lhs_e or rhs_e and return the conjuction of all these invariants.

val proj_and_detuple_eqns : (Base.int, Lang.Term.variable Base.list, Base.Int.comparator_witness) Base.Map.t -> Common.ProblemDefs.equation Base.list -> Common.ProblemDefs.equation Base.List.t

val proj_functions : Lang.Term.VarSet.t -> Lang.Term.VarSet.t * (Base.int, Lang.Term.variable Base.list, Base.Int.comparator_witness) Base.Map.t

proj_functions functions creates a map from variables in functions to tuple of new functions that represent each component of the original function, when the return type of that function is a tuple. If the return type of the function is not a tuple, then it is left unchanged. Use apply_projections to apply the projection map to a term.

returns
a tuple containing the set of new functions created and a map from non-projected functions to the list of functions corrresponding to the projections.