Module Algo.Verify

val check_solution : p:Common.ProblemDefs.psi_def -> Common.ProblemDefs.refinement_loop_state -> (Base.string * Lang.Term.variable Base.list * Lang.Term.term) Base.list -> [ `Incorrect_assumptions | `witnesss of (Lang.Term.term, Lang.Term.Terms.comparator_witness) Base.Set.t * (Lang.Term.term, Lang.Term.Terms.comparator_witness) Base.Set.t | `Correct ]

`check_solution ~p (t,u) soln`

Checks if ᴪ(p, soln) is valid, assuming `soln` has been synthesized using the term set `t` with expansion continuation `u`. Returns the set (t,u) for the next step in the witness-guided refinement loop. This bounded checking procedure takes advantage of the only partial bounding, and does not perform bounded checking by expanding unnecessary terms.

val bounded_check : ?use_concrete_witness:Base.bool -> p:Common.ProblemDefs.psi_def -> (Base.string * Lang.Term.variable Base.list * Lang.Term.term) Base.list -> Common.ProblemDefs.equation Base.option

Perform a bounded check of the solution. As opposed to check_solution this does not take advantage of partial bounding techniques. Consequently, it does not require sets of terms as arguments but only a problem definition and a solution. Returns the first equation for which checking has failed (the first element is the counterexample). Returns None if the check passed.

val invert : Lang.PMRS.t -> Lang.Term.Constant.t -> Lang.Term.term Base.list Base.option

Solve an equation of the form f(x1,..,xn) = c where f is a recursive function and t is a constant. Returns Some list of terms, one for each input of the function f, if the equation admits a solution. Returns None if the equation has no solution.