Module Algo__AState

type psi_def = {

psi_target : Lang.PMRS.t;
The target recursion skeleton in the problem.

psi_reference : Lang.PMRS.t;
The reference function in the problem.

psi_repr : Lang.PMRS.t;
The representation function in the problem.

psi_tinv : Lang.PMRS.t Base.option;
The requires predicate of the target recursion skeleton.

psi_repr_is_identity : Base.bool;
A boolean that is true iff psi_repr is identity.

psi_lifting : Lang.RType.t Base.list;
The current lifting: the output type of psi_target should be !_alpha extended with the tuple of types psi_lifting.

}

val _tau : Lang.RType.t Base.ref

The type τ in the paper, input type of the reference function.

val _theta : Lang.RType.t Base.ref

The type θ in the paper, input type of the target recursion skeleton.

val _alpha : Lang.RType.t Base.ref

The type D in the paper: the output type of the reference and the target recursion skeleton. The first element is the pure type output type of the functions, and the second element is an optional term that represents the additional predicate on the output of the reference function. The term is assumed to define a function (fun (free variables of term) -> term).

val _span : int Base.ref

Not useful for now.

val refinement_steps : int Base.ref

val secondary_refinement_steps : int Base.ref

val reinit : unit -> Base.unit

type soln = {

soln_rec_scheme : Lang.PMRS.t;
The recursion scheme this object is a solution of.

soln_implems : (Syguslib.Sygus.symbol * Lang.Term.variable Base.list * Lang.Term.term) Base.list;
The association list containing the implementation of each unknown. Each element of the list is a triple of unknown name, list of arguments, and body of the unknown implementation.

}

Representation of a solution, i.e. a function (as a PMRS) together with an association list from unknown name to its implementation.

type equation = {

eterm : Lang.Term.term;
The term from which the equation originates.

eprecond : Lang.Term.term Base.option;
An optional precondition to the equation.

elhs : Lang.Term.term;
The left-hand side of the equation, containing no unknowns.

erhs : Lang.Term.term;
The right-hand side of the equation, possibly with unknowns.

eelim : (Lang.Term.term * Lang.Term.term) Base.list;
The substitution used to eliminate recursion.

}

Represents an equational constraint.

type spurious_cause =

| ViolatesTargetRequires

| NotInReferenceImage
A counterexample can be spurious either because it wiolates the target's requires or because some values assigned to recursion elimination variables are not in the reference function's image.

type ctex_stat =

| Valid

| Spurious of spurious_cause

| Unknown
A counterexample is either valid, or spurious with some reason, or unknown.

type ctex = {

ctex_eqn : equation;
The equation the counterexample relates to.

ctex_vars : Lang.Term.VarSet.t;
The variables in the model.

ctex_model : Lang.Term.term Lang.Term.VarMap.t;
The model of the counterexample, mapping variables to terms. The terms should be constants.

ctex_stat : ctex_stat;
The spuriousness status of the counterexample.

}

A counterexample related to an equation and some info on the validity of the counterexample.

type equation_system = equation Base.list

val solved_eqn_system : equation_system Base.option Base.ref

Contains the last equation system used to solved the problem when a solution is found.

type term_state_detail = {

term : Lang.Term.term;

lemmas : Lang.Term.term Base.list;

lemma_candidate : (Base.string * Lang.Term.variable Base.list * Lang.Term.term) Base.option;

negative_ctexs : ctex Base.list;

positive_ctexs : ctex Base.list;

recurs_elim : (Lang.Term.term * Lang.Term.term) Base.list;

scalar_vars : Lang.Term.variable Base.list;

current_preconds : Lang.Term.term Base.option;

}

type term_state = (Lang.Term.term, term_state_detail, Lang.Term.Terms.comparator_witness) Base.Map.t

type lifting = {

tmap : ((Base.int * Lang.Term.term) * Lang.Term.term) Base.list;

}

The type to describe liftings. tmap is a map from terms to the expression of the lifting.

type refinement_loop_state = {

t_set : Lang.Term.TermSet.t;

u_set : Lang.Term.TermSet.t;

term_state : term_state;

lifting : lifting;

}

The state of the main refinement loop. Currently, it is entirely determined by the sets T and U, accompanied with a set of lemmas that can be used during the generation of constraints.

val pp_equation : Base.Formatter.t -> equation -> Base.unit

val pp_ctex : Base.Formatter.t -> ctex -> Base.unit

val pp_implems : Base.Formatter.t -> (Syguslib.Sygus.symbol * Lang.Term.variable Base.list * Lang.Term.term) Base.list -> unit

val pp_soln : ?⁠use_ocaml_syntax:bool -> Base.Formatter.t -> soln -> Base.unit