Module Algo__Counterexamples

val smt_unsatisfiability_check : Lang.Term.VarSet.t -> Algo.AState.equation Base.list -> Base.unit

type unrealizability_ctex = {

i : Base.int;

j : Base.int;

ci : Algo.AState.ctex;

cj : Algo.AState.ctex;

}

A counterexample to realizability is a pair of models: a pair of maps from variable ids to terms.

val pp_unrealizability_ctex : Base.Formatter.t -> unrealizability_ctex -> Base.unit

val reinterpret_model : ((Base.string, 'a, 'b) Base.Map.t * (Base.string, 'c, 'd) Base.Map.t) -> (Lang.Term.variable * Lang.Term.variable) Base.List.t -> (Lang__Term.Variable.t, 'a, Lang__Term.Variable.comparator_witness) Base.Map.t * (Lang__Term.Variable.t, 'c, Lang__Term.Variable.comparator_witness) Base.Map.t

val unrealizability_ctex_of_constmap : (Base.int * Base.int) -> (Algo.AState.equation * Algo.AState.equation) -> ((Lang.Term.VarSet.elt, Lang__Term.Variable.comparator_witness) Base.Set.t * (Lang.Term.VarSet.elt, Lang__Term.Variable.comparator_witness) Base.Set.t) -> (Lang.Term.variable * Lang.Term.variable) Base.List.t -> (Base.String.t, Lang.Term.term, 'a) Base.Map.t -> unrealizability_ctex

val skeleton_match : unknowns:Lang.Term.VarSet.t -> Lang.Term.term -> Lang.Term.term -> (Lang.Term.term * Lang.Term.term) Base.list Base.option

val components_of_unrealizability : unknowns:Lang.Term.VarSet.t -> Algo.AState.equation -> Algo.AState.equation -> ((Lang.Term.term * Lang.Term.term) Base.list * (Lang.Term.term * Lang.Term.term)) Base.option

val check_unrealizable : Lang.Term.VarSet.t -> Algo.AState.equation_system -> unrealizability_ctex Base.list

Check if system of equations defines a functionally realizable synthesis problem. If any equation defines an unsolvable problem, an unrealizability_ctex is added to the list of counterexamples to be returned. If the returned list is empty, the problem may be solvable/realizable. If the returned list is not empty, the problem is not solvable / unrealizable.

val check_image_sat : p:Algo.AState.psi_def -> Algo.AState.ctex -> Lang.SmtInterface.AsyncSmt.response * Base.int Lwt.u

val check_image_unsat : p:Algo.AState.psi_def -> Algo.AState.ctex -> Lang.SmtInterface.AsyncSmt.response * Base.int Lwt.u

val check_ctex_in_image : p:Algo.AState.psi_def -> Algo.AState.ctex -> Algo.AState.ctex

check_ctex_in_image ~p ctex checks whether the recursion elimination's variables values in the model of ctex are in the image of (p.psi_reference o p.psi_repr).

val find_original_var_and_proj : Lang.Term.variable -> ('a * Lang.Term.term) -> ('a * int) Base.Option.t

val mk_model_sat_asserts : Algo.AState.ctex -> ('a -> Lang.Term.term) -> (Lang.Term.variable -> 'a Base.Option.t) -> Smtlib.SmtLib.smtTerm Base.List.t

val check_tinv_unsat : p:Algo.AState.psi_def -> Lang.PMRS.t -> Algo.AState.ctex -> Lang.SmtInterface.AsyncSmt.response * Base.int Lwt.u

check_tinv_unsat ~p tinv c checks whether the counterexample c satisfies the predicate tinv in the synthesis problem p. The function returns a promise of a solver response and the resolver associated to that promise (the promise is cancellable). If the solver response is unsat, then there is a proof that the counterexample c does not satisfy the predicate tinv. In general, if the reponse is not unsat, the solver either stalls or returns unknown.

val check_tinv_sat : p:Algo.AState.psi_def -> Lang.PMRS.t -> Algo.AState.ctex -> Lang.SmtInterface.AsyncSmt.response * Base.int Lwt.u

check_tinv_sat ~p tinv ctex checks whether the counterexample ctex satisfies the invariant tinv (a PMRS). The function returns a pair of a promise of a solver response and a resolver for that promise. The promise is cancellable.

val satisfies_tinv : p:Algo.AState.psi_def -> Lang.PMRS.t -> Algo.AState.ctex -> Algo.AState.ctex

val classify_ctexs : p:Algo.AState.psi_def -> Algo.AState.ctex Base.list -> Algo.AState.ctex Base.list

Classify counterexamples into positive or negative counterexamples with respect to the Tinv predicate in the problem.