Module Algo__Expand

val mk_recursion_elimination_term : Algo.AState.psi_def -> (Lang.Term.term * Lang.Term.term) Base.option

val subst_recursive_calls : Algo.AState.psi_def -> Lang.Term.term Base.list -> (Lang.Term.term * Lang.Term.term) Base.list * Lang.Term.TermSet.t

val maximally_reduced_app : Algo.AState.psi_def -> Lang.Term.term -> Lang.Term.term Base.list -> Base.bool

val nonreduced_terms : Algo.AState.psi_def -> Lang.Term.VarSet.t -> Lang.Term.term -> (Lang.Term.variable * Lang.Term.term Base.list) Base.list

val nonreduced_terms_all : Algo.AState.psi_def -> Lang.Term.term -> (Lang.Term.variable * Lang.Term.term Base.list) Base.list

val replace_rhs_of_main : ?⁠verbose:Base.bool -> ?⁠for_mr:Base.bool -> Algo.AState.psi_def -> Lang.PMRS.t -> Lang.Term.term -> Lang.Term.term

Replace subterms that correspond to the right-hand side of the main rule to avoid capturing f(v) when v is a variable.

val replace_rhs_of_mains : Algo.AState.psi_def -> Lang.Term.term -> Lang.Term.term

Applies `replace_rhs_of_main` with all main functions in the problem definition (psi_def)

val simple : ?⁠verbose:Base.bool -> ?⁠max_height:Base.int -> Lang.Term.term -> (Lang.Term.term, Lang.Term.Terms.comparator_witness) Base.Set.t * (Lang.Term.term, Lang.Term.Terms.comparator_witness) Base.Set.t

Simple term expansion procedure. simple t expands the term t into two sets of terms, the first set is a set of *bounded* terms, the second set is a set of unbounded terms. In an algorithm, one should keep track of the unbounded terms as well and recursively expand them as necessary.

val make_bounded : Lang.Term.term -> Lang.Term.term

Heuristic to bound terms. Unbounded symbols will be bound using some arbitrary expansion.

val is_mr : Algo.AState.psi_def -> Lang.PMRS.t -> Lang.Term.term -> Lang.Term.VarSet.t -> Base.bool

val is_mr_all : Algo.AState.psi_def -> Lang.Term.term -> Base.bool

val to_maximally_reducible : Algo.AState.psi_def -> Lang.Term.term -> Lang.Term.TermSet.t * Lang.Term.TermSet.t

val expand_loop : Base.int Base.ref -> (Smtlib.Solvers.solver_response -> Lang.Term.term -> Smtlib.Solvers.solver_response) -> ?⁠r_stop:(Smtlib.SmtLib.solver_response -> Base.bool) -> ?⁠r_complete:Smtlib.SmtLib.solver_response -> Lang.Term.TermSet.t -> Smtlib.Solvers.solver_response

expand_loop provides basic functionality for bounded checking. expand_loop steps f start will run a bounded checking loop, starting with the term set start and performing !Lib.Config.num_expansions_check steps, applying the checking function f for each term obtained during the expansion of the term in the set start into bounded terms.

If a check returns an answer for which r_stop is true, then the function returns this answer. By default r_stop checks whether the answer is SmtLib.Sat.

If every check returns UNSAT, then the procedure returns SmtLib.Unknown, unless !Config.no_bounded_sat_as_unsat is true, in which case it returns r_complete. By default, r_complete is SmtLib.Unsat.

If the procedure has explored all possible expansions of the terms in the input set, then it returns r_complete (all the check having returned UNSAT for all possible terms).

parameter r_stop

is a function that returns true whenever the loop should stop, given the solver response to a call to f.

parameter r_complete

is the answer that should be given if all the

val lwt_expand_loop : Base.int Base.ref -> (Smtlib.Solvers.Asyncs.response -> Lang.Term.term -> Smtlib.Solvers.Asyncs.response) -> ?⁠r_stop:(Smtlib.SmtLib.solver_response -> Base.bool) -> ?⁠r_complete:Smtlib.SmtLib.solver_response -> Lang.Term.TermSet.t Lwt.t -> Smtlib.Solvers.Asyncs.response

lwt_expand_loop provides the same functionality as expand_loop, but the solver calls are threaded arount a lwt promise. Use lwt_expand_loop in place of expand_loop when you want to run concurrent solver instances.