Module Algo__Projection

val mk_projs : Lang.RType.t Base.list -> Lang.RType.t Base.list -> Lang.Term.Variable.t -> Lang__Term.variable Base.List.t

Construct the projection functions for the given variable.

val projection_eqns : Lang.Term.term -> Lang.Term.term -> (Lang.Term.term * Lang.Term.term) Base.List.t

val invar : (Lang.Term.term, 'a) Base.Set.t -> Lang.Term.term -> Lang.Term.term -> Lang.Term.term Base.option

invar invariants lhs_e rhs_e filters the terms in `invariants` that have no free variable in common with lhs_e or rhs_e and return the conjuction of all these invariants.

val proj_and_detuple_eqns : (Base.int, Lang.Term.variable Base.list, Base.Int.comparator_witness) Base.Map.t -> Algo.AState.equation Base.list -> Algo.AState.equation Base.List.t

val proj_unknowns : Lang.Term.VarSet.t -> (Lang__Term.Variable.t, Lang__Term.Variable.comparator_witness) Base.Set.t * (Base.Int.t, Lang__Term.variable Base.List.t, Base.Int.comparator_witness) Base.Map.t