Module Codegen.Dafny

type d_attribute =

| DAttrAssumption

| DAttrAutoReq

| DAttrAutocontract

| DAttrAxiom

| DAttrCompile

| DAttrDecl

| DAttrFuel of Base.string * Base.int * Base.int Base.option

| DAttrHeapQuantifier

| DAttrImported

| DAttrInduction of Base.bool * Base.string Base.list
DAttrInduction(b,l) is induction b if l is empty, induction l otherwise. l is a list consisting of bound variables.

| DAttrLayerQuantifier

| DAttrNaiveType of Base.bool Base.option * Base.string Base.option

| DAttrOpaque of Base.bool
DAttrOpaque b is opaque_full is b is true.

| DAttrTailRecursion of Base.bool

| DAttrTimeLimitMultiplier of Base.int

| DAttrTypeQuantifier

type d_qualified_name = Base.string Base.list

type d_decl_modifier =

| DDmNone

| DDmAbstract

| DDmGhost

| DDmStatic

val d_decl_modifier_to_string : d_decl_modifier -> string

type d_variance = Base.int * Base.string

type d_generic_param = d_variance Base.option * Base.string

A generic parameter has an optional variance ( *, +, ! or - ) and a type name.

and d_domain_type =

| DTyBool
The bool type.

| DTyChar
The char type.

| DTyInt
The int type.

| DTyReal
The real type.

| DTyOrdinal
The ordinal type.

| DTyBitVector
The bitvector type (incomplete implementation).

| DTyObject
The Object type.

| DTyFiniteSet of d_domain_type
The set<T> type where T is a d_domain_type.

| DTyInfiniteSet of d_domain_type
The iset<T> type where T is a d_domain_type.

| DTyMultiset of d_domain_type
The multiset<T> type where T is a d_domain_type.

| DTyFiniteMap of d_domain_type * d_domain_type
The map<K,T> type where K,T are d_domain_type.

| DTyInfiniteMap of d_domain_type * d_domain_type
The imap<K,T> type where K,T are d_domain_type.

| DTySequence of d_domain_type
The seq<T> type where T is a d_domain_type.

| DTyNat
The nat type.

| DTyString
The string type.

| DTyArray of d_domain_type
The array<T> type.

| DTyTuple of d_domain_type Base.list
The tuple<T1,T2, .., Tn> type.

| DTyNamed of Base.string * d_domain_type Base.list
A named type is a type typename<typeparams>. This is a simplification of the NamedType of the Dafny reference. For example, DTyNamed(["list", [DTyInt]]) is the type list<int>.

| DTyComposite of (Base.string * d_domain_type Base.list) Base.list

type d_ident_type = Base.string * d_domain_type

type d_datatype_constr_decl = {

dconstr_attributes : d_attribute Base.option;
Optional attribute of the constructor

dconstr_name : Base.string;
The constructor name.

dconstr_args : (Base.string Base.option * d_domain_type) Base.list;
The constructor arguments are a list of types, optionally with a name. For example, a list datatype declaration can be: datatype List<T> = Nil | Cons(head: T, tail: List<T>) where each consntructor argument is named, or datatype List<T> = Nil | Cons(T, List<T>) which is simpler, but then no destructor is created in Dafny.

}

A datatype constructor declaration, e.g. Cons(head: int, tail: list). A datatype constructor is also called a DatatypeMember in the Dafny reference.

type d_body =

| Body of Base.string
A string body for hardcoded parts.

| DMatch of d_body * (Lang.Term.pattern * d_body) Base.list
A pattern matching construct.

| DTerm of Lang.Term.term
A wrapper for a term.

| DBlock of d_body Base.list
A block of d_body in braces, each separeted by semicolon.

| DAssert of Lang.Term.term
Assert(term)

| DCalc of d_body Base.list
Calc statement.

| DAssign of Lang.Term.term * d_body
Assign statement.

| DStmt of d_body
A body statement with a semicolon after it

type d_clause = Lang.Term.term

type d_spec_kind =

| DSpecMethod
A method specification kind.

| DSpecFunction
A function specification kind.

| DSpecLambda
A lambda specification kind.

| DSpecIterator
An iterator specification kind.

| DSpecLoop
A loop specification kind.

type d_spec = {

dspec_kind : d_spec_kind;
The kind of specification (method, function, ...) This changes how the specification is printed by filtering only the clauses allowed in the given specification.

dspec_decreases : d_clause Base.list;
The list of decreases clauses.

dspec_ensures : d_clause Base.list;
The list of ensures clauses.

dspec_modifies : d_clause Base.list;
The list of modifies clauses.

dspec_requires : d_clause Base.list;
The list of requires clauses.

dspec_reads : d_clause Base.list;
The list of reads clauses

dspec_invariant : d_clause Base.option;
The loop invariant clause.

}

This type is used for all kinds of specifications. Some fields should not appear in some specifications, for example if the spec if for a function, the modifies clauses will not be used.

val dspec_is_empty : d_spec -> Base.bool

type d_method_kind =

| DMkMethod
A simple method.

| DMkConstructor
A constructor method.

| DMkLemma of Base.bool
DMkLemma(b) is "twostate lemma" if b is true, otherwise just "lemma"

| DMkLemmaExtreme of Base.bool
DMkLemmaExtreme(true) is "greatest lemma", if false "least lemma"

The method kind: a simple method, a constructor, a lemma or an extreme lemma. Most used methods are lemmas and constructors.

type d_method_signature = {

dmsig_params : d_generic_param Base.list;
The optional signature parameters.

dmsig_ktype : d_domain_type Base.option;
Optional, for "least" and "greatest" lemmas.

dmsig_formals : d_ident_type Base.list;
The formal arguments of the method.

dmsig_returns : d_ident_type Base.list;
The return of the method.

}

type d_function_kind =

| DFkFunction of Base.bool
A simple function, with also "method" is bool is true.

| DFkPredicate of Base.bool
DFkPredicate(b) is "predicate method" if b is true, otherwise just "predicate"

| DFkPredicateExtreme of Base.bool
DMkPredicateExtreme(true) is "greatest predicate", if false "least predicate"

The function kind: a simple method, a constructor, a lemma or an extreme lemma. Most used methods are lemmas and constructors.

type d_function_signature = {

dfsig_params : d_generic_param Base.list;
The optional signature parameters.

dfsig_ktype : d_domain_type Base.option;
Optional, for "least" and "greatest" lemmas.

dfsig_formals : d_ident_type Base.list;
The formal arguments of the method.

dfsig_return : Base.string Base.option * d_domain_type Base.list;
The return type of the function.

}

type d_class_member_decl =

| DClassConstantField of d_ident_type Base.list

| DClassFunction of Base.string * d_function_kind * d_attribute Base.list * d_function_signature * d_spec * d_body

| DClassMethod of Base.string * d_method_kind * d_attribute Base.list * d_method_signature * d_spec * d_body

Class members are functions, constant fields and methods.

type d_toplevel = {

dt_modifiers : d_decl_modifier Base.list;

dt_kind : d_decl_kind;

}

and d_decl_kind =

| DSubModuleDefinition of d_attribute Base.list * d_qualified_name * d_qualified_name Base.option * d_toplevel Base.list
DSubModuleDefinition(attributes, name, refname, declarations) is the submodule definition "module" attributes name (refines refname) { declarations }

| DSubModuleImport of Base.bool * d_qualified_name
DSubModuleImport(false, qname) is "import qname". DSubModuleImport(true, qname) is "import opened qname". TODO: consider the different kind of imports in the reference.

| DSubModuleExport
TODO: implement module exports. Probably not needed for this project.

| DClassDecl of d_attribute Base.list * Base.string * d_generic_param Base.list * d_toplevel Base.list
A class declaration. Todo: add the "extends".

| DDatatypeDecl of d_attribute Base.list * Base.string * d_generic_param Base.list * d_datatype_constr_decl Base.list
A Datatype declaration, e.g. datatype type<params> = c1 | c2(param) | ..

| DCodatatypeDecl of d_attribute Base.list * Base.string * d_generic_param Base.list * d_datatype_constr_decl Base.list
A Coinductive datatype declaration, e.g. codatatype type<params> = c1 | c2(param) | ..

| DNewtypeDecl

| DSynonymTypeDecl

| DIteratorDecl

| DTraitDecl

| DClassMemberDecl of d_decl_modifier * d_class_member_decl

The kinds of top level declarations in a Dafny program.

type d_program = {

dp_includes : Base.string Base.list;

dp_topdecls : d_toplevel Base.list;

}

A Dafny program is a list of includes followed by a list of toplevel declarations.

val pp_attribute : Base.Formatter.t -> d_attribute -> Base.unit

val pp_d_qualified_name : Base.Formatter.t -> d_qualified_name -> Base.unit

val pp_d_decl_modifier : Base.Formatter.t -> d_decl_modifier -> Base.unit

val pp_d_includes : Base.Formatter.t -> Base.string Base.list -> Base.unit

val pp_d_typename_segment : (Base.string * d_domain_type Base.list) Fmt.t

val pp_d_domain_type : d_domain_type Fmt.t

val pp_d_datatype_constr_decl : Base.Formatter.t -> d_datatype_constr_decl -> Base.unit

val pp_d_ident_type : Base.Formatter.t -> d_ident_type -> Base.unit

val newl : Stdlib.Format.formatter -> unit -> unit

Force a newline by asking to print 100 spaces or a new line.

val pp_d_term : Lang.Term.term Fmt.t

val pp_clause : Base.string -> Base.Formatter.t -> d_clause -> Base.unit

val pp_d_spec : Base.Formatter.t -> d_spec -> Base.unit

val pp_d_body : d_body Fmt.t

val pp_d_generic_param : Base.Formatter.t -> d_generic_param -> unit

val pp_params : Base.Formatter.t -> d_generic_param Base.list -> Base.unit

val pp_d_method_signature : Base.Formatter.t -> d_method_signature -> Base.unit

val pp_d_function_signature : Base.Formatter.t -> d_function_signature -> Base.unit

val pp_d_method_kind : Base.Formatter.t -> d_method_kind -> Base.unit

val pp_d_function_kind : Base.Formatter.t -> d_function_kind -> Base.unit

val pp_d_class_member_decl : Base.Formatter.t -> d_class_member_decl -> Base.unit

val pp_d_decl_kind : Stdlib.Format.formatter -> d_decl_kind -> Base.unit

val pp_d_toplevel : d_toplevel Fmt.t

val pp_d_program : Base.Formatter.t -> d_program -> Base.unit

val mk_int_type : d_domain_type

val mk_bool_type : d_domain_type

val mk_char_type : d_domain_type

val mk_string_type : d_domain_type

val mk_tuple_type : d_domain_type Base.list -> d_domain_type

val mk_named_type : ?⁠params:d_domain_type Base.list -> Base.string -> d_domain_type

val mk_datatype_constr : ?⁠attr:d_attribute Base.option -> Base.string -> (Base.string Base.option * d_domain_type) Base.list -> d_datatype_constr_decl

val mk_datatype_decl : ?⁠attrs:d_attribute Base.list -> ?⁠params:d_generic_param Base.list -> Base.string -> d_datatype_constr_decl Base.list -> d_decl_kind

val mk_simple_spec : ?⁠decreases:d_clause Base.list -> requires:d_clause Base.list -> ensures:d_clause Base.list -> d_spec_kind -> d_spec

val mk_method_sig : ?⁠params:d_generic_param Base.list -> ?⁠ktype:d_domain_type Base.option -> ?⁠returns:d_ident_type Base.list -> d_ident_type Base.list -> d_method_signature

Create a method signature. By default, there are not parameters, no ktype and no returns.

val mk_lemma : ?⁠attrs:d_attribute Base.list -> Base.string -> d_method_signature -> d_spec -> d_body -> d_decl_kind

Create a lemma. By default, there are no attributes.

val mk_toplevel : ?⁠modifiers:d_decl_modifier Base.list -> d_decl_kind -> d_toplevel

val mk_func : ?⁠attrs:d_attribute Base.list -> Base.string -> d_function_signature -> d_spec -> d_body -> d_decl_kind

val mk_func_sig : ?⁠params:d_generic_param Base.list -> ?⁠ktype:d_domain_type Base.option -> ?⁠returns:(Base.string Base.option * d_domain_type Base.list) -> d_ident_type Base.list -> d_function_signature