Module Common

module AlgoLog : sig ... end

All functions in this module have no effect on the flow of the algorithm. Their only purpose is to log or print information about the current state of the algorithm.

module Env : sig ... end

module ProblemDefs : sig ... end

Common.ProblemDefs: This module contains state variables for the synthesis algorithms, as well as type definitions and printing functions for displaying solutions.

module ProblemFinder : sig ... end