Module Common.Env

type env = {

functions : Lang.PMRS.Functions.ctx;

ctx : Lang.Term.Context.t;

refinement_steps : Base.int Base.ref;

secondary_refinement_steps : Base.int Base.ref;

alpha : Lang.RType.t Base.ref;	(*
The type D in the paper: the output type of the reference and the target recursion skeleton. The first element is the pure type output type of the functions, and the second element is an optional term that represents the additional predicate on the output of the reference function. The term is assumed to define a function (fun (free variables of term) -> term).
*)

tau : Lang.RType.t Base.ref;	(*
The type τ in the paper, input type of the reference function.
*)

theta : Lang.RType.t Base.ref;	(*
The type θ in the paper, input type of the target recursion skeleton.
*)

}

val group : Lang.Term.Context.t -> Lang.PMRS.Functions.ctx -> env

val env_copy : env -> env

val env_create : unit -> env

val (>>>) : env -> ( ctx:env -> 'a ) -> 'a

Execute a function with a full context.

val (>-) : env -> ( ctx:Lang.Term.Context.t -> 'a ) -> 'a

Execute a function with a term context from a full context.

val (@>-) : env -> ( Lang.Term.Context.t -> 'a ) -> 'a

Execute a function with a term context from a full context, without a labelled argument.

val (>>-) : env -> ( fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> 'a ) -> 'a

Execute a function with a term context and function context from a full context.

val ctx_reduce : env -> ?projecting:bool -> ?unboxing:bool -> Lang.Term.term -> Lang.Term.term

Reduce a term with a context.

val (!>) : env -> Base.string

Generate a fresh variable name in an environment.

val incr_refinement : env -> unit

Increment the refinement step count.

val incr_secondary_refinement : env -> unit

Increment the secondary refinement step count.

val get_refinement_steps : env -> Base.int

val get_secondary_refinement_steps : env -> Base.int

val get_theta : env -> Lang.RType.t

val get_tau : env -> Lang.RType.t

val get_alpha : env -> Lang.RType.t

val var_type : env -> Lang.Term.Variable.t -> Lang.RType.t