Module Confsearch

module ConfGraph : sig ... end

A configuration graph represents the partial order between configurations. An algorithm can build the configuration graph using the `next` function and maintain its state using a struct of type `state`.

module Configuration : sig ... end

module Main : sig ... end

module PEnum : sig ... end