Module Confsearch.ConfGraph

A configuration graph represents the partial order between configurations. An algorithm can build the configuration graph using the `next` function and maintain its state using a struct of type `state`.

module G : sig ... end

include module type of struct include G end

type t = Graph__Imperative.Digraph.Concrete(Confsearch.Configuration.Subconf).t

module V = G.V

type vertex = V.t

module E = G.E

type edge = E.t

val is_directed : bool

val is_empty : t -> bool

val nb_vertex : t -> int

val nb_edges : t -> int

val out_degree : t -> vertex -> int

val in_degree : t -> vertex -> int

val mem_vertex : t -> vertex -> bool

val mem_edge : t -> vertex -> vertex -> bool

val mem_edge_e : t -> edge -> bool

val find_edge : t -> vertex -> vertex -> edge

val find_all_edges : t -> vertex -> vertex -> edge list

val succ : t -> vertex -> vertex list

val pred : t -> vertex -> vertex list

val succ_e : t -> vertex -> edge list

val pred_e : t -> vertex -> edge list

val iter_vertex : ( vertex -> unit ) -> t -> unit

val fold_vertex : ( vertex -> 'a -> 'a ) -> t -> 'a -> 'a

val iter_edges : ( vertex -> vertex -> unit ) -> t -> unit

val fold_edges : ( vertex -> vertex -> 'a -> 'a ) -> t -> 'a -> 'a

val iter_edges_e : ( edge -> unit ) -> t -> unit

val fold_edges_e : ( edge -> 'a -> 'a ) -> t -> 'a -> 'a

val map_vertex : ( vertex -> vertex ) -> t -> t

val iter_succ : ( vertex -> unit ) -> t -> vertex -> unit

val iter_pred : ( vertex -> unit ) -> t -> vertex -> unit

val fold_succ : ( vertex -> 'a -> 'a ) -> t -> vertex -> 'a -> 'a

val fold_pred : ( vertex -> 'a -> 'a ) -> t -> vertex -> 'a -> 'a

val iter_succ_e : ( edge -> unit ) -> t -> vertex -> unit

val fold_succ_e : ( edge -> 'a -> 'a ) -> t -> vertex -> 'a -> 'a

val iter_pred_e : ( edge -> unit ) -> t -> vertex -> unit

val fold_pred_e : ( edge -> 'a -> 'a ) -> t -> vertex -> 'a -> 'a

val create : ?size:int -> unit -> t

val clear : t -> unit

val copy : t -> t

val add_vertex : t -> vertex -> unit

val remove_vertex : t -> vertex -> unit

val add_edge : t -> vertex -> vertex -> unit

val add_edge_e : t -> edge -> unit

val remove_edge : t -> vertex -> vertex -> unit

val remove_edge_e : t -> edge -> unit

type state = {

graph : t;	(*
The graph of configurations.
*)

marks : Base.int Base.Hashtbl.M(Confsearch.Configuration.Subconf).t;	(*
A negative mark means unrealizable. A positive mark means a solution has been found. Otherwise, a mark of 0 means it has not been solved.
*)

root : Configuration.Subconf.t;	(*
The maximum configuration of the graph.
*)

super : Configuration.conf;

ctx : Common.Env.env;

}

A type to represent the state of the configuration graph exploration. We need to remember the graph and mark configurations as solved or not.

val mark_unrealizable : state -> Configuration.Subconf.t -> unit

val mark_realizable : state -> Configuration.Subconf.t -> unit

val is_unmarked : state -> Configuration.Subconf.t -> bool

val expand : ?use_po:bool -> state -> Configuration.Subconf.t -> Base.unit

`expand g conf` adds the edges from `conf` to all its refinements in `g`. If `~use_po` is set to false, then the expand algorithm does not check whether the configuration to expand is realizable or not (the partial order between configurations is not used to prune unrealizable cases). `use_po` is `true` by default.

val next : ?shuffle:bool -> state -> Configuration.Subconf.t Base.option

Find the next 0-marked configuration in the graph. Return None if there is no such configuration.

val generate_configurations : Common.Env.env -> Lang.PMRS.t -> state

Generate the inital graph of configurations of a PMRS with unknowns.