Module Confsearch.Configuration

A configuration is a set of argument for unknowns in a recursion skeleton.

type conf = Lang.Term.term Base.list Lang.Term.VarMap.t

A configuration is a map from variables (the unknown functions) to a list of arguments.

val ppm : Common.Env.env -> Base.Formatter.t -> conf -> unit

module Subconf : sig ... end

val of_varset : ( Lang.TermTypes.variable, Base.Comparator.Make(Lang__Term.Variable.T).comparator_witness ) Base.Set.t -> 'a list Lang.Term.VarMap.t

Create an empty configuration of a set of variables.

val check_pmrs : Lang.PMRS.t -> bool

Check that the unknowns of a PMRS occur at most once.

val base_type_args : Common.Env.env -> Lang.PMRS.t -> Lang.Term.VarSet.t -> ( Lang__Term.term, Base__Comparator.Make(Lang__Term.Terms.E).comparator_witness ) Base.Set.t

Generate the base type arguments given a set of variables and a PMRS that defines a set of recursive functions (nonterminals).

val max_configuration : Common.Env.env -> Lang.PMRS.t -> conf

Build the argument map of the PMRS. The argument map is effectively the largest configuration: it is a map from each unknown to all the arguments available to the unknown.

val subconf_count : conf -> Base__Int.t

Count the number of subconfigurations of a given configuration.

val configuration_of : Lang.PMRS.t -> conf

Return the configuration of a PMRS, assuming it has been checked.

val apply_configuration : Common.Env.env -> conf -> Lang.PMRS.t -> Lang.PMRS.t * Common.Env.env

Apply a configuration to a given PMRS. The environment is copied before the type information for the unknowns is changed. The copied environment is then returned.

val get_rstar : Common.Env.env -> Common.ProblemDefs.PsiDef.t -> Base.int -> Lang.Term.TermSet.t * Lang.Term.TermSet.t