Module Configuration.Subconf

type t = Base.int Base.list Utils.IntMap.t

A subconfiguration (relative to some configuration) is a map from an unknown's variable id to a list of integer representing the index of the argument in the super-configuration argument list of that unknown.

val sexp_of_t : t -> Base.Sexp.t

val equal : t -> t -> Base.bool

val compare : t -> t -> int

val hash : t -> Base.int

val of_conf : conf -> t

Return the configuration as a subconf.

val to_conf : conf -> t -> conf

Return the subconf as a configuration, given the configuration it refines.

val drop_arg : t -> t Base.list

Create as many subconfigurations as possible from dropping an argument for an unknown.