Module Confsearch.PEnum

val clear : unit -> unit

This functions enumerate possible extensions of an input problem.

val enumerate_p : ctx:Lang.Term.Context.t -> Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.PsiDef.t list

This functions enumerate possible extensions of an input problem.