Module Frontend.Front

type loc = Stdlib.Lexing.position * Stdlib.Lexing.position

type id = Base.string

type type_decl =

| TDParametric of id Base.list * id * Lang.RType.type_term

| TDSimple of id * Lang.RType.type_term

type termkind =

| FTConst of Lang.Term.Constant.t

| FTApp of term * term Base.list

| FTData of id * term Base.list

| FTVar of id

| FTAny

| FTTup of term Base.list

| FTLet of term * term * term

| FTFun of term Base.list * term

| FTBin of Lang.Term.Binop.t * term * term

| FTHOBin of Lang.Term.Binop.t

| FTUn of Lang.Term.Unop.t * term

| FTIte of term * term * term

and term = {

pos : Stdlib.Lexing.position * Stdlib.Lexing.position;

kind : termkind;

}

val mk_const : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> Lang.Term.Constant.t -> term

val mk_app : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> term -> term Base.list -> term

val mk_data : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> id -> term Base.list -> term

val mk_var : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> Base.string -> term

val mk_any : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> term

val mk_bin : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> Lang.Term.Binop.t -> term -> term -> term

val mk_un : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> Lang.Term.Unop.t -> term -> term

val mk_ite : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> term -> term -> term -> term

val mk_tup : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> term Base.list -> term

val mk_fun : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> term Base.list -> term -> term

val mk_let : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> term -> term -> term -> term

type pmrs_rule = loc * term * term

type pmrs_body = pmrs_rule Base.list

type function_body =

| PmrsBody of loc * pmrs_body

| ExprBody of loc * term

type definition =

| TypeDef of loc * type_decl	(*
A type definition.
*)

| FunDef of loc * id * id Base.list * term Base.option * term	(*
A function definition.
*)

| PMRSDef of loc * id Base.list * id * id Base.list * term Base.option * term Base.option * pmrs_body	(*
A PMRS declaration (loc, syntobjs, name, params, p_requires, p_ensures, body) is a PMRS where:

loc is a location

syntobjs is a list of identifiers of the objects to be synthesized,

name is the name of the function,

params is a list of identifiers of the parameters of the PMRS, i.e. all the arguments except the last argument that is decomposed in the recursion.

p_requires is a predicate from the input type of the PMRS (last arg of the function) to bool, corresponding to the predicate that the inputs must satsify.

p_ensures is a predicate on the output type. Every element in the image of the function must satsify this predicate.

*)

| CamlPMRSDef of loc * id Base.list * term Base.option * term Base.option * (id * id Base.list * function_body) Base.list	(*
TODO: Remove once Caml syntax is fuylly supported.
*)

| SyntObjDecl of loc * definition * id * id	(*
Declares a synthesis objective: SyntObjDecl(loc, decl, f1, f2 ) specifies that the function defined by decl must be equivalent to f1 composed with f2.
*)

| EnsuresDef of loc * id * term	(*
Adds an ensures to a function symbol id.
*)

type program = definition Base.list

val pp_fterm : term Fmt.t

val make_rules : (Base.string * Base.string Base.List.t * function_body) Base.List.t -> (loc * term * term) Base.List.t

val rebuild_pmrs_decl : loc -> id Base.list -> term Base.option -> term Base.option -> (Base.string * Base.string Base.List.t * function_body) Base.List.t -> definition

val preprocess : program -> program * (Lang.RType.ident * Lang.RType.ident * Lang.RType.ident) Base.option