Module Frontend.PmrsFront

module P = Pmrs_parser

module L = Pmrs_lexer

exception SyntaxError of Base.string

val verbose : bool Base.ref

val text : string Base.ref

val loc_fatal_errmsg : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> string -> 'a

val variable_not_found : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> string -> 'a

val parsefile : Base.string -> Front.program * (Lang.RType.ident * Lang.RType.ident * Lang.RType.ident) Base.option

val seek_types : ctx:Lang.Term.Context.t -> Front.program -> unit

Iterates through the toplevel declarations of the program and adds the type declaration to the global type environment in Lang.RType.

val fterm_to_term : ctx:Lang.Term.Context.t -> Lang.Term.VarSet.t -> ( Front.id, Lang.Term.VarSet.elt ) Base.Hashtbl.t -> Lang.Term.VarSet.t -> Front.term -> Lang.Term.term

fterm_to_term loc ~ctx g l t transforms fterm t at location loc into a term, using the variables in ctx, the global variables in g and the local variables in l.

val pmrs_of_rules : ctx:Lang.Term.Context.t -> (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ( Base.string, Lang.Term.variable ) Base.Hashtbl.t -> Lang.Term.variable Base.list -> Lang.Term.variable Base.list -> Lang.Term.variable -> (Front.term Base.option * Front.term Base.option) -> Front.pmrs_body -> Lang.PMRS.t

val translate_function : ctx:Lang.Term.Context.t -> 'a -> ( Front.id, Lang.Term.Variable.t ) Base.Hashtbl.t -> Lang.Term.Variable.t -> Lang.Term.Variable.t Base.list -> Front.term Base.option -> Front.term -> Lang.Term.Variable.t * Lang.Term.fpattern Base.List.t * Lang.Term.term Base.Option.t * Lang.Term.term

val translate : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Front.program -> ( Base.String.t, Lang.PMRS.t, Base.String.comparator_witness ) Base.Map.t