Module Lang.Analysis

val vars_of_pattern : Term.pattern -> Term.VarSet.t

val free_variables : ?⁠include_functions:bool -> Term.term -> Term.VarSet.t

val has_ite : Term.term -> Base.bool

val has_calls : _to:Term.VarSet.t -> Term.term -> Base.bool

val operators_of : Term.term -> Term.OpSet.t

val is_constant : Term.term -> Base.bool

val is_norec : Term.term -> Base.bool

val is_novariant : Term.term -> Base.bool

A term has no variant if it is a variable and the type of that variable has no variants. This means that term cannot be expanded.

val is_bounded : Term.term -> Base.bool

val subst_args : Term.fpattern Base.list -> Term.term Base.list -> (Term.term * Term.term) Base.list Base.option

val replace_calls_to : Term.VarSet.t -> Term.term -> Term.term

val apply_projections : (Base.int, Term.variable Base.list, Base.Int.comparator_witness) Base.Map.t -> Term.term -> Term.term

val replace_id_calls : func:Term.variable -> Term.term -> Term.term

val unify : Term.term Base.list -> Term.term Base.option

val matches : ?⁠boundvars:(Lang__Term.Variable.t, Lang__Term.Variable.comparator_witness) Base.Set.t -> Term.term -> pattern:Term.term -> Term.term Term.VarMap.t Base.option

matches ~boundvars ~pattern t returns a map from variables in pattern to subterms of t if the term t matches the pattern. During the matching process the variables in boundvars cannot be substituted in the pattern.

val matches_subpattern : ?⁠boundvars:(Lang__Term.Variable.t, Lang__Term.Variable.comparator_witness) Base.Set.t -> Term.term -> pattern:Term.term -> (Term.term * (Term.term * Term.term) Base.list * Term.VarSet.t) Base.option

matches_subpattern ~boundvars t ~pattern returns some triple t', subs, vs when some sub-pattern of pattern matches t.

val matches_pattern : Term.term -> Term.pattern -> Term.term Term.VarMap.t Base.option

matches_pattern t p returns Some map if t matches p, where the map binds the variables of the pattern to the relevant subterms of t. If t does not match the pattern p then None is returned.

val skeletize : functions:Term.VarSet.t -> Term.term -> Term.term Term.VarMap.t * Term.term

skeletize ~functions t returns the term t where each subterm that does not contain a call to the functions in functions has been reblaced by a box containing a fresh variable. Returns a pair of a boxed var-substitution list and the term with the boxes replacing the subexpressions.

val is_expandable_var : Term.variable -> (Term.variable * (Base.string * RType.t Base.list) Base.list) option

val expand_once : Term.term -> Term.term Base.list

val variant_no_recurse : (Base.string * RType.t Base.list) Base.list -> Term.term Base.list

Given a list of variants, returns a list of terms that corresponds to the non-recursive constructors, with their arguments instantiated as variables.

val terms_of_max_depth : Base.int -> RType.t -> Term.term Base.list

type virtual_term =

| VConstr of Base.string * virtual_term Base.list

| VChoice of virtual_term Base.list

| VVarLeaf of Term.Variable.t

val pp_virtual_term : virtual_term Fmt.t

val virtual_variant_no_recurse : (Base.string * RType.t Base.list) Base.list -> virtual_term

val virtual_term_of_max_depth : Base.int -> RType.t -> virtual_term

val pick_some : virtual_term -> Term.term Base.option * virtual_term Base.option

Pick some term in a vitual term, removing the choices made from the virtual term.

val _INT_MAX : int

Arbitrary min and max values. Actual useful values will be generated by solvers anyway.

val _INT_MIN : int

val rand_const_of_type : RType.t -> Term.term

val concretize : ?⁠model:(Base.String.t, Term.term, Base.String.comparator_witness) Base.Map.t -> Term.term -> Term.term