Module Lang.Eval

val simplify : Term.term -> Term.term

Simplifies a term using rules rewriting terms containing constants. If all the leaves of the term are constants, evaluates the term to a constant.

val in_model : ?no_simplify:bool -> ctx:Term.Context.t -> Term.term Term.VarMap.t -> Term.term -> Term.term