Module Lang.Matching

val unify : Term.term Base.list -> Term.term Base.option

val matches : ?boundvars: ( Term.VarSet.elt, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t -> ctx:Term.Context.t -> Term.term -> pattern:Term.term -> Term.term Term.VarMap.t Base.option

matches ~boundvars ~pattern t returns a map from variables in pattern to subterms of t if the term t matches the pattern. During the matching process the variables in boundvars cannot be substituted in the pattern.

val matches_subpattern : ?boundvars: ( Term.VarSet.elt, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t -> ctx:Term.Context.t -> Term.term -> pattern:Term.term -> (Term.term * (Term.term * Term.term) Base.list * Term.VarSet.t) Base.option

matches_subpattern ~boundvars t ~pattern returns some triple t', subs, vs when some sub-pattern of pattern matches t.

val matches_pattern : Term.term -> Term.pattern -> Term.term Term.VarMap.t Base.option

matches_pattern t p returns Some map if t matches p, where the map binds the variables of the pattern to the relevant subterms of t. If t does not match the pattern p then None is returned.

val skeletize : functions:Term.VarSet.t -> ctx:Term.Context.t -> Term.term -> Term.term Term.VarMap.t * Term.term

skeletize ~functions t returns the term t where each subterm that does not contain a call to the functions in functions has been reblaced by a box containing a fresh variable. Returns a pair of a boxed var-substitution list and the term with the boxes replacing the subexpressions.