Module Lang.PMRS

A pattern in a PMRS rewrite rule is limited to matching terms of the form "Constructor_name(arguments)" where each argument is a term, and the Constructor_name is the name of a valid datatype constructor.

type rewrite_rule = Term.variable * Term.variable Base.list * Term.pattern Base.option * Term.term

A PMRS rewrite rule (v, args, pattern, rhs) is the rewrite rule v(args,pattern) -> rhs. v which is a non-terminal (a recursive function in the OCaml code), args is a possibly empty list of arguments, pattern is an optional pattern, rhs is the production of the rule.

type top_function = Term.variable * Term.variable Base.list * Term.term

A toplevel function is a triple of a variable (the variable representing the function itself), a list of variables (the arguments of the function) and a term (the body of the function).

type t = {

pvar : Term.Variable.t;	(*
The main function symbol, storing general information about the PMRS. The specification associated to the PMRS may be found using: get_spec pvar The type of the PMRS (the main function of the group of mutually recursive functions) is stored as the type of pvar.
*)

pinput_typ : RType.t Base.list;	(*
The input type(s). For now, it should be a singleton list. The input type is only the type of the recursively-typed argument of the PMRS, not the parameters.
*)

poutput_typ : RType.t;	(*
Output type and optional invariant on output of function.
*)

pargs : Term.variable Base.list;	(*
Parameter arguments.
*)

psyntobjs : Term.VarSet.t;	(*
The unknowns to synthesize.
*)

prules : rewrite_rule Utils.IntMap.t;	(*
The rules of the PMRS.
*)

pnon_terminals : Term.VarSet.t;	(*
Non-terminals of the PMRS.
*)

pmain_symb : Term.variable;	(*
The main symbol of the PMRS.
*)

porder : Base.int;	(*
The order of the PMRS (mostly useless for now).
*)

plogic : SmtLogic.logic_info;

}

type 'a xresult = ('a, (Base.string * Base.Sexp.t) Base.list) Base.Result.t

A result type with error messages in the form of title x s-expression.

type 'a sresult = ('a, (Base.string * Term.term) Base.list) Base.Result.t

A result type with error messages in the form of title x term.

type variables = Term.variable Base.Map.M(Base.String).t

The type for maps from variable name to variable.

val _globals : (Base.int, t) Base.Hashtbl.t

Table of all the PMRS in the file, indexed by the function variable id.

val find_by_name : Base.string -> Term.variable Base.option

Find a PMRS by name. The name of a PMRS is the name of its pvar.

val _nonterminals : (Base.int, t) Base.Hashtbl.t

Mapping nonterminals to the PMRS they belong to, index by the nonterminal variable id.

val find_nonterminal_by_name : Base.string -> Term.variable Base.option

Find a non-terminal (represented by a variable) from its name.

val lhs : (Term.variable * Term.variable Base.List.t * Term.pattern option * Term.term) -> Term.term

Generate the term that corresponds to the left hand side of a rewrite rule in a PMRS.

val reinit : unit -> unit

Reinitialize tables of PMRS and nonterminals.

val pp_rewrite_rule : Base.Formatter.t -> rewrite_rule -> Base.unit

val pp : Base.Formatter.t -> t -> Base.unit

val pp_ocaml : Base.Formatter.t -> t -> Base.unit

Pretty-print a PMRS as a set of OCaml functions.

val update_order : t -> t

Update the order of the pmrs.

val set_logic_info_of_pmrs : t -> t

val clear_pmrs_types : t -> t

Clear the type information stored in the different components of the PMRS. Clears the type information of:

the local variables in each rule of the PMRS.

the pvar of the PMRS,

the pmain_symb of the PMRS,

the unknowns in the PMRS.

val infer_pmrs_types : t -> t

val unify_two_with_update : (RType.t * RType.t) -> (RType.t * RType.t) -> Base.unit

val unify_one_with_update : (RType.t * RType.t) -> unit

val extract_rec_input_typ : t -> RType.t

val func_to_pmrs : Term.Variable.t -> Term.fpattern Base.list -> Term.term -> t

val build_match_cases : t -> 'a -> Term.variable Base.List.t -> rewrite_rule Base.list -> (Term.term * Term.match_case Base.list) Base.option

val single_rule_case : Term.variable Base.List.t -> (Term.variable Base.List.t * Term.term) -> Term.term

val vars_and_formals : t -> Term.variable -> Lang__Term.variable Base.List.t

val func_of_pmrs : t -> Term.function_descr Base.list

val inverted_rule_lookup : ?boundvars:(Lang__Term.variable, Base__Comparator.Make(Lang__Term.Variable.T).comparator_witness) Base.Set.t -> ('a, Term.variable * Term.variable Base.List.t * Term.pattern option * Term.term, 'b) Base.Map.t -> Term.term -> Term.term Base.list -> ('a, Term.term, 'b) Base.Map.t

inverted_rule_lookup searches for rules whose rhs match (func args), and return a map from rule id to the lhs of the rules matching (func args), with the appropriate substitutions performed.

val subst_rule_rhs : p:t -> (Term.term * Term.term) Base.list -> t

Apply a substitution to all the right hand side of the PMRS rules.

val depends : t -> t Base.list

Given an input PMRS, returns a list of PMRS that this PMRS depends on.

val update_output_type : t -> t

Updates and returns the output type of a PMRS.