Module PMRS.Functions

Context for:

all PMRS in the file, indexed by the function variable id.

all the nonterminals, indexed by the function variable id.

type ctx = {

globals : ( Base.int, t ) Base.Hashtbl.t;

nonterminals : ( Base.int, Base.int ) Base.Hashtbl.t;

}

val create : Base.unit -> ctx

Create an empty context, with an empty table of functions.

val clear : ctx -> Base.unit

Clear the function tables.

val copy : ctx -> ctx

Copy a function context (copies the tables)

val register_global : ctx -> t -> Base.unit

Add a global function to the table of functions globals inside the context.

val find_global : ctx -> Base.int -> t Base.option

Search for a PMRS with the given id in the global table of functions.

val find_by_name : ctx -> Base.string -> Term.variable Base.option

Find a PMRS by name. The name of a PMRS is the name of its pvar.

val register_nonterminal : ctx -> Base.int -> t -> Base.unit

Register a non-terminal symbol within a PRMS

val find_nonterminal : ctx -> Base.int -> t Base.option

Find a non-terminal symbol in the current context (returns the PMRS that defines it)

val find_nonterminal_by_name : ctx -> Base.string -> Term.variable Base.option

Find a non-terminal (represented by a variable) from its name.

val update : ctx -> t -> Base.unit

update ctx p registers the PMRS p and all its non-terminals inside the context ctx.