Module Lang.RType

val _tid : int Base.ref

val _tvar_idx : int Base.ref

type ident = Base.string

type typekind =

| TyInt
The int type.

| TyBool
The bool type.

| TyString
The string type.

| TyChar
The char type.

| TyFun of type_term * type_term
The function type.

| TyTyp of ident

| TyParam of ident

| TyConstr of type_term Base.list * type_term

| TySum of type_term Base.list

| TyVariant of ident * type_term Base.list

Used for type declarations

and type_term = {

pos : Stdlib.Lexing.position * Stdlib.Lexing.position;

tkind : typekind;

}

val dummy_loc : Stdlib.Lexing.position * Stdlib.Lexing.position

val mk_t_int : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term

val mk_t_bool : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term

val mk_t_string : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term

val mk_t_char : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term

val mk_t_typ : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ident -> type_term

val mk_t_param : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ident -> type_term

val mk_t_constr : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term Base.list -> type_term -> type_term

val mk_t_sum : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term Base.list -> type_term

val mk_t_variant : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ident -> type_term Base.list -> type_term

val mk_t_fun : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term -> type_term -> type_term

type t =

| TInt

| TBool

| TString

| TChar

| TNamed of ident

| TTup of t Base.list

| TFun of t * t

| TParam of t Base.list * t

| TVar of Base.int

val get_fresh_tvar : unit -> t

val pp : t Fmt.t

val fun_typ_unpack : t -> t Base.list * t

val fun_typ_pack : t Base.list -> t -> t

val base_name : t -> Base.string Base.option

val _unsafe_pair_t_ : t -> Base.string * t

val _types : (ident, t) Base.Hashtbl.t

This hashtable maps type names to the type term of their declaration. It is initialized with the builtin types int, bool, char and string.

val _variant_to_tname : (Base.string, Base.string) Base.Hashtbl.t

This hashtable maps variant names to the type name. Variant names must be unique!

val _tname_to_variants : (Base.string, Base.string Base.list) Base.Hashtbl.t

val _variants : (Base.string, t * t Base.list) Base.Hashtbl.t

val get_type : Base.string -> t option

val dump_types : Base.Formatter.t -> unit -> unit

val reinit : unit -> Base.unit

val substitute : old:t -> by:t -> in_:t -> t

val sub_all : (t * t) Base.list -> t -> t

val subtype_of : t -> t -> Base.bool

val occurs : Base.int -> t -> Base.bool

type substitution = (Base.int * t) Base.list

A substitution is a list pairs of type variable id, type. Applied to a (parametric) type to subsitute tyep parameters with types.

val unify_one : t -> t -> (substitution, Base.Sexp.t) Base.Result.t

val mkv : substitution -> (t * t) Base__List.t

val unify : (t * t) Base__List.t -> (substitution, Base.Sexp.t) Base.Result.t

val merge_subs : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> substitution -> substitution -> substitution

val type_of_variant : Base.string -> (t * t Base.list) Base.option

val instantiate_variant : type_term Base.list -> (ident * Base.int) Base.list -> t Base.List.t

val add_variant : variant:Base.string -> typename:Base.string -> (t * t Base.list) -> unit

val add_all_variants : params:(ident * Base.int) Base.list -> main_type:t -> typename:ident -> type_term Base.list -> Base.unit

val get_variants : t -> (Base.string * t Base.list) Base.list

get_variants t returns a list of variants of a given type. If the type is a sum type with constructors, it returns a list of pairs of constructors, list of the types of the arguments of the constructor. If the type has no variant, the list is empty.

val is_recursive_variant : (Base.string * t Base.list) -> bool

val add_type : ?⁠params:ident Base.list -> typename:Base.string -> type_term -> (unit, Base.Sexp.t) Base.result

`add_type ?params ~typename tterm` adds a type with name typename and parameters params (default empty list) and type term tterm to the global store.

val tuple_map : (t Base.list * Base.string) Base.list Base.ref

val make_tuple_name : t Base.list -> Base.string

val get_tuple_type_name : t Base.list -> Base.string

val get_tuple_constr_name : t Base.list -> Base.string

val get_tuple_proj_name : t Base.list -> Base.int -> Base.string

val reduce : case:((t -> 'a) -> t -> 'a Base.option) -> init:'a -> join:('a -> 'a -> 'a) -> t -> 'a

val is_base : t -> bool

val is_recursive : t -> Base.bool

val is_datatype : t -> bool

val is_user_defined : t -> Base.bool

val get_datatype_depends : t -> t Base.List.t