Module Lang.SmtInterface

val rtype_of_smtSort : Smtlib.SmtLib.smtSort -> RType.t option

`rtype_of_smtSort smt_sort` return `Some t` if `t` is a type representing the smt sort `smt_sort`, and `None` if no such type exists.

val sort_of_rtype : RType.t -> Smtlib.SmtLib.smtSort

val smt_of_term : Term.term -> Smtlib.SmtLib.smtTerm

`smt_of_term t` generates a smt-term from the term t. If t is not a valid term, raises errors. TODO: return a Result.t instead of failing.

val sorted_vars_of_vars : Term.VarSet.t -> Smtlib.SmtLib.smtSortedVar list

Convert a set of variables into a list of smt sorted vars.

val smt_of_pmrs : PMRS.t -> Smtlib.SmtLib.command list

Convert a PMRS into a set of commands. Once the commands have been passed to the solver, the PMRS can be used as a function in the solver.

val mk_def_fun_command : string -> (string * RType.t) list -> RType.t -> Term.term -> Smtlib.SmtLib.command

val mk_assert : Smtlib.SmtLib.smtTerm -> Smtlib.SmtLib.command

val term_of_smt : (string, Term.variable, Base.String.comparator_witness) Base.Map.t -> Smtlib.SmtLib.smtTerm -> Term.term

Convert a term to a smt-term in an environment (a map from string to variables).

type term_model = (string, Term.term, Base.String.comparator_witness) Base.Map.t

A term model: a map from string (variable names) to terms. In most cases these terms will be constants.

val model_to_constmap : Smtlib.SmtLib.solver_response -> term_model

Translate a solver response (a Smtlib.Solvers.solver_response) to a term-model. Returns an empty map if the response is not a SExps (a list of s-expressions) that define a model obtained through (get-model).

val model_to_subst : Term.VarSet.t -> Smtlib.SmtLib.solver_response -> (Term.term * Term.term) list

Translate a solver response (a Smtlib.Solvers.solver_response) to a term substitution list. Returns an empty list if the response is not a SExps (a list of s-expressions) that defines a model obtained through (get-model).

val request_different_models : term_model -> int -> Smtlib.Solvers.online_solver -> term_model list

request_different_models m i s uses solver s to find i models different from m that satisfy the current stack of assertions.

val decls_of_vars : Term.VarSet.t -> Smtlib.SmtLib.command list

Generate a list of commands from a set of variables. The list of commands may then be passsed to a Smtlib.Solvers.online_solver instance through the Smtlib.Solvers.declare_all command.