Module SmtInterface.Commands

val decls_of_vars : ctx:Term.Context.t -> Term.VarSet.t -> Smtlib.SmtLib.command list

val decl_of_tuple_type : RType.t -> Smtlib.SmtLib.command list

val mk_preamble : logic:Smtlib.Logics.logic -> ?induction:bool -> ?incremental:bool -> ?load_defs:bool -> ?models:bool -> ?proofs:bool -> unit -> Smtlib.SmtLib.command list

Produces a list of commands to be used in the preamble of a solver session.

parameter logic
a SMT logic provided as a string.

parameter induction
is set to false by default. If set to true, then the quant-ind option of the solver is set of true and if !Config.induction_proof_tlimit is positive, then the tlimit option of the solver is used to set a timeout on exection of the proof.

parameter incremental
set to false by default.

parameter load_defs
is set to true by default, and the definitions of min and max are added to the solver.

parameter models
is set to true by default. Set to false if you don't need models during the session.

parameter proofs
is set to false by default. Set to true if you want to extract proofs from the solver.

val mk_def_fun : ctx:Term.Context.t -> string -> (string * RType.t) list -> RType.t -> Term.term -> Smtlib.SmtLib.command