Module Lang.SmtLogic

val theory_of : RType.t -> Smtlib.Logics.theory

type logic_info = {

theory : Smtlib.Logics.theory;

linearity : Base.bool;

datatypes : Base.bool;

}

val logic_info_join : logic_info -> logic_info -> logic_info

val base_logic_info : logic_info

val term_requires_datatype : ctx:Term.Context.t -> Term.term -> bool

val infer_logic : ?quantifier_free:bool -> ?with_uninterpreted_functions:bool -> ?for_induction:Base.bool -> ?hint:string option -> ?logic_infos:logic_info Base.List.t -> ctx:Term.Context.t -> Term.term Base.list -> Smtlib.Logics.logic

Infer the logic to use in calls using terms.

parameter quantifier_free
is true by default, set to false if quantifiers will be used in formulas.

parameter with_uninterpreted_functions
is false by default, set to true if uninterpreted functions will be used.

parameter hint
is an optional logic hint. The current implementation returns exaclty this hint but this will change.

parameter logic_infos
a list of logic_infos, possibly from context program