Module Lang.SygusInterface

module SygusSolver : sig ... end

val rtype_of_sort : ctx:Term.Context.t -> Syguslib.Sygus.sygus_sort -> RType.t Base.option

Translates a sygus sort to a type. Returns Some type if the sort corresponds to a valid type, otherwise None.

val sort_of_rtype : ctx:Term.Context.t -> RType.t -> Syguslib.Sygus.sygus_sort

Translate a type to a sygus sort. If the type is a type variable, it is replaced by the sort Int. A term of unassigned type appearing in a constraint should mean the type does not matter, and it can be soundly replaced by Int.

val requires_dt_theory : ctx:Term.Context.t -> RType.t -> Base.bool

Returns true if the type argument requires the logic to include datatypes to reason about objects of that type.

val logic_of_operators : ?nonlinear:Base.bool -> Term.OpSet.t -> Base.string

Returns a string corresponding to the logic that allows to reason about all the operators in the set. Does not include the "DT" prefix required for reasoning about datatypes (see requires_dt_theory).

val dt_extend_base_logic : Base.string -> Base.string

Extend the base logic provided as a string to the same logic with datatypes.

val sygus_of_term : ctx:Term.Context.t -> Term.term -> Syguslib.Sygus.sygus_term

Translates a term to a sygus term.

val term_of_sygus : fctx:PMRS.Functions.ctx -> ctx:Term.Context.t -> ( Base.string, Term.variable, Base.String.comparator_witness ) Base.Map.t -> Syguslib.Sygus.sygus_term -> Term.term

Translates a sygus-term to a term in an environment. The environment is a map from variable name to variable.

val grammar_production_of_skeleton : Rewriter.Skeleton.t -> ints:Syguslib.Sygus.sygus_term -> bools:Syguslib.Sygus.sygus_term -> (Syguslib.Sygus.sygus_term * Syguslib.Sygus.sygus_sort) Base.list -> Syguslib.Sygus.sygus_term Base.list

Returns a list of terms that corresponds to the Skeleton interpreted as a grammar.

val sorted_vars_of_types : ctx:Term.Context.t -> RType.t Base.list -> Syguslib.Sygus.sorted_var Base.list

Given a list of types, returns a list of sorted sygus variables. Returns as many variables as there are elements in the input lists, and a fresh name is used for each created variable.

val sorted_vars_of_vars : ctx:Term.Context.t -> Term.variable Base.list -> Syguslib.Sygus.sorted_var Base.list

Create the list of sorted vars corresponding to the list of variables, in the same order.

val mk_synthfun : ctx:Term.Context.t -> Base.string -> Term.variable Base.list -> RType.t -> Syguslib.Sygus.grammar_def Base.option -> Syguslib.Sygus.command

Create a function to synthesize from its name, arguments, return type and optional grammar.

val mk_synthinv : ctx:Term.Context.t -> Base.string -> Term.variable Base.list -> Syguslib.Sygus.grammar_def Base.option -> Syguslib.Sygus.command

Create an invariant to synthesize from its name, arguments and optional grammar.

val wait_on_failure : Base.int Base.ref -> (Syguslib.Sygus.solver_response * 'a) Lwt.t -> (Syguslib.Sygus.solver_response * 'a) Lwt.t

Bind to the response of a solver to that if that response is a failure, the promise is deferrred by !Config.wait_parallel_tlimit.

module HLSolver : sig ... end