Module Term.Attributes

Variables have unique integer ids but two variables can have the same name. Additional information can be added via variable attributes. For example, a variable can be a Terminal or a NonTerminal in the context of a pattern matching recursion scheme.

module Elt : sig ... end

module AS : sig ... end

include module type of struct include AS end

type nonrec t = ( Elt.t, Elt.comparator_witness ) Base__Set.t

type elt = Elt.t

val singleton : Elt.t -> ( Elt.t, Elt.comparator_witness ) Base.Set.t

val empty : ( Elt.t, Elt.comparator_witness ) Base.Set.t