Module Term.Variable

module T : sig ... end

type t = TermTypes.variable

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val compare : t -> t -> int

val equal : t -> t -> bool

val (=) : t -> t -> bool

val hash : 'a -> int

type comparator_witness = Base.Comparator.Make(T).comparator_witness

val comparator : ( t, comparator_witness ) Base.Comparator.t

val vtype_assign : TermTypes.Context.t -> t -> RType.t -> unit

Assign a type to a variable.

val vtype : TermTypes.Context.t -> t -> RType.t option

Return the type of the variable, or None if it has no assigned types.

val id : t -> int

Return the id of the variable (function for v.vid)

val get_name : TermTypes.Context.t -> int -> string option

Get the name of a variable identified by its integer id, and return None is there is no such variable.

val clear_type : TermTypes.Context.t -> t -> unit

Forget the global type of the variable.

val vtype_or_new : TermTypes.Context.t -> t -> RType.t

Return the type of a variable, or create a new type and return it if the variable doesn't have a type. That new type becomes the type of the variable.

val update_var_types : TermTypes.Context.t -> (RType.t * RType.t) list -> unit

Apply a type substitution to all the knownn variables.

val mk : ?attrs: ( TermTypes.Attributes.elt, TermTypes.Attributes.Elt.comparator_witness ) Base.Set.t -> ?t:RType.t option -> TermTypes.Context.t -> string -> t

Create variable. mk ~attrs ~t:(Some typ) name creates a variable with name name ,type typ and attributes attrs. The variable is registered as well as its type if some type is provided, otherwise a type variable is generated.

val is_anonymous : t -> bool

Returns true if the variable has the Anonymous attribute.

val make_anonymous : t -> t

Add the Anonymous attribute to the variable.

val is_builtin : t -> bool

Returns true if the variable has the Builtin attribute.

val make_builtin : t -> t

Add the Builtin attribute to the variable.

val has_attr : TermTypes.Attributes.elt -> t -> bool

Check whether the variable has an attribute.

val is_nonterminal : t -> bool

is_nonterminal v returns true if v is a nonterminal in a PMRS. Inspects the attributes of the variable.

val same_name : t -> t -> bool

same_name v1 v2 eturns true if v1 and v2 have the same name.

val pp : TermTypes.Context.t -> Stdlib.Format.formatter -> t -> unit

val pp_id : Stdlib.Format.formatter -> t -> unit

val pp_typed : TermTypes.Context.t -> Stdlib.Format.formatter -> t -> unit

val free : TermTypes.Context.t -> t -> unit

val print_summary : Stdlib.Format.formatter -> TermTypes.Context.t -> unit