Module Lang.TermTypes

module O = Base.Option

module Attributes : sig ... end

Variables have unique integer ids but two variables can have the same name. Additional information can be added via variable attributes. For example, a variable can be a Terminal or a NonTerminal in the context of a pattern matching recursion scheme.

type variable = {

vname : Base.string;

vid : Base.int;

vattrs : Attributes.t;

}

val dummy_loc : Stdlib.Lexing.position * Stdlib.Lexing.position

val pp_nice : bool Base.ref

Terms.

module Binop : sig ... end

module Unop : sig ... end

module Operator : sig ... end

module OpSet : sig ... end

module Constant : sig ... end

type fpattern =

| FPatAny

| FPatVar of variable

| FPatTup of fpattern Base.list

Simple patterns for function arguments: a fpattern is either a variable or a tuple of patterns.

type pattern =

| PatAny

| PatVar of variable

| PatConstant of Constant.t

| PatTuple of pattern Base.list

| PatConstr of Base.string * pattern Base.list

More complex patterns are used in match-cases. In the current implementation, this is only used as a way to translate a PMRS back to a set of mutually recursive functions.

type termkind =

| TApp of term * term Base.list	(*
A function application.
*)

| TBin of Binop.t * term * term	(*
A binary operation.
*)

| TBox of term	(*
A boxed term, used in transformation to "protect" its contents.
*)

| TConst of Constant.t	(*
A constant.
*)

| TData of Base.string * term Base.list	(*
A datatype constructor.
*)

| TFun of fpattern Base.list * term	(*
A function, where each argument capture by a f-pattern.
*)

| TIte of term * term * term	(*
A conditional.
*)

| TMatch of term * match_case Base.list	(*
A pattern matching construct.
*)

| TSel of term * Base.int	(*
A tuple projection. TSet(TTup tl, i) is List.nth tl i.
*)

| TTup of term Base.list	(*
A tuple.
*)

| TUn of Unop.t * term	(*
A unary operation.
*)

| TVar of variable	(*
A variable.
*)

and match_case = pattern * term

and term = {

tpos : Stdlib.Lexing.position * Stdlib.Lexing.position;

tkind : termkind;

ttyp : RType.t;

}

type function_descr = {

f_var : variable;

f_args : pattern Base.list;

f_body : term;

}

module Context : sig ... end