Module Variables.VarSet

module V : sig ... end

type nonrec t = ( TermTypes.variable, Base.Comparator.Make(Variable.T).comparator_witness ) Base.Set.t

type elt = TermTypes.variable

val empty : ( elt, Base.Comparator.Make(Variable.T).comparator_witness ) Base.Set.t

val singleton : elt -> ( elt, Base.Comparator.Make(Variable.T).comparator_witness ) Base.Set.t

val union_list : ( elt, Base.Comparator.Make(Variable.T).comparator_witness ) Base.Set.t list -> ( elt, Base.Comparator.Make(Variable.T).comparator_witness ) Base.Set.t

val elements : ( 'a, 'b ) Base.Set.t -> 'a list

val of_list : elt list -> ( elt, Base.Comparator.Make(Variable.T).comparator_witness ) Base.Set.t

val map : ( 'a -> elt ) -> ( 'a, 'b ) Base.Set.t -> V.t

val filter_map : t -> f:( elt -> elt option ) -> t

val has_name : t -> string -> bool

has_name vs n returns true if there is a variable with name n in the set vs

val find_by_name : t -> string -> elt option

Find a variable by its name.

val vids_of_vs : t -> int list

Returns the list of ids of the variables in the set.

val has_vid : t -> int -> bool

has_vid vs i returns true if there is a variable with id id in the set vs

val find_by_id : t -> int -> elt option

Find a variable by its id.

val bindings : t -> (int * elt) list

Returns the association list of variable ids and variables.

val names : t -> string list

Return the list of names of the variables in the set.

val filter_by_type : TermTypes.Context.t -> t -> RType.t -> t

Filter a set with a type.

val record : TermTypes.Context.t -> t -> (string * RType.t) list

Create a record type from a set of variables (an association list where keys are the variable names and data is the variable type).

val to_env : t -> ( string, elt, Base.String.comparator_witness ) Base.Map.t

Create an environment (a map from names to variables) from a variable set.

val prime : TermTypes.Context.t -> t -> (elt * elt) list

Given a set of variables, return a list of pairs such that for each element v in the set, there is a pair v,v' in the list where v' is a fresh variable (with a name similar to v)

val add_prefix : t -> string -> t

Change the names of the variables in the set by adding a string prefix. The variables are still equal to the input variable, in the sense that they have the same id, but different names.

val iset : t -> int list -> t

iset vs ids returns the set of variables in vs that have their id in ids.

val pp_var_names : TermTypes.Context.t -> Stdlib.Format.formatter -> ( elt, 'a ) Base.Set.t -> unit

Pretty-print a set just by printing variable names.

val pp : TermTypes.Context.t -> Stdlib.Format.formatter -> ( elt, 'a ) Base.Set.t -> unit

Pretty print a set by printing variable names with their type. Seee pp_var_names for alternative.

val dump : Stdlib.Format.formatter -> ( elt, 'a ) Base.Set.t -> unit

Print the variable set in a non-human readable format.

val of_sh : ( 'a, elt ) Base.Hashtbl.t -> ( elt, Base.Comparator.Make(Variable.T).comparator_witness ) Base.Set.t