Module Lang__Eval

val simplify : Lang.Term.term -> Lang.Term.term

Simplifies a term using rules rewriting terms containing constants. If all the leaves of the term are constants, evaluates the term to a constant.

val in_model : Lang.Term.term Lang.Term.VarMap.t -> Lang.Term.term -> Lang.Term.term