Module Lang__SmtInterface

module Stats : Smtlib.Solvers.Statistics

module SmtLog : Smtlib.Solvers.Logger

module AsyncSmt : sig ... end

module SyncSmt : sig ... end

val string_of_smtSymbol : Smtlib.SmtLib.smtSymbol -> Base.string

val rtype_of_smtSort : Smtlib.SmtLib.smtSort -> Lang.RType.t Base.option

val sort_of_rtype : Lang.RType.t -> Smtlib.SmtLib.smtSort

val dec_parametric : Lang.RType.t -> Lang.RType.t Base.list -> Smtlib.SmtLib.smtSort

val decl_of_tup_type : Lang.RType.t Base.list -> Smtlib.SmtLib.smtSymbol * Smtlib.SmtLib.command

val declare_datatype_of_rtype : Lang.RType.t -> (Smtlib.SmtLib.smtSymbol Base.list * Smtlib.SmtLib.command) Base.List.t

val smtPattern_of_pattern : Lang.Term.pattern -> Smtlib.SmtLib.smtPattern

val term_of_const : Lang.Term.Constant.t -> Smtlib.SmtLib.smtTerm

val smt_of_term : Lang.Term.term -> Smtlib.SmtLib.smtTerm

val smt_of_case : Lang.Term.match_case -> Smtlib.SmtLib.match_case

val constant_of_smtConst : Smtlib.SmtLib.smtSpecConstant -> Lang.Term.Constant.t

type id_kind =

| ICstr of Base.string

| IVar of Lang.Term.variable

| IBinop of Lang.Term.Binop.t

| IUnop of Lang.Term.Unop.t

| IBool of Base.bool

| INotDef

val id_kind_of_s : (Base.string, Lang.Term.variable, 'a) Base.Map.t -> Base.string -> id_kind

val term_of_smt : (Base.string, Lang.Term.variable, Base.String.comparator_witness) Base.Map.t -> Smtlib.SmtLib.smtTerm -> Lang.Term.term

val sorted_vars_of_vars : Lang.Term.VarSet.t -> Smtlib.SmtLib.smtSortedVar Base.list

type term_model = (Base.string, Lang.Term.term, Base.String.comparator_witness) Base.Map.t

val constmap_of_s_exprs : (Base.string, Lang.Term.term, Base.String.comparator_witness) Base.Map.t -> Base.Sexp.t Base.list -> (Base.string, Lang.Term.term, Base.String.comparator_witness) Base.Map.t

val model_to_constmap : Smtlib.SmtLib.solver_response -> (Base.String.t, Lang.Term.term, Base.String.comparator_witness) Base.Map.t

val model_to_varmap : Lang.Term.VarSet.t -> Smtlib.SmtLib.solver_response -> Lang.Term.term Lang.Term.VarMap.t

val request_different_models : term_model -> Base.int -> SyncSmt.online_solver -> (Base.String.t, Lang.Term.term, Base.String.comparator_witness) Base.Map.t list

Given a term model, request additional models from the solver from the point where the current model was obtained. Simply calls (get-model) multiple times, asserting that the values must not be equal to the value in the previous model at each new call.

val maybe_decl_of_tuple_type : Lang.RType.t -> Smtlib.SmtLib.command list

val decls_of_vars : Lang.Term.VarSet.t -> Smtlib.SmtLib.command Base.List.t

val smtPattern_of_term : Lang.Term.term -> Smtlib.SmtLib.smtPattern Base.option

val build_match_cases : Lang.PMRS.t -> 'a -> Lang.Term.variable Base.List.t -> Lang.PMRS.rewrite_rule Base.list -> (Smtlib.SmtLib.smtTerm * Smtlib.SmtLib.match_case Base.list) Base.option

val single_rule_case : 'a -> 'b -> Lang.Term.variable Base.List.t -> (Lang.Term.variable Base.List.t * Lang.Term.term) -> Smtlib.SmtLib.smtTerm

val vars_and_formals : Lang.PMRS.t -> Lang.Term.variable -> Lang.RType.t * Lang__Term.variable Base.List.t * (Smtlib.SmtLib.smtSymbol * Smtlib.SmtLib.smtSort) Base.List.t

val _smt_of_pmrs : Lang.PMRS.t -> (Smtlib.SmtLib.smtSymbol Base.list * Smtlib.SmtLib.command) Base.list * Smtlib.SmtLib.command Base.list

val mk_def_fun_command : Base.string -> (Base.string * Lang.RType.t) Base.list -> Lang.RType.t -> Lang.Term.term -> Smtlib.SmtLib.command

val mk_assert : Smtlib.SmtLib.smtTerm -> Smtlib.SmtLib.command

val smt_of_pmrs : Lang.PMRS.t -> Smtlib.SmtLib.command Base.list