Module Lang__Term

module O = Base.Option

val dummy_loc : Stdlib.Lexing.position * Stdlib.Lexing.position

module Attributes : sig ... end

Variables have unique integer ids but two variables can have the same name. Additional information can be added via variable attributes. For example, a variable can be a Terminal or a NonTerminal in the context of a pattern matching recursion scheme.

type variable = {

vname : Base.string;

vid : Base.int;

vattrs : Attributes.t;

}

module Variable : sig ... end

module VarSet : sig ... end

module Binop : sig ... end

module Unop : sig ... end

module Operator : sig ... end

module OpSet : sig ... end

module Constant : sig ... end

type fpattern =

| FPatVar of variable

| FPatTup of fpattern Base.list

Simple patterns for function arguments: a fpattern is either a variable or a tuple of patterns.

type pattern =

| PatAny

| PatVar of variable

| PatConstant of Constant.t

| PatTuple of pattern Base.list

| PatConstr of Base.string * pattern Base.list

More complex patterns are used in match-cases. In the current implementation, this is only used as a way to translate a PMRS back to a set of mutually recursive functions.

type termkind =

| TApp of term * term Base.list
A function application.

| TBin of Binop.t * term * term
A binary operation.

| TBox of term
A boxed term, used in transformation to "protect" its contents.

| TConst of Constant.t
A constant.

| TData of Base.string * term Base.list
A datatype constructor.

| TFun of fpattern Base.list * term
A function, where each argument capture by a f-pattern.

| TIte of term * term * term
A conditional.

| TMatch of term * match_case Base.list
A pattern matching construct.

| TSel of term * Base.int
A tuple projection. TSet(TTup tl, i) is List.nth tl i.

| TTup of term Base.list
A tuple.

| TUn of Unop.t * term
A unary operation.

| TVar of variable
A variable.

and match_case = pattern * term

and term = {

tpos : Stdlib.Lexing.position * Stdlib.Lexing.position;

tkind : termkind;

ttyp : Lang.RType.t;

}

type function_descr = {

f_var : variable;

f_args : pattern Base.list;

f_body : term;

}

val _globals : (Base.string, Variable.t * fpattern Base.list * term Base.option * term) Base.Hashtbl.t

val find_global : Base.string -> variable Base.option

val fpat_ty : fpattern -> Lang.RType.t

val fpat_vars : fpattern -> VarSet.t

val pattern_of_term : term -> pattern

val mk_pat_any : pattern

val mk_pat_var : variable -> pattern

val mk_pat_const : Constant.t -> pattern

val mk_pat_tuple : pattern Base.list -> pattern

val mk_pat_constr : Base.string -> pattern Base.list -> pattern

val mk_var : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> variable -> term

val var_or_none : term -> variable Base.option

val ext_var_or_none : term -> variable Base.list Base.option

val mk_const : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> Constant.t -> term

val mk_box : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> term -> term

val mk_app : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> ?⁠typ:Lang.RType.t option -> term -> term Base.list -> term

val mk_app_v : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> ?⁠typ:Lang.RType.t option -> variable -> term Base.list -> term

val mk_bin : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> ?⁠typ:Lang.RType.t option -> Binop.t -> term -> term -> term

val mk_assoc : Binop.t -> term Base.list -> term Base.option

val mk_data : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> Base.string -> term Base.list -> term

val mk_fun : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> fpattern Base.list -> term -> term

val mk_ite : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> ?⁠typ:Lang.RType.t option -> term -> term -> term -> term

val mk_tup : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> term Base.list -> term

val mk_sel : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> ?⁠typ:Lang.RType.t option -> term -> Base.int -> term

val mk_un : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> ?⁠typ:Lang.RType.t option -> Unop.t -> term -> term

val mk_match : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> term -> match_case Base.list -> term

val mk_let : ?⁠pos:(Stdlib.Lexing.position * Stdlib.Lexing.position) -> ?⁠typ:Lang.RType.t option -> (variable * term) Base.list -> term -> term

val term_of_pattern : pattern -> term

val fpat_to_term : fpattern -> term

val fpat_sub : fpattern -> fpattern -> (term * term) Base.List.t option

val fpat_sub_all : fpattern Base.List.t -> fpattern Base.List.t -> (term * term) Base__List.t option

val sexp_of_term : term -> Base.Sexp.t

val mk_composite_base_type : Lang.RType.t -> term

val term_compare : term -> term -> Base.int

val substitution : (term * term) Base__List.t -> term -> term

val term_equal : term -> term -> bool

val mk_with_fresh_vars : VarSet.t -> term -> VarSet.t * term

module VarMap : sig ... end

val rewrite_with : (term -> term) -> term -> term

`rewrite_with f t` rewrites the term t by applying the rule f bottom-up.

val rewrite_top_down : (term -> term Base.option) -> term -> term

`rewrite_top_down f t` rewrites the term t by applying the rule f top-down.

val rewrite_accum : init:'a -> f:('a -> term -> (term, 'a) Base.Either.t) -> term -> term

`rewrite_accum ~init ~f t` rewrites the term t by applying the rule f in a top-down manner, but as opposed to `rewrite_top_down` the function `f` can use an accumulator that accumulates information during the traversal.

val rewrite_types : (Lang.RType.t * Lang.RType.t) Base.list -> term -> term

val reduce : init:'a -> case:((term -> 'a) -> term -> 'a Base.option) -> join:('a -> 'a -> 'a) -> term -> 'a

`reduce ~init ~case ~join t` reduces the term by reducing each leaf to `init`, and at each node of the syntax tree, using `join` to merge the values. In a top-down traversal, if `case` returns `Some a` then the subterm is not recrusively reduced, but the value `a` is used instead.

val transform : case:((term -> term) -> term -> term Base.option) -> term -> term

val transform_at_depth : Base.int -> case:((term -> term) -> term -> term Base.option) -> term -> term

val transform_info : f:(term -> term) -> term -> term

val remove_boxes : term -> term

val var_count : Lang.RType.t -> term -> Base__Int.t

val var_count_compare : Lang.RType.t -> term -> term -> int

val term_size : term -> Base__Int.t

val term_size_compare : term -> term -> int

val term_height : term -> Base__Int.t

val term_height_compare : term -> term -> int

val proj_var : variable -> term

proj_var v returns the tuple term (v.0, v.1, ..., v.n) if v is a variable of tuple type with n compoenents, otherwise returns v as a term.

val tuplify : term -> term

tuplify t transforms every subterm of t that has a tuple type into a tuple of projections. For example if t is a term of tuple type with n components, it returns (t.1, t.2, .., t.n).

val pp_fpattern : fpattern Fmt.t

val pp_pattern : pattern Fmt.t

val pp_term : Base.Formatter.t -> term -> unit

val pp_subs : Base.Formatter.t -> (term * term) Base.list -> Base.unit

val pp_function_descr : Base.Formatter.t -> function_descr -> Base.unit

val infer_type : term -> term * Lang.RType.substitution

val erase_term_type : term -> term

val type_of : term -> Lang.RType.t

module Terms : sig ... end

module TermSet : sig ... end

val pp_term_set : Base.Formatter.t -> TermSet.t -> unit