Module Lang__Term.Terms

module E : sig ... end

include E

type t = term

val compare : term -> term -> Base.int

val equal : term -> term -> bool

val sexp_of_t : term -> Base.Sexp.t

module C : sig ... end

include C

type comparator_witness = Base__Comparator.Make(E).comparator_witness

val comparator : (E.t, comparator_witness) Base__Comparator.comparator

val substs_of_alist : (variable * term) Base.list -> (term * term) Base.list

val (+) : t -> t -> t

Create a term equal to the addition two terms.

val (-) : t -> t -> t

Create a term equal to the substraction of two terms.

val (*) : t -> t -> t

Create a term equal to the multiplication two terms.

val (/) : t -> t -> t

Create a term equal to the division two terms.

val (||) : t -> t -> t

Create a term equal to the disjunction of two terms.

val (&&) : t -> t -> t

Create a term equal to the conjuction of two terms.

val (>) : t -> t -> t

Create a term equal to the > comparison of two terms.

val (>=) : t -> t -> t

Create a term equal to the >= comparison of two terms.

val (<=) : t -> t -> t

Create a term equal to the <= comparison of two terms.

val (<) : t -> t -> t

Create a term equal to the < comparison of two terms.

val (==) : t -> t -> t

Create a term equal to the equality of two terms.

val max : t -> t -> t

Create a term equal to the max of two terms.

val min : t -> t -> t

Create a term equal to the min of two terms.

val int : Base.int -> t

Create an integer constant term.

val bool : Base.bool -> t

Create a boolean constant term.

val (~!) : term -> t

Create a negation of a term.

val ite : t -> t -> t -> t

Create an if-then-else term.

val (~^) : variable -> t

Create a term from a variable.

val typed : t -> t

Infers the type of the term, and returns the term with the correct types assigned.