Module Lib

module Parsers = Frontend.Parsers

module Utils = Utils

module Algos = Algo.Refinement

module Lang = Lang

val reinit : unit -> unit

Use reinit to reinitialize all the global variables used in Synduce when solving multiple problems.

val solve_file : ?print_info:bool -> Base.string -> Codegen.Commons.problem_descr * (Algo.AState.soln Base.option, Algo.AState.unrealizability_ctex Base.list) Base.Either.t

val get_lemma_hints : unit -> Algo.AState.equation_system

Call get_lemma_hints () after solve_file to get a list of potential useful lemmas for the proof of correctness.