Module Parsers__.PmrsFront

module P = Parsers__.Pmrs_parser

module L = Parsers__.Pmrs_lexer

exception SyntaxError of Base.string

val verbose : bool Base.ref

val text : string Base.ref

val loc_fatal_errmsg : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> string -> 'a

val variable_not_found : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> string -> 'a

val parsefile : Base.string -> Parsers__.Front.program * (Lang.RType.ident * Lang.RType.ident * Lang.RType.ident) Base.option

val seek_types : Parsers__.Front.program -> unit

Iterates through the toplevel declarations of the program and adds the type declaration to the global type environment in Lang.RType.

val fterm_to_term : 'a -> Lang.Term.VarSet.t -> (Base.String.t, Lang.Term.VarSet.elt) Base.Hashtbl.t -> Lang.Term.VarSet.t -> Parsers__.Front.term -> Lang.Term.term

```fterm_to_term loc v g l t``` transforms fterm t at location loc into a term, using the variables in v, the global variables in g and the local variables in l.

val pmrs_of_rules : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> (Base.string, Lang.Term.variable) Base.Hashtbl.t -> Lang.Term.variable Base.list -> Lang.Term.variable Base.list -> Lang.Term.variable -> (Parsers__.Front.term Base.option * Parsers__.Front.term Base.option) -> Parsers__.Front.pmrs_body -> Lang.PMRS.t

val translate_function : 'a -> (Base.String.t, Lang.Term.VarSet.elt) Base.Hashtbl.t -> Lang.Term.Variable.t -> Lang.Term.Variable.t Base.list -> Parsers__.Front.term Base.option -> Parsers__.Front.term -> Lang.Term.Variable.t * Lang.Term.fpattern Base.List.t * Lang.Term.term Base.Option.t * Lang.Term.term

val translate : Parsers__.Front.program -> (Base.String.t, Lang.PMRS.t, Base.String.comparator_witness) Base.Map.t