Module Se2gis.Counterexamples

val smt_unsatisfiability_check : ctx:Lang.Term.Context.t -> Lang.Term.VarSet.t -> Common.ProblemDefs.equation Base.list -> Base.unit

val merge_all : Common.ProblemDefs.unrealizability_witness Base.list -> Common.ProblemDefs.unrealizability_witness Base.list

val reinterpret_model : ctx:Lang.TermTypes.Context.t -> Lang.Term.VarSet.t -> (( Base.string, 'a, 'b ) Base.Map.t * ( Base.string, 'c, 'd ) Base.Map.t) -> (Lang.Term.variable * Lang.Term.variable) Base.List.t -> ( Lang.TermTypes.variable, 'a, Base.Comparator.Make(Lang__Term.Variable.T).comparator_witness ) Base.Map.t * ( Lang.TermTypes.variable, 'c, Base.Comparator.Make(Lang__Term.Variable.T).comparator_witness ) Base.Map.t

val unrealizability_witness_of_constmap : ctx:Lang.TermTypes.Context.t -> (Base.int * Base.int) -> (Common.ProblemDefs.equation * Common.ProblemDefs.equation) -> (Lang.Term.VarSet.t * ( Lang.TermTypes.variable, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t) -> (Lang.Term.variable * Lang.Term.variable) Base.List.t -> ( Base.string, Lang.Term.term, 'a ) Base.Map.t -> Common.ProblemDefs.unrealizability_witness

val skeleton_match : ctx:Lang.Term.Context.t -> unknowns:Lang.Term.VarSet.t -> Lang.Term.term -> Lang.Term.term -> (Lang.Term.term * Lang.Term.term) Base.list Base.option

val components_of_unrealizability : ctx:Lang.Term.Context.t -> unknowns: ( Lang.TermTypes.variable, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t -> Common.ProblemDefs.equation -> Common.ProblemDefs.equation -> ((Lang.Term.term * Lang.Term.term) Base.list * (Lang.Term.term * Lang.Term.term)) Base.option

val gen_info : ctx:Lang.Term.Context.t -> (Common.ProblemDefs.equation * Common.ProblemDefs.equation) -> ( Lang.Term.VarSet.elt, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t -> ( Lang.Term.VarSet.elt, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t * ( Lang.Term.VarSet.elt, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t * ( Lang.Term.VarSet.elt, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t * (Lang.Term.term * Lang.Term.term) Base.List.t * (Lang.Term.VarSet.elt * Lang.Term.VarSet.elt) list

val check_unrealizable : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Lang.Term.VarSet.t -> Common.ProblemDefs.equation_system -> Common.ProblemDefs.unrealizability_witness Base.list Lwt.t * Base.int Lwt.u

Check if system of equations defines a functionally realizable synthesis problem. If any equation defines an unsolvable problem, an unrealizability_witness is added to the list of counterexamples to be returned. If the returned list is empty, the problem may be solvable/realizable. If the returned list is not empty, the problem is not solvable / unrealizable.

val add_cause : Common.ProblemDefs.witness_stat -> Common.ProblemDefs.spurious_cause -> Common.ProblemDefs.witness_stat

val check_image_sat : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.witness -> (Utils.Stats.verif_method * Smtlib.SmtLib.solver_response) Lwt.t * Base.int Lwt.u

val check_image_unsat : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.witness -> (Utils.Stats.verif_method * Smtlib.SmtLib.solver_response) Lwt.t * Base.int Lwt.u

val check_witness_in_image : ?ignore_unknown:bool -> ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.witness -> Common.ProblemDefs.witness

check_witness_in_image ~p witness checks whether the recursion elimination's variables values in the model of witness are in the image of (p.PsiDef.reference o p.PsiDef.repr).

val find_original_var_and_proj : Lang.Term.variable -> ('a * Lang.Term.term) -> ('a * int) Base.Option.t

val mk_model_sat_asserts : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Common.ProblemDefs.witness -> ( 'a -> Lang.Term.term ) -> ( Lang.Term.variable -> 'a Base.Option.t ) -> Smtlib.SmtLib.smtTerm Base.List.t

val check_tinv_unsat : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Lang.PMRS.t -> Common.ProblemDefs.witness -> (Utils.Stats.verif_method * Smtlib.SmtLib.solver_response) Lwt.t * Base.int Lwt.u

check_tinv_unsat ~ctx ~p tinv c checks whether the counterexample c satisfies the predicate tinv in the synthesis problem p. The function returns a promise of a solver response and the resolver associated to that promise (the promise is cancellable). If the solver response is unsat, then there is a proof that the counterexample c does not satisfy the predicate tinv. In general, if the reponse is not unsat, the solver either stalls or returns unknown.

val check_tinv_sat : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Lang.PMRS.t -> Common.ProblemDefs.witness -> (Utils.Stats.verif_method * Smtlib.SmtLib.solver_response) Lwt.t * Base.int Lwt.u

check_tinv_sat ~ctx ~p tinv witness checks whether the counterexample witness satisfies the invariant tinv (a PMRS). The function returns a pair of a promise of a solver response and a resolver for that promise. The promise is cancellable.

val satisfies_tinv : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Lang.PMRS.t -> Common.ProblemDefs.witness -> Common.ProblemDefs.witness

val classify_witnesss : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.witness Base.list -> Common.ProblemDefs.witness Base.list

Classify counterexamples into positive or negative counterexamples with respect to the Tinv predicate in the problem.