Module Se2gis.Elim

val recurs_elim_of_term : ctx:Common.Env.env -> Lang.Term.term -> p:Common.ProblemDefs.PsiDef.t -> (Lang.Term.term * Lang.Term.term) Base.list * Lang.Term.variable Base.list

Given a term and a problem definition, output a recursion elimination map (a list associating some terms of recursive type to terms of bounded type). The second element of the pair returned by this function is the set of scalar variables that would appear in the term after recursion elimination has been applied.

val flatten_rec_elim_tuples : ctx:Lang.Term.Context.t -> (Lang.Term.term * Lang.Term.term) Base.list -> (Lang.Term.term * Lang.Term.term) Base.list

val subs_from_elim_to_elim : ctx:Lang.Term.Context.t -> (Lang.Term.term * Lang.Term.term) Base.list -> (Lang.Term.term * Lang.Term.term) Base.list -> (Lang.Term.term * Lang.Term.term) Base.list

val mk_f_compose_r_orig : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Lang.Term.term -> Lang.Term.term

val mk_f_compose_r_main : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Lang.Term.term -> Lang.Term.term