Module Se2gis.LemmaSynthesis

module S = Smtlib.SmtLib

module Sy = Syguslib.Sygus

module Sm = Syguslib.Semantic

val lemmas_of_context : ctx:Lang.Term.Context.t -> is_pos_witness:Base.bool -> Common.ProblemDefs.witness -> Common.ProblemDefs.term_info

val create_or_update_lemmas_with_witness : ctx:Lang.Term.Context.t -> is_pos_witness:Base.bool -> Common.ProblemDefs.lemmas -> Common.ProblemDefs.witness -> Base.unit

Given a counterexample, create a new term state that contains that counterexample, or if a term state exists for the term associated with that counterexample, update the term state by adding the counterexample's models.

val skeleton_of_tinv : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> Lang.Term.term -> Lang.Skeleton.t Base.Option.t

Applying tinv to the term of a term info can help in findiing a good lemma. This function extract a skeleton for the lemma when the input is the application of tinv to the term of the term detail.

val synthfun_of_det : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Sy.command * Base.string) Base.list

Construct the synthesis objective corresponding to the lemma of the term state detail. For each synthesis strategy implemented, return a command that defines the synthesis objective (SyGuS command). All strategies should be executed in order to find a solution efficiently.

val term_var_string : ctx:Lang.Term.Context.t -> Lang.Term.term -> Base.string

val convert_term_rec_to_witness_rec : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> Common.ProblemDefs.witness -> Base.string -> Base.string

val witness_model_to_args : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Base.string * Sy.sygus_sort) Base.list -> Common.ProblemDefs.witness -> Sy.sygus_term Base.list

val constraint_of_neg_witness : ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> Common.ProblemDefs.witness -> Sy.command

val constraint_of_pos_witness : ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> Common.ProblemDefs.witness -> Sy.command

val log_soln : ctx:Lang.Term.Context.t -> string -> Lang.Term.variable Base.List.t -> Lang.Term.term -> unit

val handle_lemma_synth_response : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> (Sy.solver_response Base.option Lwt.t * Base.int Lwt.u) -> Lang.Term.term Base.list Base.option Lwt.t

val parse_positive_example_solver_model : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> S.solver_response -> Common.ProblemDefs.term_info -> Common.ProblemDefs.witness list

val synthesize_new_lemma : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> Lang.Term.term Base.option Lwt.t

val lemma_refinement_loop : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> Common.ProblemDefs.term_info Base.option Lwt.t

val witnesss_for_lemma_synt : Common.ProblemDefs.witness -> [> `Fst of Common.ProblemDefs.witness | `Snd of Common.ProblemDefs.witness | `Trd of Common.ProblemDefs.witness ]

Partitioning function to partitiion a list into (a,b,c) where a are examples that satisfy the invariant, b are examples that do not satisfy the invariant, c are examples that are spurious for other reasons.

val witnesss_for_ensures_synt : Common.ProblemDefs.witness -> [> `Fst of Common.ProblemDefs.witness | `Snd of Common.ProblemDefs.witness | `Trd of Common.ProblemDefs.witness ]

Partitioning function to partitiion a list into (a,b,c) where a are examples that are not in the reference function's image, b are examples that are in the reference function's image, c are examples that are spurious for other reasons.

val refine_ensures_predicates : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> neg_witnesss:Common.ProblemDefs.witness Base.list -> pos_witnesss:Common.ProblemDefs.witness Base.list -> [ `CoarseningFailure | `CoarseningOk | `Unrealizable ] Lwt.t

val synthesize_lemmas : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> (Sy.solver_response * ( 'a, Common.ProblemDefs.unrealizability_witness Base.List.t ) Base.Either.t) -> Common.ProblemDefs.refinement_loop_state -> ( ( Common.ProblemDefs.refinement_loop_state, Common.ProblemDefs.unrealizability_witness Base.list ) Base.Either.t, Sy.solver_response ) Base.Result.t Lwt.t