Module Se2gis.LemmaVerif

module S = Smtlib.SmtLib

val placeholder_witness : Common.ProblemDefs.term_info -> Common.ProblemDefs.witness

val smt_of_recurs_elim_eqns : ctx:Lang.Term.Context.t -> (Lang.Term.term * Lang.Term.term) Base.list -> p:Common.ProblemDefs.PsiDef.t -> S.smtTerm

Generates a set of smt equations that correspond to the recursion elimination.

val smt_of_ensures : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> S.smtTerm Base.list

Returns a set of smt constraints that assert the know invariants on the functions in the problem.

val smt_of_tinv_app : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> S.smtTerm

val smt_of_lemma_app : Common.ProblemDefs.term_info -> S.smtTerm

val smt_of_lemma_validity : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> S.command list

val inductive_solver_preamble : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Lang.SmtInterface.AsyncSmt.solver -> Common.ProblemDefs.term_info -> unit Lwt.t

val smt_of_disallow_witness_values : ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> S.smtTerm

val set_up_to_get_model : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Lang.SmtInterface.AsyncSmt.solver -> Common.ProblemDefs.term_info -> unit Lwt.t

val mk_model_sat_asserts : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> ( 'a -> Lang.Term.term ) -> ( Lang.Term.variable -> 'a Base.Option.t ) -> Smtlib.SmtLib.smtTerm Base.List.t

val verify_lemma_bounded : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Utils.Stats.verif_method * S.solver_response) Lwt.t * Base.int Lwt.u

val verify_lemma_unbounded : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Utils.Stats.verif_method * S.solver_response) Lwt.t * Base.int Lwt.u

val verify_lemma_candidate : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Utils.Stats.verif_method * Lang.SmtInterface.SyncSmt.solver_response) Lwt.t

Verify a lemma candidate with two parallel calls: one call to an induction solver, another call to a SMT solver. The induction solver attempts to prove that the solution is correct while the smt solver attempt to find a counterexample.