Module Se2gis.Lemmas

module S = Smtlib.SmtLib

module T = Lang.Term

val empty_term_state : Common.ProblemDefs.term_state

val placeholder_witness : Common.ProblemDefs.term_info -> Common.ProblemDefs.witness

val recurs_elim_of_term : ctx:Common.Env.env -> Lang.Term.term -> p:Common.ProblemDefs.PsiDef.t -> (Lang.Term.term * Lang.Term.term) Base.list * Lang.Term.variable Base.list

Given a term and a problem definition, output a recursion elimination map (a list associating some terms of recursive type to terms of bounded type). The second element of the pair returned by this function is the set of scalar variables that would appear in the term after recursion elimination has been applied.

val flatten_rec_elim_tuples : ctx:Lang.Term.Context.t -> (Lang.Term.term * Lang.Term.term) Base.list -> (Lang.Term.term * Lang.Term.term) Base.list

val subs_from_elim_to_elim : ctx:Lang.Term.Context.t -> (Lang.Term.term * Lang.Term.term) Base.list -> (Lang.Term.term * Lang.Term.term) Base.list -> (Lang.Term.term * Lang.Term.term) Base.list

val term_state_of_context : ctx:Lang.Term.Context.t -> is_pos_witness:Base.bool -> Common.ProblemDefs.witness -> Common.ProblemDefs.term_info

val create_or_update_term_state_with_witness : ctx:Lang.Term.Context.t -> is_pos_witness:Base.bool -> Common.ProblemDefs.term_state -> Common.ProblemDefs.witness -> Common.ProblemDefs.term_state

Given a counterexample, create a new term state that contains that counterexample, or if a term state exists for the term associated with that counterexample, update the term state by adding the counterexample's models.

val skeleton_of_tinv : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> Lang.Term.term -> Lang.Rewriter.Skeleton.t Base.Option.t

Applying tinv to the term of a term detail can help in findiing a good lemma. This function extract a skeleton for the lemma when the input is the application of tinv to the term of the term detail.

val synthfun_of_det : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Syguslib.Sygus.command * Base.string) Base.list

Construct the synthesis objective corresponding to the lemma of the term state detail. For each synthesis strategy implemented, return a command that defines the synthesis objective (SyGuS command). All strategies should be executed in order to find a solution efficiently.

val term_var_string : ctx:Lang.Term.Context.t -> Lang.Term.term -> Base.string

val convert_term_rec_to_witness_rec : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> Common.ProblemDefs.witness -> Base.string -> Base.string

val witness_model_to_args : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Base.string * Syguslib.Sygus.sygus_sort) Base.list -> Common.ProblemDefs.witness -> Syguslib.Sygus.sygus_term Base.list

val constraint_of_neg_witness : ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> Common.ProblemDefs.witness -> Syguslib.Sygus.command

val constraint_of_pos_witness : ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> Common.ProblemDefs.witness -> Syguslib.Sygus.command

val log_soln : ctx:Lang.Term.Context.t -> string -> Lang.Term.variable Base.List.t -> Lang.Term.term -> unit

val handle_lemma_synth_response : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> (Syguslib.Sygus.solver_response Base.option Lwt.t * Base.int Lwt.u) -> Lang.Term.term Base.list Base.option

val mk_f_compose_r_orig : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Lang.Term.term -> Lang.Term.term

val mk_f_compose_r_main : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Lang.Term.term -> Lang.Term.term

val get_precise_lemma : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_state -> key:(Lang.Term.term * Lang.Term.term Base.option) -> Lang.Term.term Base.option

val get_lemma : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_state -> key:Lang.Term.term -> Lang.Term.term Base.option

val smt_of_recurs_elim_eqns : ctx:Lang.Term.Context.t -> (Lang.Term.term * Lang.Term.term) Base.list -> p:Common.ProblemDefs.PsiDef.t -> S.smtTerm

val smt_of_aux_ensures : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> S.smtTerm Base.list

val smt_of_tinv_app : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> S.smtTerm

val smt_of_lemma_app : Common.ProblemDefs.term_info -> S.smtTerm

val smt_of_lemma_validity : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> S.command list

val inductive_solver_preamble : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Lang.SmtInterface.AsyncSmt.solver -> Common.ProblemDefs.term_info -> unit Lwt.t

val smt_of_disallow_witness_values : ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> S.smtTerm

val set_up_to_get_model : ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Lang.SmtInterface.AsyncSmt.solver -> Common.ProblemDefs.term_info -> unit Lwt.t

val mk_model_sat_asserts : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Common.ProblemDefs.term_info -> ( 'a -> Lang.Term.term ) -> ( Lang.Term.variable -> 'a Base.Option.t ) -> Smtlib.SmtLib.smtTerm Base.List.t

val verify_lemma_bounded : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Utils.Stats.verif_method * S.solver_response) Lwt.t * Base.int Lwt.u

val verify_lemma_unbounded : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> (Utils.Stats.verif_method * S.solver_response) Lwt.t * Base.int Lwt.u

val verify_lemma_candidate : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> Utils.Stats.verif_method * Lang.SmtInterface.SyncSmt.solver_response

Verify a lemma candidate with two parallel calls: one call to an induction solver, another call to a SMT solver. The induction solver attempts to prove that the solution is correct while the smt solver attempt to find a counterexample.

val parse_positive_example_solver_model : fctx:Lang.PMRS.Functions.ctx -> ctx:Lang.Term.Context.t -> Smtlib.SmtLib.solver_response -> Common.ProblemDefs.term_info -> Common.ProblemDefs.witness list

module Interactive : sig ... end

The Interactive module contains function used in interactive mode.

val synthesize_new_lemma : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> Lang.Term.term Base.option

val lemma_refinement_loop : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> Common.ProblemDefs.term_info -> Common.ProblemDefs.term_info Base.option

val witnesss_for_lemma_synt : Common.ProblemDefs.witness -> [> `Fst of Common.ProblemDefs.witness | `Snd of Common.ProblemDefs.witness | `Trd of Common.ProblemDefs.witness ]

Partitioning function to partitiion a list into (a,b,c) where a are examples that satisfy the invariant, b are examples that do not satisfy the invariant, c are examples that are spurious for other reasons.

val witnesss_for_ensures_synt : Common.ProblemDefs.witness -> [> `Fst of Common.ProblemDefs.witness | `Snd of Common.ProblemDefs.witness | `Trd of Common.ProblemDefs.witness ]

Partitioning function to partitiion a list into (a,b,c) where a are examples that are not in the reference function's image, b are examples that are in the reference function's image, c are examples that are spurious for other reasons.

val refine_ensures_predicates : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> neg_witnesss:Common.ProblemDefs.witness Base.list -> pos_witnesss:Common.ProblemDefs.witness Base.list -> Common.ProblemDefs.refinement_loop_state -> Common.ProblemDefs.term_state * [ `CoarseningFailure | `CoarseningOk | `Unrealizable ]

val synthesize_lemmas : ctx:Common.Env.env -> p:Common.ProblemDefs.PsiDef.t -> (Syguslib.Sygus.solver_response * ( 'a, Common.ProblemDefs.unrealizability_witness Base.List.t ) Base.Either.t) -> Common.ProblemDefs.refinement_loop_state -> ( ( Common.ProblemDefs.refinement_loop_state, Common.ProblemDefs.unrealizability_witness Base.list ) Base.Either.t, Syguslib.Sygus.solver_response ) Base.Result.t