Module Se2gis.Main

val solve_problem : ctx:Common.Env.env -> Common.ProblemDefs.PsiDef.t -> Syguslib.Sygus.solver_response Common.ProblemDefs.segis_response

val find_and_solve_problem : ctx:Common.Env.env -> (string * string * string) option -> ( string, Lang.PMRS.t, Base.String.comparator_witness ) Base.Map.t -> (Common.ProblemDefs.PsiDef.t * Syguslib.Sygus.solver_response Common.ProblemDefs.segis_response) list

find_and_solve_problem ctx (Some (target, reference, representation)) solves the synthesis problem associated with the target function named target, the reference function named reference and the representation function named representation that have been parsed in the file. If None is passed as argument, the default values are ("target", "spec", "repr"). If the functions cannot be found, it will exit the program. To allow the program to search for better recursion skeletons, use ManyProgramRefinement.find_and_solve_problem