Module Smtlib

module Logics : sig ... end

module SmtLib : sig ... end

This module implements an interface to the SMT-LIB standard. Some parts of documentation will refer to a specific section of the SMT-LIB standard available online (http://smtlib.cs.uiowa.edu/papers/smt-lib-reference-v2.6-r2021-05-12.pdf).

module Solvers : sig ... end

This module contains functions to interface with syntax-guided synthesis solvers using the SMTLIB language defined in The synchronous and asynchronous solvers defined here are functors parametric on Logger and Statistics modules to automate logging on some output and collecting statistics on solver usage.