Module Smtlib.Logics

type logic =

| QF_AX	(*
Closed quantifier-free formulas over the theory of arrays with extensionality.
*)

| QF_IDL	(*
Difference Logic over the integers. In essence, Boolean combinations of inequations of the form x - y < b where x and y are integer variables and b is an integer constant.
*)

| QF_UF	(*
Unquantified formulas built over a signature of uninterpreted (i.e., free) sort and function symbols.
*)

| QF_BV	(*
Closed quantifier-free formulas over the theory of fixed-size bitvectors.
*)

| QF_RDL	(*
Difference Logic over the reals. In essence, Boolean combinations of inequations of the form x - y < b where x and y are real variables and b is a rational constant.
*)

| QF_LIA	(*
Unquantified linear integer arithmetic. In essence, Boolean combinations of inequations between linear polynomials over integer variables.
*)

| QF_UFIDL	(*
Difference Logic over the integers (in essence) but with uninterpreted sort and function symbols.
*)

| QF_UFBV	(*
Unquantified formulas over bitvectors with uninterpreted sort function and symbols.
*)

| QF_ABV	(*
Closed quantifier-free formulas over the theory of bitvectors and bitvector arrays.
*)

| QF_LRA	(*
Unquantified linear real arithmetic. In essence, Boolean combinations of inequations between linear polynomials over real variables.
*)

| QF_NIA	(*
Quantifier-free integer arithmetic.
*)

| LIA	(*
Closed linear formulas over linear integer arithmetic.
*)

| QF_ALIA	(*
Quantifier-free arrays and linear arithmetic.
*)

| QF_UFLIA	(*
Unquantified linear integer arithmetic with uninterpreted sort and function symbols.
*)

| QF_UFLRA	(*
Unquantified linear real arithmetic with uninterpreted sort and function symbols.
*)

| QF_AUFBV	(*
Closed quantifier-free formulas over the theory of bitvectors and bitvector arrays extended with free sort and function symbols.
*)

| LRA	(*
Closed linear formulas in linear real arithmetic.
*)

| QF_NRA	(*
Quantifier-free real arithmetic.
*)

| NIA	(*
Quantifier-free integer arithmetic.
*)

| ALIA	(*
Arrays and linear arithmetic.
*)

| QF_AUFLIA	(*
Closed quantifier-free linear formulas over the theory of integer arrays extended with free sort and function symbols.
*)

| QF_UFNIA	(*
Unquantified integer arithmetic with uninterpreted sort and function symbols.
*)

| UFLRA	(*
Unquantified linear real arithmetic with uninterpreted sort and function symbols.
*)

| NRA	(*
Real arithmetic.
*)

| UFNIA	(*
Non-linear integer arithmetic with uninterpreted sort and function symbols.
*)

| AUFLIA	(*
Closed formulas over the theory of linear integer arithmetic and arrays extended with free sort and function symbols but restricted to arrays with integer indices and values.
*)

| AUFLIRA	(*
Closed linear formulas with free sort and function symbols over one- and two-dimentional arrays of integer index and real value.
*)

| AUFNIRA	(*
Closed formulas with free function and predicate symbols over a theory of arrays of arrays of integer index and real value.
*)

| UFDTLIA

| UFDTLRA

| UFDTNIA

| UFDTNRA

| DTNIA

| DTLIA

| ALL	(*
Default "ALL" logics
*)

Sub-logics for the main SMT-LIB logic (see http://smtlib.cs.uiowa.edu/logics.shtml)

val to_string : logic -> string

val of_string : string -> logic

type theory =

| ArraysEx

| Core

| FixedSizeBitVectors

| FloatingPoint

| Ints

| Reals

| Reals_Ints

| Strings

| All

val base_string : ?arith:bool -> theory -> string

val join_theories : theory -> theory -> theory

val combine : ?arith:bool -> ?datatypes:bool -> ?linear:bool -> ?qf:bool -> ?uf:bool -> theory list -> logic