Module Smtlib.Solvers

module OC = Stdio.Out_channel

module IC = Stdio.In_channel

val log_out : OC.t option Core.ref

module type Logger = sig ... end

Logger for the solvers. A solver might log error, debug and verbose messages.

module EmptyLog : Logger

An empty logging module that can be used to create silent solver instances.

module type Statistics = sig ... end

A module signature to log satistics about solver usage.

module NoStat : Statistics

module Synchronous (Log : Logger) (Stats : Statistics) : sig ... end

module Asyncs (Log : Logger) (Stats : Statistics) : sig ... end

Async versions of the solver operations for use in concurrent threads.