Module Solvers.EmptyLog

An empty logger that ignores messages, and can be used to construct a solver if no logging is desired.

val error : ( Stdlib.Format.formatter -> unit -> unit ) -> unit

val debug : ( Stdlib.Format.formatter -> unit -> unit ) -> unit

val verb : ( Stdlib.Format.formatter -> unit -> unit ) -> unit

val log_file : string

val verbose : bool

val log_queries : bool