Module Syguslib

module Solvers : sig ... end

This module contains functions to interface with syntax-guided synthesis solvers using the SyGuS Language Standard Version 2 or 1 defined in SyGuS. The synchronous and asynchronous solvers defined here are functors parametric on Logger and Statistics modules to automate logging on some output and collecting statistics on solver usage.

module Sygus : sig ... end

This module contains definitions that mirror the SyGuS Language Standard Version 2.1, which can be found at https://sygus.org/assets/pdf/SyGuS-IF_2.1.pdf.