Module Syguslib.Solvers

module OC = Stdio.Out_channel

module IC = Stdio.In_channel

module type Logger = sig ... end

Logger for the solvers. A solver might log error, debug and verbose messages.

module EmptyLog : Logger

An empty logging module that can be used to create silent solver instances.

module type Statistics = sig ... end

A module signature to log satistics about solver usage.

module NoStat : Statistics

module type SolverSystemConfig = sig ... end

type solver_instance = {

s_name : Base.string;

s_pid : Base.int;

s_input_file : Base.string;

s_output_file : Base.string;

s_process : Lwt_process.process_out;

}

val online_solvers : (Base.int * solver_instance) Base.list Base.ref

val mk_tmp_sl : string -> string

val commands_to_file : Sygus.program -> Base.string -> Base.unit

module SygusSolver (Stats : Statistics) (Log : Logger) (Config : SolverSystemConfig) : sig ... end