Module type Solvers.Logger

The Logger module must provide some basic logging functionality, including error, debug and verbose messages. One can also set log_queries to true with a file, which allows the solver to write queries to a separate file.

val error : (Stdlib.Format.formatter -> unit -> unit) -> unit

val debug : (Stdlib.Format.formatter -> unit -> unit) -> unit

val verb : (Stdlib.Format.formatter -> unit -> unit) -> unit

val log_file : string

val verbose : bool

val log_queries : bool