Module Syguslib__Sygus

val use_v1 : bool Base.ref

type symbol = Base.string

type literal =

| LitNum of Base.int

| LitDec of Base.float

| LitBool of Base.bool

| LitHex of Base.string

| LitBin of Base.bool Base.list

| LitString of Base.string

type index =

| INum of Base.int

| ISym of symbol

type identifier =

| IdSimple of symbol

| IdIndexed of symbol * index Base.list

| IdQual of symbol * sygus_sort

and sygus_sort =

| SId of identifier

| SApp of identifier * sygus_sort Base.list

type sygus_term =

| SyId of identifier

| SyLit of literal

| SyApp of identifier * sygus_term Base.list

| SyExists of sorted_var Base.list * sygus_term

| SyForall of sorted_var Base.list * sygus_term

| SyLet of binding Base.list * sygus_term

and sorted_var = symbol * sygus_sort

and binding = symbol * sygus_term

type feature =

| FGrammar

| FFwdDecls

| FRecursion

type command =

| CCheckSynth

| CConstraint of sygus_term

| CDeclareVar of symbol * sygus_sort

| CInvConstraint of symbol * symbol * symbol * symbol

| CSetFeature of feature * Base.bool

| CSynthFun of symbol * sorted_var Base.list * sygus_sort * grammar_def Base.option

| CSynthInv of symbol * sorted_var Base.list * grammar_def Base.option

| CDeclareDataType of symbol * dt_cons_dec Base.list

| CDeclareDataTypes of sygus_sort_decl Base.list * dt_cons_dec Base.list Base.list

| CDeclareSort of symbol * Base.int

| CDefineFun of symbol * sorted_var Base.list * sygus_sort * sygus_term

| CDefineSort of symbol * sygus_sort

| CSetInfo of symbol * literal

| CSetLogic of symbol

| CSetOption of symbol * literal

and sygus_sort_decl = symbol * Base.int

and dt_cons_dec = symbol * sorted_var Base.list

and grammar_def = (sorted_var * grouped_rule_list) Base.list

and grouped_rule_list = sygus_gsterm Base.list

and sygus_gsterm =

| GConstant of sygus_sort

| GTerm of sygus_term

| GVar of sygus_sort

type program = command Base.list

val special_chars : Base.char Base.list

val reserved_words : Base.string Base.list

val digits : char list

val valid_ident : Base.string -> Base.bool

`valid_ident name` is `true` whenever `name` is not a reserved word, and does not contain any special character.

val symbol_of_sexp : Sexplib.Sexp.t -> symbol

val sexp_of_symbol : symbol -> Sexplib.Sexp.t

val index_of_sexp : Sexplib.Sexp.t -> index

val sexp_of_index : index -> Sexplib.Sexp.t

val identifier_of_sexp : Sexplib.Sexp.t -> identifier

val sexp_of_identifier : identifier -> Sexplib.Sexp.t

val sygus_sort_of_sexp : Sexplib.Sexp.t -> sygus_sort

val sexp_of_sygus_sort : sygus_sort -> Sexplib.Sexp.t

val char_to_bool : Base.char -> bool

val literal_of_string : Base.string -> literal

val literal_of_sexp : Sexplib.Sexp.t -> literal

val bool_list_to_bin_string : Base.bool Base.list -> Base.String.t

val sexp_of_literal : literal -> Sexplib.Sexp.t

val sorted_var_of_sexp : Sexplib.Sexp.t -> sorted_var

val sexp_of_sorted_var : sorted_var -> Sexplib.Sexp.t

val sygus_term_of_sexp : Sexplib.Sexp.t -> sygus_term

val binding_of_sexp : Sexplib.Sexp.t -> binding

val sexp_of_sygus_term : sygus_term -> Sexplib.Sexp.t

val feature_of_sexp : Sexplib.Sexp.t -> feature

val sexp_of_feature : feature -> Sexplib.Sexp.t

val sygus_sort_decl_of_sexp : Sexplib.Sexp.t -> sygus_sort_decl

val sexp_of_sort_decl : sygus_sort_decl -> Sexplib.Sexp.t

val sygus_sort_decl_list_of_sexp : Sexplib.Sexp.t -> sygus_sort_decl Base.list

val dt_cons_dec_of_sexp : Sexplib.Sexp.t -> dt_cons_dec

val sexp_of_dt_cons_dec : dt_cons_dec -> Sexplib.Sexp.t

val dt_cons_dec_list_of_sexp : Sexplib.Sexp.t -> dt_cons_dec Base.list

val sygus_gsterm_of_sexp : Sexplib.Sexp.t -> sygus_gsterm

val sexp_of_sygus_gsterm : sygus_gsterm -> Sexplib.Sexp.t

val pre_grouped_rule_of_sexp : Sexplib.Sexp.t -> symbol * sygus_sort * sygus_gsterm Base.List.t

val grammar_def_of_sexps : Sexplib.Sexp.t -> Sexplib.Sexp.t -> grammar_def

val sexp_of_grammar_def : grammar_def -> Sexplib.Sexp.t * Sexplib.Sexp.t

val command_of_sexp : Sexplib.Sexp.t -> command

val sexp_of_command : command -> Sexplib.Sexp.t

val program_of_sexp_list : Sexplib.Sexp.t Base.list -> program

val sexp_list_of_program : program -> Sexplib.Sexp.t Base.list

type solver_response =

| RSuccess of (symbol * sorted_var Base.list * sygus_sort * sygus_term) Base.list

| RInfeasible

| RFail

| RUnknown

val is_sat : solver_response -> bool

val is_failed : solver_response -> bool

val is_infeasible : solver_response -> bool

val reponse_of_sexps : Sexplib.Sexp.t Base.list -> solver_response

module SyCommand : sig ... end

module SyTerm : sig ... end

module SyIdentifier : sig ... end

module SyLiteral : sig ... end

module SySort : sig ... end

val write_command : Stdio.Out_channel.t -> command -> Base.unit

val is_setter_command : SyCommand.t -> bool

val is_well_formed : program -> Base.bool

A SyGuS input is not well-formed if it specifies a list of commands that do not meet the restrictions given in this section regarding their order. The order is specified by the following regular pattern: ```(set logic command)? (setter commands)∗(other sygus_commands)∗```

val max_definition : command

val min_definition : command

val rename : (symbol * symbol) Base.list -> sygus_term -> sygus_term