Module Utils.Config

val problem_name : string Base.ref

val debug : bool Base.ref

Toggle debugging.

val debug_msg_max_chars : int Base.ref

Limit the size of debugging messages.

val info : bool Base.ref

Toggle info and error messages. Set to true by default.

val timings : bool Base.ref

Toggle printing timing info when info is off. Set to true by default.

val reduction_limit : int Base.ref

Maximum steps of rewrites to apply during PMRS reduction (symbolic evaluation).

val show_vars : bool Base.ref

val verbose : bool Base.ref

Toggle verbose messages.

val math_display : bool Base.ref

val output_folder : Base.string Base.option Base.ref

Optional output folder for solution. If None, the solution is only printed on the standard output. If Some path, the solution is written in path/inputfilename

val set_output_folder : Base.string -> Base.unit

val get_output_file : string -> string Base.Option.t

val interactive_lemmas : bool Base.ref

Prompt user to input a precondition (lemma) for each equation, while the equations are being generated from a set of terms in Equations.make.

val interactive_lemmas_loop : bool Base.ref

val interactive_lifting : bool Base.ref

Prompt for lifting expressions.

val classify_ctex : bool Base.ref

val interactive_check_lemma : bool Base.ref

val check_unrealizable : bool Base.ref

Check whether a system of equations defines a "functionally realizable" synthesis problem. ON by default.

val check_unrealizable_smt_unsatisfiable : bool Base.ref

Check whether a synthesis problem is unrealizable by checking whether the problem with second-order quantification is unrealizable, using a SMT solver (Z3 by default).

val no_bounded_sat_as_unsat : bool Base.ref

Accept no-sat in bounded checking as unsat.

val bounded_lemma_check : bool Base.ref

val sysfe_opt : bool Base.ref

Run with the optimizations used to synthesize solutions for systems of equations. Works best when there are more than 4 cores.

val force_nonlinear : bool Base.ref

Sometimes non-linearity occurs even when the reference only has linear operators.

val tmp_folder : string

val root_folder : string

val base : string -> string

val use_cvc4 : bool Base.ref

val cvc4_binary_path : FilePath.filename option

val cvc5_binary_path : FilePath.filename option

val using_cvc5 : unit -> Base.bool

val cvc_binary_path : unit -> FilePath.filename

val z3_binary_path : FilePath.filename

val yices_binary_path : FilePath.filename option

val dryadsynth_binary_path : FilePath.filename

val use_eusolver : bool Base.ref

val eusolver_binary_path : FilePath.filename

val verification_solver : string Base.ref

val set_verification_solver : Base.string -> Base.unit

val smt_solver_log_file : string Base.ref

val smt_log_queries : bool Base.ref

val smt_solve_verbose : bool Base.ref

val generate_benchmarks : bool Base.ref

val benchmark_generation_dir : string Base.ref

val benchmark_lang_version : string Base.ref

val set_benchmark_generation_dir : Base.string -> Base.unit

val new_benchmark_file : ?hint:Base.String.t -> string -> string

val generate_proof : bool Base.ref

val proof_generation_file : string Base.ref

val set_proof_output_file : Base.string -> Base.unit

val detupling_on : bool Base.ref

Unkmowns of type i -> a * b * c .. are split into i -> a, i -> b and equations between tuples are separated into different equations. Turn off using option -t or --detupling-off

val use_syntactic_definitions : bool Base.ref

Syntactic definitions: detect syntactic definitions in the equation systems and use them. Turn off with --no-syndef

val split_solve_on : bool Base.ref

Separate systems of equations into subsystems, and solve independently each subsystem. Use option -s or --split-solving-off to turn off.

val simple_init : bool Base.ref

Initialize T using a single variable of type theta, instead of searching for a set of terms covering all unknowns. Use option --simple-init or -c to turn on.

val use_acegis : bool Base.ref

Not an optimization. Set to true to use the base abstract cegis algorithm. Option --use-acegis sets this flag to true.

val use_ccegis : bool Base.ref

Not an optimization. Set to true to use the base concrete cegis algorithm. Option --use-ccegis sets this flag to true.

val simplify_eqns : bool Base.ref

Simplify equations before feeding them to equations solver.

val optimize_grammars : int Base.ref

Use the equations as a indicator to optimize grammars, without compromising soundness.

val set_grammar_optimization_level : Base.string -> Base.unit

val pp_eqn_count : int Base.ref

When printing a system of equations, put a limit on how many equations are printed.

val expand_depth : int Base.ref

Maximum depth of pointwise expansions to perform. Careful setting this variable to high, it will lead to explosion in the number of terms expanded. A lazy expansion should be implemented.

val num_expansions_check : int Base.ref

val expand_cut : int Base.ref

val set_num_expansions_check : Base.string -> Base.unit

num_expansions_check is set by the -n or --verification option of the CLI.

val use_bmc : bool Base.ref

Use bounded model checking. From CLI, use --use-bmc to set to true.

val check_depth : int Base.ref

Depth of bounded model checking.

val set_check_depth : Base.string -> Base.unit

Bounded model checking depth is set by the -b or --bmc option of the CLI.

val induction_proof_tlimit : int Base.ref

A time limit for induction proofs. Infinity if set to negative.

val set_induction_proof_tlimit : Base.string -> Base.unit

val wait_parallel_tlimit : float Base.ref

A time limit parallel calls when waiting on first result.. Infinity if set to negative. Default is 10 mins.

val set_wait_parallel_tlimit : Base.string -> Base.unit

val rewrite_limit : int Base.ref

A limit for the number of rewriting steps applied during deduction.

val set_rewrite_limit : Base.string -> Base.unit

val fuzzing_count : int Base.ref

When a model has been found, attempt fuzzing to find models that satisfy the same constraints. Used in Counterexamples.ml.

val set_fuzzing_count : Base.string -> Base.unit

val attempt_lifting : bool Base.ref

Attempt to lift the function if there is no solution.

val max_lifting_attempts : int Base.ref

The number of times Synduce should attempt to add a lifting variable.

val set_max_lifting_attempts : Base.string -> Base.unit

val options : (unit -> unit) -> bool Stdlib.ref -> (char * string * (unit -> unit) option * (Base.string -> Base.unit) option) list