Module Utils.Log

val reference_text : string Base.ref

val extract : Base.String.t -> (Stdlib.Lexing.position * Stdlib.Lexing.position) -> Base.string

val compress : string -> string

val shorten : Base__Int.t -> Base.String.t -> Base.String.t

val log_located : Base.Formatter.t -> (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ( Stdlib.Format.formatter -> 'a -> unit ) -> 'a -> unit

val (@!) : Base.Sexp.t -> (Stdlib.Lexing.position * Stdlib.Lexing.position) -> Base.Sexp.t

val width : int

val range : Base.String.t -> (Stdlib.Lexing.position * Stdlib.Lexing.position) -> Base.string

val log_with_excerpt : Base.Formatter.t -> Base.string -> (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ( Stdlib.Format.formatter -> 'a -> unit ) -> 'a -> unit

val wrap : Base.string -> Stdlib.Format.formatter -> unit -> unit

val wrap1 : ( 'a -> 'b -> 'c, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a -> 'b -> Stdlib.Format.formatter -> unit -> 'c

val wrap2 : ( 'a -> 'b -> 'c -> 'd -> 'e, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a -> 'b -> 'c -> 'd -> Stdlib.Format.formatter -> unit -> 'e

val wrapn : Stdlib.Format.formatter -> unit -> ( 'a, Stdlib.Format.formatter, unit ) Stdlib.format -> 'a

val indent : int Base.ref

val section_indent : unit -> Base.String.t

val error : ( Base.Formatter.t -> Base.unit -> Base.unit ) -> Base.unit

val error_msg : Base.string -> Base.unit

val fatal : unit -> 'a

val loc_fatal_errmsg : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> string -> 'a

val info : ( Base.Formatter.t -> Base.unit -> Base.unit ) -> Base.unit

Info messaging is the lowest verbose output, activated by default.

val hline : Base.String.t

val sep : ?i:Base.Int.t option -> unit -> unit

val debug : ( Base.Formatter.t -> Base.unit -> Base.unit ) -> Base.unit

Debug messaging is used mostly for printing status of interactions with external solvers.

val debug_msg : Base.string -> Base.unit

val print_ok : unit -> unit

val verb : ( Stdlib.Format.formatter -> 'a -> unit ) -> 'a -> unit

Verbose output for printing status of internal methods.

val verbose : ( Stdlib.Format.formatter -> unit -> unit ) -> unit

val verbose_msg : string -> unit

val start_section : string -> unit

val end_section : unit -> unit

val status_printer : unit -> unit

val to_file : Base.string -> ( Base.Formatter.t -> Base.unit -> Base.unit ) -> Base.unit

Print a message function to a file.

val print_solvers_summary : Base.int -> Base.Formatter.t -> unit -> Base.unit