Module Utils.LogJson

val solvers_summary : unit -> Yojson.t

Outputs a JSON summarizing the time spent in each solver as well as the number of solver instances used during the execution of the algorithm.

val refinement_steps_summary : unit -> Yojson.t

Outputs a summary of the refinement steps taken by the algorithm in the form of a JSON.