Module Utils.Stats

val timed : (Base.unit -> 'a) -> Base.float * 'a

type proc_time_info = {

pti_start : Base.float;

pti_status : Base.int;

pti_elapsed : Base.float;

pti_last : Base.float;

pti_exited : Base.bool;

}

val _TIME_TABLE_ : (Base.int, proc_time_info) Base.Hashtbl.t

val _PID_TABLE_ : (Base.int, Base.string) Base.Hashtbl.t

val solver_kind_pids : (Base.string, Base.int Base.list) Base.Hashtbl.t

val get_solver_pids : unit -> (Base.string Base.Hashtbl.key * Base.int Base.list) list

val _PID_ : int Base.ref

val log_proc_start : Base.int -> unit

val log_proc_restart : Base.int -> unit

Log when a process is restarted.

val log_alive : Base.int -> unit

Update the elapsed time. When pausing a subprocess, first update the elapsed time and update the restart time when restarting the process.

val get_alive : unit -> (Base.string * Base.int Base.Hashtbl.key) Base.List.t

Get the list of all registered subprocesses still running.

val log_proc_quit : ?status:Base.int -> Base.int -> unit

Log when a process quits without error (status = 0 by default).

val get_elapsed : Base.int -> Base.float

val glob_start : unit -> unit

val get_glob_elapsed : unit -> Base__Float.t

val verif_time : float Base.ref

Total elapsed time spent in verification of solutions.

val add_verif_time : Base.float -> Base.unit

val log_solver_start : Base.int -> Base.string -> unit

val refinement_log : (Base.string * Yojson.t) Base.list Base.Stack.t

val major_step_counter : int Base.ref

val log_new_major_step : tsize:int -> usize:int -> unit -> unit

val log_major_step_end : ?failure_step:bool -> synth_time:Base.float -> verif_time:Base.float -> t:Base.int -> u:Base.int -> Base.bool -> Base.unit

type verif_method =

| BoundedChecking

| Induction

val verif_method_to_str : verif_method -> string

val counterexample_classification_method : verif_method Base.option Base.ref

val update_counterexample_classification_method : verif_method -> Base.unit

val last_lemma_synthesized : (Base.string * Base.string) Base.option Base.ref

val last_lemma_proved : verif_method Base.option Base.ref

If a lemma has been proved, the Some true means it has been proved by induction and Some false means it has been proved by bounded checking.

val set_lemma_synthesized : Base.string -> Base.string -> Base.unit

val set_last_lemma_proof_method : verif_method -> Base.unit

val log_minor_step : synth_time:Base.float -> auxtime:Base.float -> Base.bool -> Base.unit