Module Utils__.Stats

type proc_time_info = {

pti_start : Base.float;

pti_status : Base.int;

pti_elapsed : Base.float;

pti_last : Base.float;

}

val _TIME_TABLE_ : (Base.int, proc_time_info) Base.Hashtbl.t

val solver_kind_pids : (Base.string, Base.int Base.list) Base.Hashtbl.t

val _PID_ : int Base.ref

val pid_placeholder : unit -> Base.int

val log_proc_start : Base.int -> Base.unit

val log_proc_restart : Base.int -> Base.unit

Log when a process is restarted.

val log_alive : Base.int -> Base.unit

Update the elapsed time. When pausing a subprocess, first update the elapsed time and update the restart time when restarting the process.

val log_proc_quit : ?⁠status:Base.int -> Base.int -> Base.unit

Log when a process quits without error (status = 0 by default).

val get_elapsed : Base.int -> Base.float

val glob_start : unit -> unit

val get_glob_elapsed : unit -> Base__Float.t

val verif_time : float Base.ref

Total elapsed time spent in verification of solutions.

val add_verif_time : Base.float -> Base.unit

val log_solver_start : Base.int -> Base.string -> unit

val print_solvers_summary : Base.Formatter.t -> unit -> Base.unit