Module Tenyaeva_lib.Ast

Copyright 2025, Tenyaeva Ekaterina

SPDX-License-Identifier: LGPL-3.0-or-later

type ident = Base.string

val pp_ident : Ppx_deriving_runtime.Format.formatter -> ident -> Ppx_deriving_runtime.unit

val show_ident : ident -> Ppx_deriving_runtime.string

val is_keyword : string -> bool

type constant =

| Const_int of Base.int
(*
integer, e.g. 122
*)

| Const_bool of Base.bool
(*
boolean, e.g. true
*)

| Const_unit
(*
()
*)

val pp_constant : Ppx_deriving_runtime.Format.formatter -> constant -> Ppx_deriving_runtime.unit

val show_constant : constant -> Ppx_deriving_runtime.string

type rec_flag =

| Recursive
(*
recursive
*)

| NonRecursive
(*
non-recursive
*)

val pp_rec_flag : Ppx_deriving_runtime.Format.formatter -> rec_flag -> Ppx_deriving_runtime.unit

val show_rec_flag : rec_flag -> Ppx_deriving_runtime.string

type binary_op =

| Add
(*
+
*)

| Mult
(*
*
*)

| Sub
(*
-
*)

| Div
(*
/
*)

| Gt
(*
>
*)

| Lt
(*
<
*)

| Eq
(*
=
*)

| Neq
(*
<>
*)

| Gte
(*
>=
*)

| Lte
(*
<=
*)

val pp_binary_op : Ppx_deriving_runtime.Format.formatter -> binary_op -> Ppx_deriving_runtime.unit

val show_binary_op : binary_op -> Ppx_deriving_runtime.string

type unary_op =

| Negative
(*
unary minus, e.g. -5
*)

| Positive
(*
unary plus, e.g. +5
*)

| Not
(*
not
*)

val pp_unary_op : Ppx_deriving_runtime.Format.formatter -> unary_op -> Ppx_deriving_runtime.unit

val show_unary_op : unary_op -> Ppx_deriving_runtime.string

type type_annot =

| Type_int
(*
integer type - int
*)

| Type_bool
(*
boolean type - bool
*)

| Type_unit
(*
unit type - unit
*)

| Type_var of ident
(*
variable type
*)

| Type_arrow of type_annot * type_annot
(*
arrow type
*)

| Type_option of type_annot
(*
type option
*)

val pp_type_annot : Ppx_deriving_runtime.Format.formatter -> type_annot -> Ppx_deriving_runtime.unit

val show_type_annot : type_annot -> Ppx_deriving_runtime.string

type pattern =

| Pat_any
(*
matches any value without binding it - _
*)

| Pat_var of ident
(*
matches any value and binds it to a variable, e.g. x
*)

| Pat_constant of constant
(*
matches a constant value, e.g. 42, true
*)

| Pat_option of pattern Base.option
(*
matches an optional pattern, e.g. Some x or None
*)

| Pat_constraint of type_annot * pattern
(*
typed pattern, e.g. a: int
*)

val pp_pattern : Ppx_deriving_runtime.Format.formatter -> pattern -> Ppx_deriving_runtime.unit

val show_pattern : pattern -> Ppx_deriving_runtime.string

type expression =

| Expr_const of constant
(*
constant, e.g. 10
*)

| Expr_ident of ident
(*
variable, e.g. x
*)

| Expr_option of expression Base.option
(*
optonal expression, e.g. Some x
*)

| Expr_constraint of type_annot * expression
(*
typed expression, e.g. a: int
*)

| Expr_binop of binary_op * expression * expression
(*
binary operation, e.g. 1 + 5
*)

| Expr_unop of unary_op * expression
(*
unary operation, e.g. -7
*)

| Expr_fun of pattern * expression
(*
function, e.g. fun (x, y) -> x + y
*)

| Expr_apply of expression * expression
(*
application, e.g. (fun (x, y) -> x + y) (1, 2)
*)

| Expr_if of expression * expression * expression Base.option
(*
conditional expression, e.g. if a then b else c
*)

| Expr_let of rec_flag * value_binding * value_binding Base.list * expression
(*
let, e.g. let x = 5
*)

| Expr_function of case * case Base.list
(*
function, e.g. fun (x, y) -> x + y
*)

| Expr_match of expression * case * case Base.list
(*
pattern matching, e.g. match x with | 0 -> "zero" | _ -> "nonzero"
*)

and value_binding = {

vb_pat : pattern;
(*
the pattern being bound, e.g. x, (a, b)
*)

vb_expr : expression;
(*
the expression being assigned, e.g. 42, fun x -> x + 1
*)

}

and case = {

case_pat : pattern;
(*
the pattern to match, e.g. x, _
*)

case_expr : expression;
(*
the expression to evaluate if the pattern matches
*)

}

val pp_expression : Ppx_deriving_runtime.Format.formatter -> expression -> Ppx_deriving_runtime.unit

val show_expression : expression -> Ppx_deriving_runtime.string

val pp_value_binding : Ppx_deriving_runtime.Format.formatter -> value_binding -> Ppx_deriving_runtime.unit

val show_value_binding : value_binding -> Ppx_deriving_runtime.string

val pp_case : Ppx_deriving_runtime.Format.formatter -> case -> Ppx_deriving_runtime.unit

val show_case : case -> Ppx_deriving_runtime.string

type structure_item =

| Str_eval of expression
(*
an expression to be evaluated but not bound, e.g. 1 + 2
*)

| Str_value of rec_flag * value_binding * value_binding Base.list
(*
a value or function binding, e.g. let x = 1
*)

val pp_structure_item : Ppx_deriving_runtime.Format.formatter -> structure_item -> Ppx_deriving_runtime.unit

val show_structure_item : structure_item -> Ppx_deriving_runtime.string

type structure = structure_item Base.list

full program

val pp_structure : Ppx_deriving_runtime.Format.formatter -> structure -> Ppx_deriving_runtime.unit

val show_structure : structure -> Ppx_deriving_runtime.string