Module ZRusML_lib.Parser

Copyright 2023-2024, Rustam Shangareev and Danil Yevdokimov

SPDX-License-Identifier: LGPL-2.1

val parse : 'a Angstrom.t -> Base.string -> ('a, Base.string) Base.result

parse p s applies the parser p on string s.

val prog : Ast.prog Angstrom.t

prog is the main parser for the entire program.