Package b0 0.0.1

WARNING. This package is unstable and work in progress, do not depend on it.

B0 describes software construction and deployments using modular and customizable definitions written in OCaml.

It also provides the B00 build library which provides arbitrary build abstraction with reliable, efficient incremental rebuilds.

Manuals

For B0 and B00 users.

The B0 manual is a conceptual overview of the B0 system and B0 files. It is recommended reading.

The B00 manual describes the B00 build library.

For B0 programmers.

The B0 driver development manual describes how to make your own drivers to operate on B0 files.

The TODO page

Old manual fragments and TODO which will eventually disapear.

B00

B00 is the build model used by B0. It can be used for its own good to devise domain specific build system – see for example the brzo or odig tools.

The b0.b00.std and b0.b00.kit libraries also provide sharp tools for quick "scripts" and command line tools even if you are not using b0.b00.

Library b0.b00.std

The b0.b00.std library has a few things that should be in the stdlib and a few others that should not be but are useful for B0 and B00 based programs.

B00_std

Library b0.b00

The build library.

B000

B000_conv

B00

Library b0.b00.kit

Toolkit for working with B00 and B0. A few things do not depend on the b0.b00 library but are useful to have under your hand.

IO

B00_base64

B00_fexts

B00_findex

B00_htmlg

B00_http

B00_lines

B00_serialk_json

B00_serialk_sexp

B00_serialk_text

Dev tools and info

B00_cmark

B00_github

B00_ocaml

B00_odoc

B00_os

B00_vcs

UI fragments

Mostly Cmdliner user interface fragments.

B00_ui

B00_std_ui

B00_trace

Viewers and editors

B00_editor

B00_pager

B00_pdf_viewer

B00_www_browser

B0

B0 is the system for describing software construction and deployments via B0 files.

Library b0

B0_def

B0_meta

B0_build

B0_unit

B0_pack

Library b0.kit

Toolkit for B0 files.

B0_kit

B0_srcs

B0_ocaml

Library b0.driver

The driver library provides programmatic access to the definitions of B0 files. Only needed if you want to process B0 files your own own way.

B0_file

B0_driver

Package info

authors

The b0 programmers

changes-files

CHANGES.md

depends

cmdliner

ocaml

ocamlbuild

ocamlfind

topkg

homepage

https://erratique.ch/software/b0

issues

https://github.com/b0-system/b0/issues

license

BSD2

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/b0/doc

readme-files

README.md

repo

git+https://erratique.ch/repos/b0.git#b00

tags

build

dev

org:b0-system

org:erratique

version

0.0.1