Package down 0.0.3

Down is an unintrusive user experience upgrade for the ocaml toplevel (REPL).

Simply load the zero dependency Down library in the ocaml toplevel and you get line edition, history, session support and identifier completion and documentation (courtesy of ocp-index).

Consult the manual.

Quick start

Simply run ocaml and load the library:

> ocaml
...
# #use "down.top";;
Down loaded. Type Down.help () for more info.
# List.con^t
  List.cons : 'a -> 'a list -> 'a list
  List.concat : 'a list list -> 'a list
# List.con

You can add this #use invocation to your ~/.ocamlinit file.

For ocamlnat you will likely have to invoke it with -noinit and you must #use "down.nattop".

More in the manual.

API

Down

Package info

authors

The down programmers

changes-files

CHANGES.md

depends

ocaml

ocamlbuild

ocamlfind

topkg

homepage

https://erratique.ch/software/down

issues

https://github.com/dbuenzli/down/issues

license

ISC

license-files

LICENSE.md

maintainers

Daniel Bünzli <daniel.buenzl i@erratique.ch>

online-doc

https://erratique.ch/software/down/doc

readme-files

README.md

repo

git+https://erratique.ch/repos/down.git

tags

dev

org:erratique

repl

toplevel

version

0.0.3