Module Interpreter.Eval

Main interpreter module

Parameters

module M : MonadFail

Signature

val mapm : ('a -> 'b M.t) -> 'a list -> 'b list M.t

mapm f l applies f to each element of l returning the resulting list if all applications succeed and failing otherwise

val flat_mapm : ('a -> 'b list M.t) -> 'a list -> 'b list M.t

flat_mapm f l applies f to each element of l returning the resulting flattened list if all applications succeed and failing otherwise

type var_t =

| IndArray of string IMap.t	(*
Simple variable or indexed array
*)

| AssocArray of string SMap.t	(*
Associative array
*)

Container for values of variables

val pp_var_t : Ppx_deriving_runtime.Format.formatter -> var_t -> Ppx_deriving_runtime.unit

val show_var_t : var_t -> Ppx_deriving_runtime.string

type fun_t = Ast.compound

Container for functions

val pp_fun_t : Ppx_deriving_runtime.Format.formatter -> fun_t -> Ppx_deriving_runtime.unit

val show_fun_t : fun_t -> Ppx_deriving_runtime.string

type environment = {

vars : var_t SMap.t;	(*
Variables available in the current scope
*)

funs : fun_t SMap.t;	(*
Functions available in the current scope
*)

stdin : string;	(*
Current stdin contents
*)

stdout : string;	(*
Current stdout contents
*)

stderr : string;	(*
Current stderr contents
*)

retcode : int;	(*
Return code of the last operation
*)

}

Complete environment

val pp_environment : Ppx_deriving_runtime.Format.formatter -> environment -> Ppx_deriving_runtime.unit

val show_environment : environment -> Ppx_deriving_runtime.string

val get_var : SMap.key -> environment -> var_t option

val set_var : SMap.key -> var_t -> environment -> var_t SMap.t

val get_fun : SMap.key -> environment -> fun_t option

val set_fun : SMap.key -> fun_t -> environment -> fun_t SMap.t

val ev_var : environment -> Ast.var -> string M.t

Evaluate variable

val ev_arithm : environment -> Ast.arithm -> int M.t

Evaluate arithmetic expression

val ev_param_exp : environment -> Ast.param_exp -> string M.t

Evaluate parameter expansion

val ev_filename_exp : 'a -> string -> Stdlib.String.t list M.t

Evaluate filename expansion

val ev_word : environment -> Ast.word -> (environment * Stdlib.String.t list) M.t

Evaluate word

val ev_words : environment -> ?c:(Stdlib.String.t list -> bool) -> ?e:string -> Ast.word list -> (environment * Stdlib.String.t list) M.t

Evaluate word list

val ev_assignt : environment -> Ast.assignt -> environment M.t

Evaluate assignment

val ev_cmd : environment -> Ast.cmd -> environment M.t

Evaluate command

val ev_redir : environment -> 'a M.t

Evaluate redirection

val ev_pipeline_list : environment -> environment M.t

Evaluate pipeline list

val ev_pipeline : environment -> environment M.t

Evaluate pipeline

val ev_compound : environment -> environment M.t

Evaluate compound

val ev_while_loop : environment -> environment M.t

Evaluate while loop

val ev_for_list_loop : environment -> environment M.t

Evaluate for loop (list form)

val ev_for_expr_loop : environment -> environment M.t

Evaluate for loop (expression form)

val ev_if_stmt : environment -> environment M.t

Evaluate if statement

val ev_case_stmt : environment -> environment M.t

Evaluate case statement

val ev_script_elem : environment -> environment M.t

Evaluate script element

val ev_script : environment -> Ast.script -> environment M.t

Evaluate Bash script