Module Counter.Efficiency

type t = {

consecutive : int array;

mutable consecutive_blocked : int;

mutable no_more_binary : int;

mutable no_more_unary : int;

mutable clashing_instance : int;

mutable time_correction : int;

}

val write_t : Stdlib.Buffer.t -> t -> unit

Output a JSON value of type t.

val string_of_t : ?len:int -> t -> string

Serialize a value of type t into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_t : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> t

Input JSON data of type t.

val t_of_string : string -> t

Deserialize JSON data of type t.