Module Ast.Ast

type ident = string

val pp_ident : Ppx_deriving_runtime.Format.formatter -> ident -> Ppx_deriving_runtime.unit

val show_ident : ident -> Ppx_deriving_runtime.string

type block = statement list

and unop =

| Not	(*
not
*)

| USub	(*

*)

and logic_binop =

| And	(*
and
*)

| Or	(*
or
*)

and arithm_binop =

| Add	(*

*)

| Mul	(*
*
*)

| Sub	(*

*)

| Div	(*
/
*)

| Pow	(*
^
*)

and compare_binop =

| Le	(*
<=
*)

| Ge	(*
>=
*)

| Lt	(*
>
*)

| Gt	(*
<
*)

| Eq	(*
==
*)

| Ne	(*
~=
*)

and string_binop =

| Concat

and binop =

| LOp of logic_binop	(*
Logic operators: and, or
*)

| AOp of arithm_binop	(*
Arithmetic operators: +. -. /, *, ^
*)

| COp of compare_binop	(*
Comparative operators: <=, >=, <, >, ==, ~=
*)

| SOp of string_binop	(*
String operators: ..
*)

and expr_table_entry =

| JustExpr of expr	(*
Just an expressinon, its index will be determined by position
*)

| PairExpr of expr * expr	(*
Key/Value Pair (key=value)
*)

and lua_function = ident list * block

and const =

| LuaBool of bool	(*
true | false
*)

| LuaNumber of float	(*
1, 2.3, 3, ...
*)

| LuaString of string	(*
"abc"
*)

| LuaFunction of lua_function	(*
function(args) block end
*)

| LuaNil	(*
nil
*)

and expr =

| LuaConst of const	(*
number or string or ..., written in const
*)

| LuaVariable of ident	(*
obv. variable
*)

| LuaTableGet of expr * expr	(*
exprexpr
*)

| LuaTableInit of expr_table_entry list	(*
xpr, expr, expr, ...
*)

| LuaBinOp of binop * expr * expr	(*
expr binop expr
*)

| LuaUnOp of unop * expr	(*
unop expr
*)

| LuaExprApply of apply	(*
expr(expr, ..., expr)
*)

and apply =

| LuaCall of expr * expr list	(*
expr(expr, ..., expr)
*)

and statement =

| LuaDo of block	(*
do sttmn sttmnt end
*)

| LuaSet of lvalue list * expr list	(*
a, b, c = d, e, f
*)

| LuaWhile of expr * block	(*
while expr do block
*)

| LuaRepeat of block * expr	(*
repeat block until expr
*)

| LuaIf of expr * block * elseif_block list * block option	(*
if expr then block elseif block else block end
*)

| LuaFornum of ident * expr * expr * expr option * block	(*
for var = expr, expr, ?expr, block
*)

| LuaForin of ident list * expr list * block	(*
for expr, expr, ... in expr, expr, ... do block
*)

| LuaLocal of ident list * expr list	(*
local a = ?expr
*)

| LuaReturn of expr option	(*
return expr
*)

| LuaBreak	(*
break
*)

| LuaStatementApply of apply	(*
just a fun call, needed because of side effects
*)

| LuaFunctionDeclare of lvalue * ident list * block	(*
function name(args) body end
*)

| LuaExpr of expr	(*
just an expression, will be printed, for REPL
*)

and lvalue =

| Index of lvalue * expr	(*
a123..nnn
*)

| Ident of ident	(*
var
*)

and elseif_block = expr * block

elseif expr then block

val pp_block : Ppx_deriving_runtime.Format.formatter -> block -> Ppx_deriving_runtime.unit

val show_block : block -> Ppx_deriving_runtime.string

val pp_unop : Ppx_deriving_runtime.Format.formatter -> unop -> Ppx_deriving_runtime.unit

val show_unop : unop -> Ppx_deriving_runtime.string

val pp_logic_binop : Ppx_deriving_runtime.Format.formatter -> logic_binop -> Ppx_deriving_runtime.unit

val show_logic_binop : logic_binop -> Ppx_deriving_runtime.string

val pp_arithm_binop : Ppx_deriving_runtime.Format.formatter -> arithm_binop -> Ppx_deriving_runtime.unit

val show_arithm_binop : arithm_binop -> Ppx_deriving_runtime.string

val pp_compare_binop : Ppx_deriving_runtime.Format.formatter -> compare_binop -> Ppx_deriving_runtime.unit

val show_compare_binop : compare_binop -> Ppx_deriving_runtime.string

val pp_string_binop : Ppx_deriving_runtime.Format.formatter -> string_binop -> Ppx_deriving_runtime.unit

val show_string_binop : string_binop -> Ppx_deriving_runtime.string

val pp_binop : Ppx_deriving_runtime.Format.formatter -> binop -> Ppx_deriving_runtime.unit

val show_binop : binop -> Ppx_deriving_runtime.string

val pp_expr_table_entry : Ppx_deriving_runtime.Format.formatter -> expr_table_entry -> Ppx_deriving_runtime.unit

val show_expr_table_entry : expr_table_entry -> Ppx_deriving_runtime.string

val pp_lua_function : Ppx_deriving_runtime.Format.formatter -> lua_function -> Ppx_deriving_runtime.unit

val show_lua_function : lua_function -> Ppx_deriving_runtime.string

val pp_const : Ppx_deriving_runtime.Format.formatter -> const -> Ppx_deriving_runtime.unit

val show_const : const -> Ppx_deriving_runtime.string

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

val pp_apply : Ppx_deriving_runtime.Format.formatter -> apply -> Ppx_deriving_runtime.unit

val show_apply : apply -> Ppx_deriving_runtime.string

val pp_statement : Ppx_deriving_runtime.Format.formatter -> statement -> Ppx_deriving_runtime.unit

val show_statement : statement -> Ppx_deriving_runtime.string

val pp_lvalue : Ppx_deriving_runtime.Format.formatter -> lvalue -> Ppx_deriving_runtime.unit

val show_lvalue : lvalue -> Ppx_deriving_runtime.string

val pp_elseif_block : Ppx_deriving_runtime.Format.formatter -> elseif_block -> Ppx_deriving_runtime.unit

val show_elseif_block : elseif_block -> Ppx_deriving_runtime.string

type ast = block

val pp_ast : Ppx_deriving_runtime.Format.formatter -> ast -> Ppx_deriving_runtime.unit

val show_ast : ast -> Ppx_deriving_runtime.string

type t = ast