Synduce Library

Basic Functionality

The Lib library provides some basic functionality for Synduce to be used as a library.

Main Algorithms

The Algo library contains the main algorithmic components of Synduce. The entry point of this library is the module:

Code Generation

The Codegen library contains all the code generators of Synduce. The current version only generates sketches of Dafny proofs.

Internal Representations

Lang contains the main internal representation definitions and interfaces with SMT-LIB and SyGuS components. Some program analysis algorithms are also implemented in Lang, but those algorithms should be reusable for other synthesis problems and not only the ones of Synduce.

Front End

Frontend provides the front-end functionality of Synduce.

SmtLib Interface

Smtlib provides an implementation to interact with SMT-LIB compliant solvers. The internal representation interfaces with this implementation through Lang.SmtInterface.

SyGuS Interface

The library syguslib-utils provides an implementation to interact with SyGuS solvers. The internal representation interfaces with this implementation through Lang.SygusInterface.

Utils and Configuration

Utils contains mainly utility functions. The module Utils.Config contains configuration settings that can be set by the user.