Module Lang.Alpha

val get_ids : Base.unit -> (Base.int, Base.string) Base.Hashtbl.t

Get a reference to the table of ids mapping integers ids to names.

val reinit : Base.unit -> Base.unit

Reinitialize all state values in the name.

val new_id : Base.unit -> Base.int

Returns a fresh, unused id.

val forget : Base.int -> Base.string -> Base.unit

`forget id name` Forget the assignments of id and name.

val fresh : ?⁠s:Base.string -> Base.unit -> Base.string

Creates a fresh name with base ~s (default "x").

val mk_with_id : Base.int -> Base.string -> (Base.int -> 'a) -> 'a

mk_with_id i s f applies f to i is i < 0, otherwise creates a fresh id, reigsters the name s to that fresh id, and returns the result of applying f to that fresh id.