Module Lang.Alpha

type t = {

ids : ( Base.int, Base.string ) Base.Hashtbl.t;

names : ( Base.string, Base.int Base.list ) Base.Hashtbl.t;

max_id : Base.int Base.ref;

}

An environment for names associated to ids.

val create : Base.unit -> t

val copy : t -> t

val reinit : t -> Base.unit

Reinitialize all state values in the name.

val new_id : t -> Base.int

Returns a fresh, unused id.

val forget : t -> Base.int -> Base.string -> Base.unit

`forget id name` Forget the assignments of id and name.

val fresh : ?s:Base.string -> t -> Base.string

Creates a fresh name with base ~s (default "x").

val mk_with_id : t -> Base.int -> Base.string -> ( Base.int -> 'a ) -> 'a

mk_with_id i s f applies f to i is i < 0, otherwise creates a fresh id, reigsters the name s to that fresh id, and returns the result of applying f to that fresh id.

val get_exn : t -> Base.int -> Base.string

Get the name that corresponds to a given id.

val check_source_variable_name : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> Base.string -> Base.unit

Check that a variable name from the source file obeys some basic restriction with synduce.

module Nice : sig ... end