Module Lang.Analysis

val vars_of_pattern : Term.pattern -> Term.VarSet.t

val free_variables : ?include_functions:bool -> ctx:Term.Context.t -> Term.term -> Term.VarSet.t

val has_ite : Term.term -> Base.bool

val has_calls : ctx:Term.Context.t -> _to:Term.VarSet.t -> Term.term -> Base.bool

val operators_of : Term.term -> Term.OpSet.t

val is_constant : Term.term -> Base.bool

val is_norec : ctx:Term.Context.t -> Term.term -> Base.bool

val is_novariant : ctx:Term.Context.t -> Term.term -> Base.bool

A term has no variant if it is a variable and the type of that variable has no variants. This means that term cannot be expanded.

val is_bounded : ctx:Term.Context.t -> Term.term -> Base.bool

val subst_args : ctx:Term.Context.t -> Term.fpattern Base.list -> Term.term Base.list -> ( Term.fpattern Base.list * (Term.term * Term.term) Base.list, Base.string ) Base.Result.t

val argset_of : ctx:Term.Context.t -> Term.variable -> Term.term -> ( Term.VarSet.elt, Base.Comparator.Make(Lang__Variables.Variable.T).comparator_witness ) Base.Set.t

val replace_calls_to : ctx:Term.Context.t -> Term.VarSet.t -> Term.term -> Term.term

replace_calls_to functions t replaces all calls to functions in functions by a fresh variable.

val replace_id_calls : func:Term.variable -> Term.term -> Term.term

val rename_nicely : ctx:Term.Context.t -> (Base.string * Term.variable Base.list * Term.term) Base.list -> (Base.string * Term.variable Base.list * Term.term) Base.list

val is_expandable_var : ctx:Term.Context.t -> Term.variable -> (Term.variable * (Base.string * RType.t Base.list) Base.list) option

val expand_once : ctx:Term.Context.t -> Term.term -> Term.term Base.list

val variant_no_recurse : ctx:Term.Context.t -> (Base.string * RType.t Base.list) Base.list -> Term.term Base.list

Given a list of variants, returns a list of terms that corresponds to the non-recursive constructors, with their arguments instantiated as variables.

val terms_of_max_depth : ctx:Term.Context.t -> Base.int -> RType.t -> Term.term Base.list

type virtual_term =

| VConstr of Base.string * virtual_term Base.list

| VChoice of virtual_term Base.list

| VVarLeaf of Term.Variable.t

val pp_virtual_term : ctx:Term.Context.t -> virtual_term Fmt.t

val virtual_variant_no_recurse : ctx:Term.Context.t -> (Base.string * RType.t Base.list) Base.list -> virtual_term

val virtual_term_of_max_depth : ctx:Term.Context.t -> Base.int -> RType.t -> virtual_term

val pick_some : ctx:Term.Context.t -> virtual_term -> Term.term Base.option * virtual_term Base.option

Pick some term in a vitual term, removing the choices made from the virtual term.

val _INT_MAX : int

Arbitrary min and max values. Actual useful values will be generated by solvers anyway.

val _INT_MIN : int

val rand_const_of_type : ctx:Term.Context.t -> RType.t -> Term.term

val concretize : ?model: ( Base.String.t, Term.term, Base.String.comparator_witness ) Base.Map.t -> ctx:Term.Context.t -> Term.term -> Term.term