Module Lang.PMRS

type rewrite_rule = Term.variable * Term.variable Base.list * Term.pattern Base.option * Term.term

A PMRS rewrite rule (v, args, pattern, rhs) is the rewrite rule v(args,pattern) -> rhs. v which is a non-terminal (a recursive function in the OCaml code), args is a possibly empty list of arguments, pattern is an optional pattern, rhs is the production of the rule.

type top_function = Term.variable * Term.variable Base.list * Term.term

A toplevel function is a triple of a variable (the variable representing the function itself), a list of variables (the arguments of the function) and a term (the body of the function).

type t = {

pvar : Term.Variable.t;	(*
The main function symbol, storing general information about the PMRS. The specification associated to the PMRS may be found using: get_spec pvar The type of the PMRS (the main function of the group of mutually recursive functions) is stored as the type of pvar.
*)

pinput_typ : RType.t Base.list;	(*
The input type(s). For now, it should be a singleton list. The input type is only the type of the recursively-typed argument of the PMRS, not the parameters.
*)

poutput_typ : RType.t;	(*
Output type and optional invariant on output of function.
*)

pargs : Term.Variable.t Base.list;	(*
Parameter arguments.
*)

psyntobjs : Variables.VarSet.t;	(*
The unknowns to synthesize.
*)

prules : rewrite_rule Utils.IntMap.t;	(*
The rules of the PMRS.
*)

pnon_terminals : Term.VarSet.t;	(*
Non-terminals of the PMRS.
*)

pmain_symb : Term.variable;	(*
The main symbol of the PMRS.
*)

porder : Base.int;	(*
The order of the PMRS (mostly useless for now).
*)

plogic : SmtLogic.logic_info;

}

The main PMRS type.

module Functions : sig ... end

Context for:

val update_order : t -> t

Update the order of a PMRS.

val set_logic_info_of_pmrs : ctx:Term.Context.t -> t -> t

Set the logic info field of a PMRS.

1 Utility functions, transformation, conversion to functions.

}

val func_to_pmrs : ctx:Term.Context.t -> Term.Variable.t -> Term.fpattern Base.list -> Term.term -> t

Translate a recursive function to a PMRS (currently only work for identity).

val func_of_pmrs : ctx:Term.Context.t -> t -> Term.function_descr Base.list

Translate a PMRS to a set of (mutually) recursive functions.

val inverted_rule_lookup : ?boundvars:( Term.Variable.t, Term.Variable.comparator_witness ) Base.Set.t -> ctx:Term.Context.t -> ( 'a, Term.Variable.t * Term.Variable.t Base.list * Term.pattern Base.option * Term.term, 'b ) Base.Map.t -> Term.term -> Term.term Base.list -> ( 'a, Term.term, 'b ) Base.Map.t

inverted_rule_lookup searches for rules whose rhs match (func args), and return a map from rule id to the lhs of the rules matching (func args), with the appropriate substitutions performed.

val subst_rule_rhs : ctx:Term.Context.t -> p:t -> (Term.term * Term.term) Base.list -> t

Apply a substitution to all the right hand side of the PMRS rules.

val depends : glob:Functions.ctx -> ctx:Term.Context.t -> t -> t Base.list

Given an input PMRS, returns a list of PMRS that this PMRS depends on.

val lhs : ctx:Term.Context.t -> (Term.variable * Term.variable Base.list * Term.pattern Base.option * Term.term) -> Term.term

Generate the term that corresponds to the left hand side of a rewrite rule in a PMRS.

val update_output_type : ctx:Term.Context.t -> t -> t

Updates and returns the output type of a PMRS.

1 PMRS and Types

}

val clear_pmrs_types : ctx:Term.Context.t -> t -> t

Clear the type information inside a PMRS.

val infer_pmrs_types : ctx:Term.Context.t -> t -> t

Infer the types inside a PMRS.

val unify_two_with_vartype_update : ctx:Term.Context.t -> (RType.t * RType.t) -> (RType.t * RType.t) -> RType.substitution

val unify_one_with_update : ctx:Term.Context.t -> (RType.t * RType.t) -> Base.unit

val extract_rec_input_typ : t -> RType.t

1 Pretty printing functions.

}

val pp_rewrite_rule : ctx:Term.Context.t -> Base.Formatter.t -> rewrite_rule -> Base.unit

Pretty print a rewrite rule, as in a PMRS.

val pp : ctx:Term.Context.t -> Base.Formatter.t -> ?short:Base.bool -> t -> Base.unit

Pretty print a PMRS.

val pp_ocaml : ctx:Term.Context.t -> Base.Formatter.t -> ?short:Base.bool -> t -> Base.unit

Pretty print a PMRS using Ocaml syntax (as a set of recursive functions).