Module Lang.RType

val _tid : int Base.ref

val _tvar_idx : int Base.ref

type ident = Base.string

type typekind =

| TyInt	(*
The int type.
*)

| TyBool	(*
The bool type.
*)

| TyString	(*
The string type.
*)

| TyChar	(*
The char type.
*)

| TyFun of type_term * type_term	(*
The function type.
*)

| TyTyp of ident

| TyParam of ident

| TyConstr of type_term Base.list * type_term

| TySum of type_term Base.list

| TyVariant of ident * type_term Base.list

Used for type declarations

and type_term = {

pos : Stdlib.Lexing.position * Stdlib.Lexing.position;

tkind : typekind;

}

val dummy_loc : Stdlib.Lexing.position * Stdlib.Lexing.position

val mk_t_int : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term

val mk_t_bool : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term

val mk_t_string : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term

val mk_t_char : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term

val mk_t_typ : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ident -> type_term

val mk_t_param : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ident -> type_term

val mk_t_constr : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term Base.list -> type_term -> type_term

val mk_t_sum : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term Base.list -> type_term

val mk_t_variant : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> ident -> type_term Base.list -> type_term

val mk_t_fun : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> type_term -> type_term -> type_term

type t =

| TInt

| TBool

| TString

| TChar

| TNamed of Base.string

| TTup of t Base.list

| TFun of t * t

| TParam of t Base.list * t

| TVar of Base.int

val t_of_sexp : Base__.Ppx_sexp_conv_lib.Sexp.t -> t

val sexp_of_t : t -> Base__.Ppx_sexp_conv_lib.Sexp.t

val pp : t Fmt.t

val fun_typ_unpack : t -> t Base.list * t

fun_type_unpack t uncurries the type t as a function. Returns a pair of a list (the type of the arguments of the function) and a type (the output type of the function). If t is not a function type, then (,t) is returned.

val fun_typ_pack : t Base.list -> t -> t

fun_typ_pack args t construct the type of a function args -> t: for example fun_typ_pack [t1;t2] t_out is TFun(t1, TFun(t2, tout)).

val base_name : t -> Base.string Base.option

val smt_name : t -> Base.string Base.option

val _unsafe_pair_t_ : t -> Base.string * t

Type environment module

type env = {

variant_typenames : ( Base.string, Base.string ) Base.Hashtbl.t;

typenames_variants : ( Base.string, Base.string Base.list ) Base.Hashtbl.t;

variant_types : ( Base.string, t * t Base.list ) Base.Hashtbl.t;

typenames : ( Base.string, t ) Base.Hashtbl.t;

idx : Base.int Base.ref;

}

val create : unit -> env

val copy : env -> env

val get_fresh_tvar : env -> t

val get_type : env -> Base.string -> t option

val dump_types : Base.Formatter.t -> env -> unit

val reinit : env -> Base.unit

val substitute : old:t -> by:t -> in_:t -> t

val sub_all : (t * t) Base.list -> t -> t

val subtype_of : t -> t -> Base.bool

val occurs : Base.int -> t -> Base.bool

type substitution = (Base.int * t) Base.list

A substitution is a list pairs of type variable id, type. Applied to a (parametric) type to subsitute tyep parameters with types.

val unify_one : t -> t -> ( substitution, Base.Sexp.t ) Base.Result.t

val mkv : substitution -> (t * t) Base__List.t

val unify : (t * t) Base__List.t -> ( substitution, Base.Sexp.t ) Base.Result.t

val merge_subs : (Stdlib.Lexing.position * Stdlib.Lexing.position) -> substitution -> substitution -> substitution

val t_equals : t -> t -> bool

val type_of_variant : env -> Base.string -> (t * t Base.list) Base.option

val instantiate_variant : type_term Base.list -> (ident * Base.int) Base.list -> t Base.List.t

val add_variant : env -> variant:Base.string -> typename:Base.string -> (t * t Base.list) -> unit

val add_all_variants : env -> params:(ident * Base.int) Base.list -> main_type:t -> typename:ident -> type_term Base.list -> Base.unit

val get_variants : env -> t -> (Base.string * t Base.list) Base.list

get_variants t returns a list of variants of a given type. If the type is a sum type with constructors, it returns a list of pairs of constructors, list of the types of the arguments of the constructor. If the type has no variant, the list is empty.

val is_recursive_variant : env -> (Base.string * t Base.list) -> bool

val add_type : env -> ?params:ident Base.list -> typename:Base.string -> type_term -> ( unit, Base.Sexp.t ) Base.result

`add_type ?params ~typename tterm` adds a type with name typename and parameters params (default empty list) and type term tterm to the global store.

val reduce : case:( ( t -> 'a ) -> t -> 'a Base.option ) -> init:'a -> join:( 'a -> 'a -> 'a ) -> t -> 'a

val is_base : t -> bool

val is_function : t -> bool

val is_recursive : env -> t -> Base.bool

val is_datatype : env -> t -> bool

val is_user_defined : t -> Base.bool

val get_datatype_depends : env -> t -> t Base.List.t