Module Lang.SmtInterface

module Stats : Smtlib.Solvers.Statistics

module SmtLog : Smtlib.Solvers.Logger

module AsyncSmt : sig ... end

module SyncSmt : sig ... end

val rtype_of_smtSort : ctx:Term.Context.t -> Smtlib.SmtLib.smtSort -> RType.t option

val sort_of_rtype : RType.t -> Smtlib.SmtLib.smtSort

val dec_parametric : RType.t -> RType.t list -> Smtlib.SmtLib.smtSort

val decl_of_tup_type : RType.t list -> Smtlib.SmtLib.smtSymbol * Smtlib.SmtLib.command

val declare_datatype_of_rtype : ctx:Term.Context.t -> RType.t -> (Smtlib.SmtLib.smtSymbol list * Smtlib.SmtLib.command) list

val smtPattern_of_pattern : ctx:Term.Context.t -> Term.pattern -> Smtlib.SmtLib.smtPattern

val term_of_const : Term.Constant.t -> Smtlib.SmtLib.smtTerm

val smt_of_term : ctx:Term.Context.t -> Term.term -> Smtlib.SmtLib.smtTerm

val smt_of_case : ctx:Term.Context.t -> Term.match_case -> Smtlib.SmtLib.match_case

val constant_of_smtConst : Smtlib.SmtLib.smtSpecConstant -> Term.Constant.t

type id_kind =

| ICstr of string

| IVar of Term.variable

| IBinop of Term.Binop.t

| IUnop of Term.Unop.t

| IBool of bool

| ITupCstr

| INotDef

val term_of_smt : fctx:PMRS.Functions.ctx -> ctx:Term.Context.t -> ( string, Term.variable, Base.String.comparator_witness ) Base.Map.t -> Smtlib.SmtLib.smtTerm -> Term.term

val sorted_vars_of_vars : ctx:Term.Context.t -> Term.VarSet.t -> Smtlib.SmtLib.smtSortedVar list

type term_model = ( string, Term.term, Base.String.comparator_witness ) Base.Map.t

val model_to_constmap : fctx:PMRS.Functions.ctx -> ctx:Term.Context.t -> SyncSmt.solver_response -> ( string, Term.term, Base.String.comparator_witness ) Base.Map.t

val model_to_varmap : fctx:PMRS.Functions.ctx -> ctx:Term.Context.t -> Term.VarSet.t -> SyncSmt.solver_response -> Term.term Term.VarMap.t

val request_different_models : fctx:PMRS.Functions.ctx -> ctx:Term.Context.t -> term_model -> int -> Smtlib.Solvers.Synchronous(SmtLog)(Stats).online_solver -> ( string, Term.term, Base.String.comparator_witness ) Base.Map.t list

val request_different_models_async : fctx:PMRS.Functions.ctx -> ctx:Term.Context.t -> term_model Lwt.t -> int -> Smtlib.Solvers.Asyncs(SmtLog)(Stats).solver -> ( string, Term.term, Base.String.comparator_witness ) Base.Map.t list Lwt.t

val smtPattern_of_term : Term.term -> Smtlib.SmtLib.smtPattern option

val mk_assert : Smtlib.SmtLib.smtTerm -> Smtlib.SmtLib.command

val smt_of_pmrs : fctx:PMRS.Functions.ctx -> ctx:Term.Context.t -> PMRS.t -> Smtlib.SmtLib.command list

Generate a set of definitions that define a PMRS. This can consist of datatype declarations and recursive definitions. If the PMRS uses parameters, then the definitions for the parameters are included in the set of definitions.

module Commands : sig ... end