Module Lib

module Parsers = Parsers

module Utils = Utils

module Algos = Algo.PmrsAlgos

type problem_descr = {

pd_target : Lang.Term.function_descr Base.list;

pd_reference : Lang.Term.function_descr Base.list;

pd_repr : Lang.Term.function_descr Base.list;

}

val reinit : unit -> unit

Use reinit to reinitialize all the global variables used in Synduce when solving multiple problems.

val pp_problem_descr : Base.Formatter.t -> problem_descr -> unit

val solve_file : ?⁠print_info:bool -> Base.string -> problem_descr * Algo.AState.soln Base.option

val get_lemma_hints : unit -> Algo.AState.equation_system

Call get_lemma_hints () after solve_file to get a list of potential useful lemmas for the proof of correctness.