Module Smtlib.SmtLib

This module implements an interface to the SMT-LIB standard. Some parts of documentation will refer to a specific section of the SMT-LIB standard available online (http://smtlib.cs.uiowa.edu/papers/smt-lib-reference-v2.6-r2021-05-12.pdf).

module OC = Stdio.Out_channel

module IC = Stdio.In_channel

type numeral = Base.int

The type of numbers.

type smtSymbol =

| SSimple of Base.string

| SQuoted of Base.string

A SMT symbol is either a string identifier (e.g. x) or a quoted symbol (e.g. | this|). See Section 3.1 of the SMT-LIB language standard.

val symb_equal : smtSymbol -> smtSymbol -> bool

val str_of_symb : smtSymbol -> Base.string

type smtIndex =

| INum of numeral

| ISym of smtSymbol

An SMT index is either a number or a symbol (Section 3.3).

type smtIdentifier =

| Id of smtSymbol

| IdC of smtSymbol * smtIndex Base.list	(*
For example, (_ vector-add 4 5) is an indexed identifier.
*)

A SMT identifier is either a symbol or an indexed symbol (Section 3.3).

type smtSort =

| SmtSort of smtIdentifier	(*
For example SmtSort(IdC(SSimple("BitVec"), INum(3))) is (_ BitVec 3)
*)

| Comp of smtIdentifier * smtSort Base.list	(*
For example Comp(Id(SSimple("Set")),SmtSort(Id(SSimple("Int")))) is (Set Int)
*)

A SMT Sort (a type) is an identifier or a composite type.

type smtSortedVar = smtSymbol * smtSort

A SMT sorted var is a symbol associated with a sort.

type smtQualIdentifier =

| QI of smtIdentifier

| QIas of smtIdentifier * smtSort

A qualified identifier is either a simple identifier QI ... or an identifier qualified with a sort QIas(x,y) is (as x y).

type smtAttribute =

| ASymb of Base.string

| ASymbVal of Base.string * Base.string

An SMT attribute that is either a string symbol or a symbol with a string value.

type smtPattern =

| Pat of smtSymbol

| PatComp of smtSymbol * smtSymbol Base.list

A SMT pattern is either a variable (a SMT symbol) or a constructor (a pair of a constructor name with a lsit of arguments).

type smtSpecConstant =

| SCNumeral of numeral

| SCDecimal of numeral * Base.int * numeral

| SCHexaDecimal of Base.string

| SCBinary of Base.bool Base.list

| SCString of Base.string

The type of SMT special constants (Section 3.1).

type smtTerm =

| SmtTSpecConst of smtSpecConstant

| SmtTQualdId of smtQualIdentifier

| SmtTApp of smtQualIdentifier * smtTerm Base.list

| SmtTLet of var_binding Base.list * smtTerm

| SmtTForall of smtSortedVar Base.list * smtTerm

| SmtTExists of smtSortedVar Base.list * smtTerm

| SmtTMatch of smtTerm * match_case Base.list

| SmtTAnnot of smtTerm * smtAttribute Base.list

and match_case = smtPattern * smtTerm

and var_binding = smtSymbol * smtTerm

type smtSortDec = smtSymbol * numeral

type smtSelectorDec = smtSymbol * smtSort

type smtConstructorDec = smtSymbol * smtSelectorDec Base.list

type datatype_dec =

| DDConstr of smtConstructorDec Base.list

| DDParametric of smtSymbol Base.list * smtConstructorDec Base.list

type func_dec = smtSymbol * smtSortedVar Base.list * smtSort

type func_def = smtSymbol * smtSortedVar Base.list * smtSort * smtTerm

type prop_literal =

| PL of smtSymbol

| PLNot of smtSymbol

type info_flag = Base.int

type solver_option = Base.string * Base.string

type command =

| Assert of smtTerm

| CheckSat

| CheckSatAssuming of prop_literal Base.list

| DeclareConst of smtSymbol * smtSort

| DeclareDatatype of smtSymbol * datatype_dec

| DeclareDatatypes of smtSortDec Base.list * datatype_dec Base.list

| DeclareFun of smtSymbol * smtSort Base.list * smtSort

| DeclareSmtSort of smtSymbol * numeral

| DefineFun of func_def

| DefineFunRec of func_def

| DefineFunsRec of func_dec Base.list * smtTerm Base.list

| DefineSmtSort of smtSymbol * smtSymbol Base.list * smtSort

| Echo of Base.string

| Exit

| GetAssertions

| GetAssignment

| GetInfo of info_flag

| GetModel

| GetOption of Base.string

| GetProof

| GetUnsatAssumptions

| GetUnsatCore

| GetValue of smtTerm Base.list

| Pop of numeral

| Push of numeral

| Reset

| ResetAssertions

| SetInfo of Base.string

| SetLogic of smtSymbol

| SetOption of solver_option

| Simplify of smtTerm * solver_option Base.list

type script = command Base.list

val sexps_of_option : solver_option -> Base.Sexp.t Base.list

val option_of_sexp : Base.Sexp.t -> solver_option Base.option

val sexp_of_info_flag : info_flag -> Base.Sexp.t

val info_flag_of_sexp : Base.Sexp.t -> info_flag

val sexp_of_numeral : numeral -> Base.Sexp.t

val numeral_of_sexp : Base.Sexp.t -> numeral Base.option

val sexp_of_smtSpecConstant : smtSpecConstant -> Base.Sexp.t

val smtSpecConstant_of_sexp : Base.Sexp.t -> smtSpecConstant Base.option

val sexp_of_smtSymbol : smtSymbol -> Base.Sexp.t

val smtSymbol_of_sexp : Base.Sexp.t -> smtSymbol Base.option

val sexp_of_smtIndex : smtIndex -> Base.Sexp.t

val smtIndex_of_sexp : Base.Sexp.t -> smtIndex Base.option

val sexp_of_smtIdentifier : smtIdentifier -> Base.Sexp.t

val smtIdentifier_of_sexp : Base.Sexp.t -> smtIdentifier Base.option

val sexp_of_smtSort : smtSort -> Base.Sexp.t

val smtSort_of_sexp : Base.Sexp.t -> smtSort Base.option

val sexp_of_smtAttribute : smtAttribute -> Base.Sexp.t

val smtAttribute_of_sexp : Base.Sexp.t -> smtAttribute Base.option

val sexp_of_sortedVar : smtSortedVar -> Base.Sexp.t

val sortedVar_of_sexp : Base.Sexp.t -> smtSortedVar Base.option

val sexp_of_smtQualIdentifier : smtQualIdentifier -> Base.Sexp.t

val smtQualIdentifier_of_sexp : Base.Sexp.t -> smtQualIdentifier Base.option

val sexp_of_smtPattern : smtPattern -> Base.Sexp.t

val smtPattern_of_sexp : Base.Sexp.t -> smtPattern Base.option

val sexp_of_smtTerm : smtTerm -> Base.Sexp.t

val sexp_of_binding : var_binding -> Base.Sexp.t

val sexp_of_match_case : match_case -> Base.Sexp.t

val smtTerm_of_sexp : Base.Sexp.t -> smtTerm Base.option

val match_case_of_sexp : Base.Sexp.t -> match_case Base.option

val binding_of_sexp : Base.Sexp.t -> var_binding Base.option

val sexp_of_smtSelectorDec : smtSelectorDec -> Base.Sexp.t

val smtSelectorDec_of_sexp : Base.Sexp.t -> smtSelectorDec Base.option

val sexp_of_smtConstructorDec : smtConstructorDec -> Base.Sexp.t

val smtConstructorDec_of_sexp : Base.Sexp.t -> smtConstructorDec Base.option

val sexp_of_datatype_dec : datatype_dec -> Base.Sexp.t

val datatype_dec_of_sexp : Base.Sexp.t -> datatype_dec Base.option

val sexp_of_func_dec : func_dec -> Base.Sexp.t

val func_dec_of_sexpl : Base.Sexp.t Base.list -> func_dec Base.option

val func_dec_of_sexp : Base.Sexp.t -> func_dec Base.option

val sexp_of_func_def : func_def -> Base.Sexp.t Base.list

val func_def_of_sexp : Base.Sexp.t Base.list -> func_def Base.option

val sexp_of_prop_literal : prop_literal -> Base.Sexp.t

val prop_literal_of_sexp : Base.Sexp.t -> prop_literal Base.option

val sexp_of_smtSortDec : smtSortDec -> Base.Sexp.t

val smtSortDec_of_sexp : Base.Sexp.t -> smtSortDec Base.option

val sexp_of_command : command -> Base.Sexp.t

val command_of_sexp : Base.Sexp.t -> command Base.option

val write_command : OC.t -> command -> Base.unit

val pp_script : Base.Formatter.t -> script -> unit

val mk_symb : Base.string -> smtSymbol

val mk_int_sort : smtSort

val mk_bool_sort : smtSort

val mk_seq_sort : smtSort -> smtSort

val mk_tuple_sort : smtSort Base.list -> smtSort

val mk_simple_sort : Base.string -> smtSort

val mk_qi : Base.string -> smtQualIdentifier

val mk_int : Base.int -> smtTerm

val mk_false : smtTerm

val mk_true : smtTerm

val mk_nil : smtTerm

val mk_simple_app : Base.string -> smtTerm Base.list -> smtTerm

val mk_var : Base.string -> smtTerm

val mk_ite : smtTerm -> smtTerm -> smtTerm -> smtTerm

val mk_add : smtTerm -> smtTerm -> smtTerm

val mk_sub : smtTerm -> smtTerm -> smtTerm

val mk_mul : smtTerm -> smtTerm -> smtTerm

val mk_div : smtTerm -> smtTerm -> smtTerm

val mk_and : smtTerm -> smtTerm -> smtTerm

val mk_assoc_and : smtTerm Base.list -> smtTerm

val mk_or : smtTerm -> smtTerm -> smtTerm

val mk_eq : smtTerm -> smtTerm -> smtTerm

val mk_lt : smtTerm -> smtTerm -> smtTerm

val mk_gt : smtTerm -> smtTerm -> smtTerm

val mk_le : smtTerm -> smtTerm -> smtTerm

val mk_ge : smtTerm -> smtTerm -> smtTerm

val mk_not : smtTerm -> smtTerm

val mk_let : var_binding Base.list -> smtTerm -> smtTerm

val mk_exists : smtSortedVar Base.list -> smtTerm -> smtTerm

val mk_forall : smtSortedVar Base.list -> smtTerm -> smtTerm

val mk_assert : smtTerm -> command

val mk_const_decl : Base.string -> smtSort -> command

val mk_check_sat : command

val mk_fun_decl : Base.string -> smtSort Base.list -> smtSort -> command

val mk_fun_def : Base.string -> (Base.string * smtSort) Base.list -> smtSort -> smtTerm -> command

val mk_named_assert : Base.string -> smtTerm -> command

val mk_set_option : Base.string -> Base.string -> command

val mk_set_logic : Logics.logic -> command

val mk_simplify : ?options:solver_option Base.list -> smtTerm -> command

val mk_print_success : command

val mk_exit : command

val mk_pop : numeral -> command

val mk_push : numeral -> command

val mk_min_def : command

mk_min_def is the following SMT-LIB command:

  (define-function min ((x Int) (y Int)) Int (ite (<= x y) x y))

val mk_max_def : command

mk_max_def is the following SMT-LIB command:

  (define-function max ((x Int) (y Int)) Int (ite (>= x y) x y))

val mk_inv_def : command

mk_inv_def is the following SMT-LIB command:

  (define-function inv ((x Int)) Int (div 1 x))

val smtSymbol_of_decl : command -> smtSymbol option

val remove_duplicate_decls : Base.String.t Base.Hash_set.t -> command Base.list -> command Base__List.t

type solver_response =

| Error of Base.string

| Sat	(*
The solver responded sat
*)

| Unsat	(*
The solver responded unsat
*)

| Unknown	(*
The solver responded unknown
*)

| Success	(*
The solver responded success (the :print-success option is on)
*)

| Unsupported	(*
The solver reponded that the input is not supported.
*)

| SExps of Base.Sexp.t Base.list	(*
Any other type of response is a list of S-Expressions, which can be interpreted later.
*)

val parse_response : Base.Sexp.t Base.list -> solver_response

val pp_solver_response : Stdlib.Format.formatter -> solver_response -> unit