Module Smtlib.Solvers

This module contains functions to interface with syntax-guided synthesis solvers using the SMTLIB language defined in The synchronous and asynchronous solvers defined here are functors parametric on Logger and Statistics modules to automate logging on some output and collecting statistics on solver usage.

Log and Stats modules

module type Logger = sig ... end

The Logger module must provide some basic logging functionality, including error, debug and verbose messages. One can also set log_queries to true with a file, which allows the solver to write queries to a separate file.

module EmptyLog : Logger

An empty logger that ignores messages, and can be used to construct a solver if no logging is desired.

module type Statistics = sig ... end

The statistics modules allows logging start/termination time of subprocesses used to solve the syntax-guided synthesis instances. It should also provide a function that returns the time elapsed in the program.

module NoStat : Statistics

An empty Stats module that does nothing.

module Synchronous (Log : Logger) (Stats : Statistics) : sig ... end

module Asyncs (Log : Logger) (Stats : Statistics) : sig ... end