Module Utils

module Config : sig ... end

module Stats : sig ... end

module Log : sig ... end

module LogJson : sig ... end

module IntMap : sig ... end

module StringMap : sig ... end

type z =

| Z : z

type 'n s =

| S : 'n -> 'n s

val identity : 'a -> 'a

val first : ('a * 'b) -> 'a

val second : ('a * 'b) -> 'b

val (<|) : ( 'a -> 'b ) -> ( 'c -> 'a ) -> 'c -> 'b

val product : 'a Base.list Base.list -> 'a Base.list Base.list

val time : ( unit -> 'a ) -> 'a

val (>>?|) : 'a Base.Option.t -> ( 'a -> 'b ) -> 'b Base.Option.t

val (>>!|) : ( 'a, 'b ) Base.Result.t -> ( 'a -> 'c ) -> ( 'c, 'b ) Base.Result.t

val (>>=?) : 'a Base.Option.t -> ( 'a -> 'b Base.Option.t ) -> 'b Base.Option.t

val (>>=!) : ( 'a, 'b ) Base.Result.t -> ( 'a -> ( 'c, 'b ) Base.Result.t ) -> ( 'c, 'b ) Base.Result.t

val repeat : int -> ( 'a -> 'a ) -> init:'a -> 'a

val satom : string -> Sexplib.Sexp.t

val slist : Sexplib.Sexp.t list -> Sexplib.Sexp.t

val blast : ( 'a, 'b Base.List.t ) Base.Result.t list -> ( 'a list, 'b Base.List.t ) Base.Result.t

val pair : 'a -> 'b -> 'a * 'b

val index_list : 'a Base.List.t -> (int * 'a) Base.List.t

val sub_list : 'a Base.list -> Base.int Base.list -> 'a Base__List.t

val trim : Base.string -> string

val ast : Stdlib.Format.formatter -> unit -> unit

val colon : Stdlib.Format.formatter -> unit -> unit

val vbar : Stdlib.Format.formatter -> unit -> unit

val dot : Stdlib.Format.formatter -> unit -> unit

val rightarrow : Stdlib.Format.formatter -> unit -> unit

val leftarrow : Stdlib.Format.formatter -> unit -> unit

val sep_and : Base.Formatter.t -> Base.unit -> Base.unit

val pp_link : Stdlib.Format.formatter -> string -> unit

val list_or_space : sep:unit Fmt.t -> f:'a Fmt.t -> Stdlib.Format.formatter -> 'a list -> unit

list_or_space prints the list using f for each element, and sep for the separator. If the list is empty, prints a space. If the list is not empty, the printed list is surrounded by two spaces.

val option_or_space : f:( Stdlib.Format.formatter -> 'a -> unit ) -> Stdlib.Format.formatter -> 'a option -> unit

option_or_space prints the option using f for the data, and sep for the separator. If the option is None, prints a space. If the option is not None, the printed data is surrounded by two spaces.

val to_subscript_unicode : Base.int -> Base.String.t

val list_map_snd : ('a * 'b) Base.list -> f:( 'b -> 'c ) -> ('a * 'c) Base.list

val cartesian_nary_product : 'a Base.list Base.list -> 'a Base.list Base.list

val combinations : 'a Base.list -> ('a * 'a) Base.list

val all_or_none : 'a Base.option Base.list -> 'a Base.list Base.option

val subsets : 'a list -> 'a list Base.List.t

type 'a continue_or_stop =

| Continue of 'a

| Stop of 'a

val lwt_until : f:( 'a -> 'b -> 'a continue_or_stop Lwt.t ) -> init:'a Lwt.t -> 'b Base.list -> 'a Lwt.t

val relative_to_root : Base.string -> string