Module Utils.Config

val problem_name : string Base.ref

val debug : bool Base.ref

Toggle debugging.

val compact : bool Base.ref

Print messages in compact form, non human readable.

val debug_msg_max_chars : int Base.ref

Limit the size of debugging messages.

val info : bool Base.ref

Toggle info and error messages. Set to true by default.

val json_out : bool Base.ref

Json output toggle.

val json_progressive : bool Base.ref

When Json output is on, set true to get partial updates in Json format. When the tool searches for a single solution, print intermediate refinement steps. When the tool searches for multiples ones, print intermediate results (unrealizable, failures and solutions).

val print_unrealizable_configs : bool Base.ref

val timings : bool Base.ref

Toggle printing timing info when info is off. Set to true by default.

val reduction_limit : int Base.ref

Maximum steps of rewrites to apply during PMRS reduction (symbolic evaluation).

val refinement_rounds_warning_limit : int Base.ref

After 50 refinement rounds, synthesized is probably using too many eager optimizations and the solution does not follow occam's razor.

val show_vars : bool Base.ref

val verbose : bool Base.ref

Toggle verbose messages.

val math_display : bool Base.ref

val output_folder : Base.string Base.option Base.ref

Optional output folder for solution. If None, the solution is only printed on the standard output. If Some path, the solution is written in path/inputfilename

val set_output_folder : Base.string -> Base.unit

val get_output_file : string -> string Base.Option.t

val set_json_out : unit -> Base.unit

Turn of all output except json.

val pp_eqn_count : int Base.ref

When printing a system of equations, put a limit on how many equations are printed.

val interactive_lemmas : bool Base.ref

Prompt user to input a precondition (lemma) for each equation, while the equations are being generated from a set of terms in Equations.make.

val interactive_lemmas_loop : bool Base.ref

val interactive_lifting : bool Base.ref

Prompt for lifting expressions.

val classify_witness : bool Base.ref

val interactive_check_lemma : bool Base.ref

val check_unrealizable : bool Base.ref

Check whether a system of equations defines a "functionally realizable" synthesis problem. ON by default.

val check_unrealizable_smt_unsatisfiable : bool Base.ref

Check whether a synthesis problem is unrealizable by checking whether the problem with second-order quantification is unrealizable, using a SMT solver (Z3 by default).

val no_bounded_sat_as_unsat : bool Base.ref

Accept no-sat in bounded checking as unsat.

val bounded_lemma_check : bool Base.ref

val sysfe_opt : bool Base.ref

Run with the optimizations used to synthesize solutions for systems of equations. Works best when there are more than 4 cores.

val force_nonlinear : bool Base.ref

Sometimes non-linearity occurs even when the reference only has linear operators.

val node_failure_behavior : bool Base.ref

When `node_failure_behaviour` is set to `true` (the default value), the failure to solve a given configuration is considered as an unrealizability result. That is, the configuration will not be refined further. If `node_failure_behaviour` is set to `false`, the failure will only interpreted as realizable: subconfigurations will be explored.

val tmp_folder : string

val root_folder : string

val base : string -> string

val use_cvc4 : bool Base.ref

val using_cvc4_tuples : bool Base.ref

You can use mkTuple and (Tuple ..) types with CVC4, but it doesn't seem to work with CVC5.

val no_grammar_for_constants : bool Base.ref

Don't generate a special grammar for constants if true.

val cvc4_binary_path : FilePath.filename option

val cvc5_binary_path : FilePath.filename option

val using_cvc5 : unit -> Base.bool

val cvc_binary_path : unit -> FilePath.filename

val z3_binary_path : FilePath.filename

val yices_binary_path : FilePath.filename option

val dryadsynth_binary_path : FilePath.filename

val use_eusolver : bool Base.ref

val eusolver_binary_path : FilePath.filename

val verification_solver : string Base.ref

val set_verification_solver : Base.string -> Base.unit

val smt_solver_log_file : string Base.ref

val smt_log_queries : bool Base.ref

val smt_solve_verbose : bool Base.ref

val generate_benchmarks : bool Base.ref

val benchmark_generation_dir : string Base.ref

val benchmark_lang_version : string Base.ref

val set_benchmark_generation_dir : Base.string -> Base.unit

val new_benchmark_file : ?hint:Base.String.t -> string -> string

val generate_proof : bool Base.ref

val proof_generation_file : string Base.ref

val set_proof_output_file : Base.string -> Base.unit

val included_files : Base.string Base.list Base.ref

val set_included_files : Base.string -> Base.unit

module Optims : sig ... end

val options : ( unit -> Base.unit ) -> bool Stdlib.ref -> (char * string * ( unit -> Base.unit ) option * ( Base.string -> Base.unit ) option) list