Module Config.Optims

val turn_off : bool Base.ref -> Base.unit

val detupling_on : bool Base.ref

Unkmowns of type i -> a * b * c .. are split into i -> a, i -> b and equations between tuples are separated into different equations. Turn off using option -t or --detupling-off

val use_syntactic_definitions : bool Base.ref

Syntactic definitions: detect syntactic definitions in the equation systems and use them. Turn off with --no-syndef

val use_rewrite_solver : bool Base.ref

Rewrite sovler: find solutions by rewriting expressions and isolate possible solutions for the functions. Turn off with --no-rew

val make_partial_correctness_assumption : bool Base.ref

Partial correctenss asumptions: don't throw away partial solutions Turn off with --no-assumptions

val make_partial_lemma_sketches : bool Base.ref

Turn on with --lemma-sketches.

val split_solve_on : bool Base.ref

Separate systems of equations into subsystems, and solve independently each subsystem. Use option -s or --split-solving-off to turn off.

val simple_init : bool Base.ref

Initialize T using a single variable of type theta, instead of searching for a set of terms covering all unknowns. Use option --simple-init or -c to turn on.

val bound_after_verif : bool Base.ref

In some cases, synthesizing lemmas nca be too hard; the algorithm can revert to bounded mode after a failed round of solution synthesis.

val use_segis : bool Base.ref

Not an optimization. Set to true to use the base abstract cegis algorithm. Option --use-segis sets this flag to true.

val use_cegis : bool Base.ref

Not an optimization. Set to true to use the base concrete cegis algorithm. Option --use-cegis sets this flag to true.

val simplify_eqns : bool Base.ref

Simplify equations before feeding them to equations solver. Turn off with --no-simplify

val optimize_grammars : int Base.ref

Use the equations as a indicator to optimize grammars, without compromising soundness.

val set_grammar_optimization_level : Base.string -> Base.unit

val expand_depth : int Base.ref

Maximum depth of pointwise expansions to perform. Careful setting this variable to high, it will lead to explosion in the number of terms expanded. A lazy expansion should be implemented.

val num_expansions_check : int Base.ref

val expand_cut : int Base.ref

val set_num_expansions_check : Base.string -> Base.unit

num_expansions_check is set by the -n or --verification option of the CLI.

val use_bmc : bool Base.ref

Use bounded model checking. From CLI, use --use-bmc to set to true.

val check_depth : int Base.ref

Depth of bounded model checking.

val set_check_depth : Base.string -> Base.unit

Bounded model checking depth is set by the -b or --bmc option of the CLI.

val induction_proof_tlimit : int Base.ref

A time limit for induction proofs. Infinity if set to negative.

val set_induction_proof_tlimit : Base.string -> Base.unit

val wait_parallel_tlimit : float Base.ref

A time limit parallel calls when waiting on first result.. Infinity if set to negative. Default is 10 mins.

val set_wait_parallel_tlimit : Base.string -> Base.unit

val rewrite_limit : int Base.ref

A limit for the number of rewriting steps applied during deduction.

val set_rewrite_limit : Base.string -> Base.unit

val fuzzing_count : int Base.ref

When a model has been found, attempt fuzzing to find models that satisfy the same constraints. Used in Counterexamples.ml.

val set_fuzzing_count : Base.string -> Base.unit

val attempt_lifting : bool Base.ref

Attempt to lift the function if there is no solution.

val max_lifting_attempts : int Base.ref

The number of times Synduce should attempt to add a lifting variable.

val set_max_lifting_attempts : Base.string -> Base.unit

val max_solutions : int Base.ref

The maximum number of different solutions Synduce attempts to find. If set to (-1), Synduce only attempt to solve the given skeleton.

val set_max_solutions : Base.string -> Base.unit

Set the maximum number of solutions. The string argument must represent a number between 1 and 32.

val rstar_limit : int Base.ref

The number of expansions an algorithm goes through to identify an algorithm run.

val set_rstar_limit : Base.string -> Base.unit

Set the number of expansions the algorithm uses to identify an algorithm run.

val some_eager_optim_on : unit -> Base.bool

val turn_off_eager_optims : unit -> Base.unit

val num_threads : int Base.ref

val set_num_threads : Base.string -> Base.unit